% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-lists.R
\name{anticipated_media}
\alias{anticipated_media}
\alias{movies_anticipated}
\alias{shows_anticipated}
\title{Anticipated media}
\usage{
movies_anticipated(limit = 10, extended = c("min", "full"),
  query = NULL, years = NULL, genres = NULL, languages = NULL,
  countries = NULL, runtimes = NULL, ratings = NULL,
  certifications = NULL, networks = NULL, status = NULL)

shows_anticipated(limit = 10, extended = c("min", "full"),
  query = NULL, years = NULL, genres = NULL, languages = NULL,
  countries = NULL, runtimes = NULL, ratings = NULL,
  certifications = NULL, networks = NULL, status = NULL)
}
\description{
These functions return the most anticipated movies/shows on trakt.tv.
}
\section{The Dynamic Lists on trakt.tv}{


These functions access the automatically updated lists provided by trakt.tv.
Each function comes in two flavors: Shows or movies. The following descriptions
are adapted directly from the
\href{https://trakt.docs.apiary.io/#reference/movies/popular/get-popular-movies}{API reference}.
\itemize{
\item \link[=popular_media]{Popular}: Popularity is calculated using the rating percentage and the number of
ratings.
\item \link[=trending_media]{Trending}: Returns all movies/shows being watched right now.
Movies/shows with the most users are returned first.
\item \link[=played_media]{Played}: Returns the most played (a single user can watch multiple times)
movies/shows in the specified time \code{period}.
\item \link[=watched_media]{Watched}: Returns the most watched (unique users) movies/shows in the specified
time \code{period}.
\item \link[=collected_media]{Collected}: Returns the most collected (unique users) movies/shows in the
specified time \code{period}.
\item \link[=anticipated_media]{Anticipated}: Returns the most anticipated movies/shows based on the number of
lists a movie/show appears on.
\item \link[=updated_media]{Updates}: Returns all movies/shows updated since the specified UTC \code{start_date}.
In this case, the upper bound for \code{limit} is 100.
}

The functions for \strong{Played}, \strong{Watched}, \strong{Collected} and \strong{Played} each return
the same additional variables besides the media information: \code{watcher_count},
\code{play_count}, \code{collected_count}, \code{collector_count}.
}

\examples{
\dontrun{
# Get 15 the most anticipated upcoming shows on Netflix that air this year
current_year <- format(Sys.Date(), "\%Y")
shows_anticipated(limit = 15, networks = "Netflix", years = current_year)
}
}
\seealso{
Other Dynamic lists: \code{\link{played_media}},
  \code{\link{popular_media}},
  \code{\link{trending_media}}, \code{\link{watched_media}}
}
\concept{Dynamic lists}
