% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{apidoc}
\alias{apidoc}
\title{Get API docs keys}
\usage{
apidoc(section, method, key = NULL)
}
\arguments{
\item{section}{E.g. "movies"}

\item{method}{E.g "summary"}

\item{key}{E.g. "url", "endpoint"}
}
\value{
\code{character(1)}
}
\description{
Get API docs keys
}
\keyword{internal}
