% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{apiurl}
\alias{apiurl}
\title{Get a formatted API url for an endpoint}
\usage{
apiurl(section, method, prefix = "@source ")
}
\arguments{
\item{section}{E.g. "movies"}

\item{method}{E.g. "summary"}
}
\value{
Markdown-formatted url
}
\description{
Get a formatted API url for an endpoint
}
\examples{
apiurl("lists", "popular")
}
\keyword{internal}
