% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{build_trakt_url}
\alias{build_trakt_url}
\title{Assemble a trakt.tv API URL}
\usage{
build_trakt_url(...)
}
\arguments{
\item{...}{Unnamed arguments will be concatenated with \code{/} as separators to
form the path of the API URL, e.g. the arguments \code{movies},
\code{tron-legacy-2012} and \code{releases} will be concatenated to
\code{movies/tron-legacy-2012/releases}. Additional \strong{named} arguments will be
used as query parameters, usually \code{extended = "full"} or others.}
}
\value{
A URL: \code{character} of length 1.
}
\description{
\code{build_trakt_url} assembles a trakt.tv API URL from different arguments.
The result should be fine for use with \link{trakt_get}, since that's what this
function was created for.
}
\examples{
build_trakt_url("shows", "breaking-bad", extended = "full")
build_trakt_url("shows", "popular", page = 3, limit = 5)

# Path can also be partially assembled already
build_trakt_url("users/jemus42", "ratings")

build_trakt_url("shows", "popular", page = 1, limit = 5)
}
\seealso{
Other utility functions: 
\code{\link{pad_episode}()}
}
\concept{utility functions}
