% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-internal.R
\docType{data}
\name{certifications}
\alias{certifications}
\title{Certifications}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 5 columns.}
\usage{
certifications
}
\description{
Currently only US certification are available.
}
\keyword{datasets}
