% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract-fix.R
\name{check_filter_arg}
\alias{check_filter_arg}
\title{Check filter arguments}
\usage{
check_filter_arg(filter, filter_type = c("query", "years", "genres",
  "languages", "countries", "runtimes", "ratings", "certifications",
  "networks", "status"))
}
\arguments{
\item{filter}{The username input.}
}
\description{
Check filter arguments
}
\keyword{internal}
