% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_trending.R
\name{comments_trending}
\alias{comments_trending}
\alias{comments_recent}
\alias{comments_updates}
\title{Get trending, recent or recently updated comments}
\usage{
comments_trending(comment_type = c("all", "reviews", "shouts"),
  type = c("all", "movies", "shows", "seasons", "episodes", "lists"),
  include_replies = FALSE, limit = 10L)

comments_recent(comment_type = c("all", "reviews", "shouts"),
  type = c("all", "movies", "shows", "seasons", "episodes", "lists"),
  include_replies = FALSE, limit = 10L)

comments_updates(comment_type = c("all", "reviews", "shouts"),
  type = c("all", "movies", "shows", "seasons", "episodes", "lists"),
  include_replies = FALSE, limit = 10L)
}
\arguments{
\item{comment_type}{\code{character(1) ["all"]}: The type of comment, one of "all",
"reviews" or "shouts".}

\item{type}{\code{character(1) ["all"]}: The type of media to filter by, one of
"all", "movies", "shows", "seasons", "episodes" or "lists".}

\item{include_replies}{\code{logical(1) [FALSE]}: Whether to include replies.}

\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced via \code{as.integer()}.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get trending, recent or recently updated comments
}
\section{Functions}{
\itemize{
\item \code{comments_trending}: Get trending comments.

\item \code{comments_recent}: Get recently made comments.

\item \code{comments_updates}: Get recently updated comments.
}}

\examples{
\dontrun{
# Trending reviews
comments_trending("reviews")

# Recent shouts (short comments)
comments_recent("shouts")

# Recently updated comments
comments_updates()
}
}
