% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_trending.R
\name{comments_updates}
\alias{comments_updates}
\title{Get recently updated/edited comments}
\source{
\code{comments_updates()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/comments/updates/get-recently-updated-comments}{/comments/updates/:comment_type/:type?include_replies=}.
}
\usage{
comments_updates(
  comment_type = c("all", "reviews", "shouts"),
  type = c("all", "movies", "shows", "seasons", "episodes", "lists"),
  include_replies = FALSE,
  limit = 10L
)
}
\arguments{
\item{comment_type}{\code{character(1) ["all"]}: The type of comment, one of "all",
"reviews" or "shouts".}

\item{type}{\code{character(1) ["all"]}: The type of media to filter by, one of
"all", "movies", "shows", "seasons", "episodes" or "lists".}

\item{include_replies}{\code{logical(1) [FALSE]}: Whether to include replies.}

\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced via \code{as.integer()}.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get recently updated/edited comments
}
\examples{
\dontrun{ # quite slow
# Recently updated comments
comments_updates(limit = 2)
}
}
\seealso{
Other comment methods: 
\code{\link{comments_comment}()},
\code{\link{comments_trending}()},
\code{\link{media_comments}},
\code{\link{user_comments}()},
\code{\link{user_list_comments}()}
}
\concept{comment methods}
