% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{convert_datetime}
\alias{convert_datetime}
\title{Quick datetime conversion}
\usage{
convert_datetime(object)
}
\arguments{
\item{object}{The input object. Must be \code{data.frame} or \code{list}}
}
\value{
The same object with converted datetimes
}
\description{
Searches for datetime variables and converts them to \code{POSIXct} via \pkg{lubridate}.
}
\keyword{internal}

