% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-internal.R
\docType{data}
\name{countries}
\alias{countries}
\title{Country codes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 213 rows and 2 columns.}
\usage{
countries
}
\description{
Country codes
}
\keyword{datasets}
