% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-dynamic-lists.R
\name{dynamic_lists}
\alias{dynamic_lists}
\title{See which movies / shows are popular across various metrics}
\arguments{
\item{start_date}{\code{character(1)}: A date in the past from which
on to count updates. If no date is supplied, the default is to
use yesterday relative to the current date. Value must either
be standard \code{YYYY-MM-DD} format or an object of class \link[base:Dates]{Date},
which will then be coerced via \link[base:as.character]{as.character()}.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
According to the API docs, popularity is calculated based both on ratings
and the number of ratings (i.e., votes). Trending items are those being
watched right now, where items with the most users currently watching are
returned first. Anticipation is measured by the number of user-created lists
an items is part of while not being released yet.
}
\section{The Dynamic Lists on trakt.tv}{


These functions access the automatically updated lists provided by trakt.tv.
Each function comes in two flavors: Shows or movies. The following descriptions
are adapted directly from the
\href{https://trakt.docs.apiary.io/#reference/movies/popular/get-popular-movies}{API reference}.
\itemize{
\item \link[=popular_media]{Popular}: Popularity is calculated using the rating percentage and the number of
ratings.
\item \link[=trending_media]{Trending}: Returns all movies/shows being watched right now.
Movies/shows with the most users are returned first.
\item \link[=played_media]{Played}: Returns the most played (a single user can watch multiple times)
movies/shows in the specified time \code{period}.
\item \link[=watched_media]{Watched}: Returns the most watched (unique users) movies/shows in the specified
time \code{period}.
\item \link[=collected_media]{Collected}: Returns the most collected (unique users) movies/shows in the
specified time \code{period}.
\item \link[=anticipated_media]{Anticipated}: Returns the most anticipated movies/shows based on the number of
lists a movie/show appears on.
\item \link[=updated_media]{Updates}: Returns all movies/shows updated since the specified UTC \code{start_date}.
In this case, the upper bound for \code{limit} is 100.
}

The functions for \strong{Played}, \strong{Watched}, \strong{Collected} and \strong{Played} each return
the same additional variables besides the media information: \code{watcher_count},
\code{play_count}, \code{collected_count}, \code{collector_count}.
}

\seealso{
\link{search_filters}, for more information on search filters.
}
