% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-common.R
\name{extended_info}
\alias{extended_info}
\title{Extended information}
\arguments{
\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\description{
Extended information
}
