% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract-media.R
\name{fix_datetime}
\alias{fix_datetime}
\title{Quick datetime conversion}
\usage{
fix_datetime(response)
}
\arguments{
\item{response}{The input object. Must be \code{data.frame}(ish) or named \code{list}.}
}
\value{
The same object with converted datetimes
}
\description{
Searches for datetime variables and converts them to \code{POSIXct} via \pkg{lubridate}.
}
\keyword{internal}
