% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fix_ids}
\alias{fix_ids}
\title{Fix IDs object}
\usage{
fix_ids(ids)
}
\description{
Fix IDs object
}
\keyword{Always}
\keyword{NA}
\keyword{NULL}
\keyword{characters,}
\keyword{explicit}
\keyword{internal}
\keyword{replace}
\keyword{return}
\keyword{with}
