% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{futurama}
\alias{futurama}
\title{Futurama episodes}
\format{A \link[tibble:tibble-package]{tibble} with 60 rows and 18 variables:
\describe{
\item{episode_abs}{Overall episode number}
\item{episode, season}{Episode within each season and season number}
\item{epid}{Episode ID in \code{s00e00} format}
\item{runtime}{Runtime in minutes}
\item{title}{Episode title}
\item{year}{Year of first airing}
\item{overview}{Episode summary}
\item{rating}{Rating (1-10) on  trakt.tv}
\item{votes}{Number of votes ontrakt.tv}
\item{first_aired, first_aired.string}{Original airdate in UTC as
\code{POSIXct} and \code{character}}
\item{trakt, tvdb, tmdb}{Episode IDs for trakt.tv, TVDb, and TMDb}
}}
\usage{
futurama
}
\description{
This data comes from \url{https://trakt.tv}.
}
\examples{
futurama
}
\seealso{
Other Episode datasets: \code{\link{gameofthrones}}
}
\concept{Episode datasets}
\keyword{datasets}
