% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gameofthrones}
\alias{gameofthrones}
\title{Game of Thrones episodes}
\format{A \link[tibble:tibble-package]{tibble} with 67 rows and 17 variables:
\describe{
\item{episode_abs}{Overall episode number}
\item{episode, season}{Episode within each season and season number}
\item{title}{Episode title}
\item{overview}{Episode summary}
\item{rating}{Rating (1-10) on  trakt.tv}
\item{votes}{Number of votes ontrakt.tv}
\item{viewers}{Viewers according to Wikipedia}
\item{director, writer}{Directing and writing credits}
\item{comment_count}{Number of comments on episode page}
\item{first_aired, updated_at}{Original air date and last update in UTC as \code{POSIXct}}
\item{runtime}{Runtime in minutes}
\item{trakt, tvdb, tmdb, tvrage}{Episode IDs for trakt.tv, TVDb, and TMDb}
\item{year}{Year of first airing}
\item{epid}{Episode ID in \code{s00e00} format}
}}
\usage{
gameofthrones
}
\description{
This data comes from \url{https://trakt.tv} and
\url{https://en.wikipedia.org/wiki/List_of_Game_of_Thrones_episodes}.
}
\examples{
gameofthrones
}
\seealso{
Other Episode datasets: \code{\link{futurama}}
}
\concept{Episode datasets}
\keyword{datasets}
