% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-utils.R
\name{get_client_secret}
\alias{get_client_secret}
\title{Get the client secret from the environment}
\usage{
get_client_secret()
}
\description{
Get's the client secrete either from the environment variable
\code{trakt_client_secret}, or from the environment variable \code{tRakt_key},
which holds an encryption key to decrypt the packaged encrypted secret.
}
\note{
See \url{https://httr2.r-lib.org/articles/wrapping-apis.html#package-keys-and-secrets}
}
\keyword{internal}
