% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-oauth.R
\name{get_token}
\alias{get_token}
\title{Get a trakt.tv API OAuth token}
\usage{
get_token(cache = TRUE)
}
\arguments{
\item{cache}{\code{\link{TRUE}}: Cache the token to the OS-specific cache directory. See \code{\link[rappdirs:user_cache_dir]{rappdirs::user_cache_dir()}}.}
}
\description{
This is an unfortunately home-brewed version of what \emph{should} be a simple call to \code{\link[httr2:req_oauth_device]{httr2::oauth_flow_device()}},
but since the <trakt.tv> API plays ever so slightly fast and mildly loose with RFC 8628, that's not possible.
}
\note{
RFC 8628 expects the device token request to have the field "device_code",
but the trakt.tv API expects a field named "code". That's it. It's kind of silly.
}
\examples{
if (FALSE) {

get_token(cache = TRUE)

}

}
