% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{get_trakt_credentials}
\alias{get_trakt_credentials}
\title{Set the required trakt.tv API credentials}
\usage{
get_trakt_credentials(apikey = NULL, username = NULL, set.headers = TRUE,
  client.id = NULL, client.secret = NULL)
}
\arguments{
\item{apikey}{Explicitly set your API key}

\item{username}{Explicitly set your trakt.tv username (Not used yet)}
}
\value{
Nothing
}
\description{
\code{get_trakt_credentials} searches for your credentials and stores them
in the appropriate \code{option} variables.
It also sets the HTTP header required for v2 API calls.
To make this work, place a \code{key.json} file either in the working directory
or in \code{~/.config/trakt/key.json}.
Arguments to this function take precedence over any key file.
}
\note{
This function includes both the old v1 API key as well
as the v2 API keys (client id and client secret).
Please note that no oauth2 methods are supported yet,
only client id really matters.
}
\examples{
\dontrun{
get_trakt_credentials()
}
}

