% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{getNameFromURL}
\alias{getNameFromURL}
\title{Get info from a show URL}
\usage{
getNameFromURL(url, epid = FALSE, getslug = FALSE)
}
\arguments{
\item{url}{Input URL. must be a \code{character}, but not a valid URL.}

\item{epid}{Whether the episode ID (\code{sXXeYY} format) should be extracted.
Defaults to \code{FALSE}.}

\item{getslug}{Whether the \code{slug} should be extracted. Defaults to \code{FALSE}.}
}
\value{
A \code{list} containing at least the show name.
}
\description{
\code{getNameFromURL} extracts some info from a show URL
}
\note{
This is pointless.
}
\examples{
getNameFromURL("http://trakt.tv/show/fargo/season/1/episode/2", T, T)
getNameFromURL("http://trakt.tv/show/breaking-bad", T, F)
}

