% Generated by roxygen2 (4.0.1): do not edit by hand
\name{\%+\%}
\alias{\%+\%}
\title{Easy string pasting}
\usage{
a \%+\% b
}
\arguments{
\item{a,b}{Character strings to paste together}
}
\value{
A \code{character} string.
}
\description{
\code{\%+\%} is an infix version of \code{paste0} or \code{paste(…, sep="")}
}
\note{
Blatantly copied from \href{http://adv-r.had.co.nz/Functions.html#infix-functions}{Hadley Wickham}
}
\examples{
\dontrun{
"string one" \%+\% " string two"
}
}

