% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-internal.R
\docType{data}
\name{languages}
\alias{languages}
\title{Language codes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 150 rows and 2 columns.}
\source{
\url{https://trakt.docs.apiary.io/#reference/languages/list/get-languages}
}
\usage{
languages
}
\description{
Language codes
}
\examples{
head(languages)
}
\keyword{datasets}
