% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_lists.R
\name{media_lists}
\alias{media_lists}
\alias{movies_lists}
\alias{shows_lists}
\alias{seasons_lists}
\alias{episodes_lists}
\alias{people_lists}
\title{Get lists containing a movie, show, season, episode or person}
\source{
\code{movies_lists()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/movies/lists/get-lists-containing-this-movie}{/movies/:id/lists/:type/:sort}.

\code{shows_lists()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/shows/lists/get-lists-containing-this-show}{/shows/:id/lists/:type/:sort}.

\code{seasons_lists()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/seasons/lists/get-lists-containing-this-season}{/shows/:id/seasons/:season/lists/:type/:sort}.

\code{episodes_lists()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/episodes/lists/get-lists-containing-this-episode}{/shows/:id/seasons/:season/episodes/:episode/lists/:type/:sort}.
}
\usage{
movies_lists(
  id,
  type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L,
  extended = "min"
)

shows_lists(
  id,
  type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L,
  extended = "min"
)

seasons_lists(
  id,
  season,
  type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L,
  extended = "min"
)

episodes_lists(
  id,
  season,
  episode,
  type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L,
  extended = "min"
)

people_lists(
  id,
  type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L,
  extended = "min"
)
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("tRakt")} for more details.}

\item{type}{\code{character(1) ["all"]}: The type of list, one of "all", "personal",
"official" or "watchlists".}

\item{sort}{\code{character(1) ["popular"]}: Sort lists by one of "popular", "likes",
"comments", "items", "added" or "updated".}

\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced via \code{as.integer()}.}

\item{extended}{\code{character}: Level of detail for the API response.
\itemize{
\item \code{"min"} (default): Minimal info (title, year, IDs). Omits the \code{extended} query param.
\item \code{"full"}: Complete info including overview, ratings, runtime, etc.
\item \code{"images"}: Minimal info plus image URLs (returned as a list-column).
\item \code{"full,images"}: Complete info plus images.
\item \code{"metadata"}: Collection endpoints only; adds video/audio metadata.
}

Multiple values can be combined as a comma-separated string (e.g. \code{"full,images"})
or a character vector (e.g. \code{c("full", "images")}).}

\item{season, episode}{\code{integer(1) [1L]}: The season and episode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get lists containing a movie, show, season, episode or person
}
\section{Functions}{
\itemize{
\item \code{movies_lists()}: Lists containing a movie.

\item \code{shows_lists()}: Lists containing a show.

\item \code{seasons_lists()}: Lists containing a season.

\item \code{episodes_lists()}: Lists containing an episode.

\item \code{people_lists()}: Lists containing a person.

}}
\examples{
\dontshow{if (trakt_api_available()) withAutoprint(\{ # examplesIf}
movies_lists("190430", type = "personal", limit = 5)
\dontshow{\}) # examplesIf}
\dontshow{if (trakt_api_available()) withAutoprint(\{ # examplesIf}
shows_lists("46241")
\dontshow{\}) # examplesIf}
\dontshow{if (trakt_api_available()) withAutoprint(\{ # examplesIf}
seasons_lists("46241", season = 1)
\dontshow{\}) # examplesIf}
\dontshow{if (trakt_api_available()) withAutoprint(\{ # examplesIf}
episodes_lists("46241", season = 1, episode = 1)
\dontshow{\}) # examplesIf}
\dontshow{if (trakt_api_available()) withAutoprint(\{ # examplesIf}
people_lists("david-tennant")

people_lists("emilia-clarke", sort = "items")
\dontshow{\}) # examplesIf}
}
\seealso{
Other list methods:
\code{\link[=lists_popular]{lists_popular()}},
\code{\link[=user_list]{user_list()}},
\code{\link[=user_list_comments]{user_list_comments()}},
\code{\link[=user_list_items]{user_list_items()}},
\code{\link[=user_lists]{user_lists()}}

Other movie data:
\code{\link{anticipated_media}},
\code{\link{collected_media}},
\code{\link{media_aliases}},
\code{\link{media_comments}},
\code{\link{media_people}},
\code{\link[=media_ratings]{media_ratings()}},
\code{\link[=media_stats]{media_stats()}},
\code{\link{media_translations}},
\code{\link{media_watching}},
\code{\link[=movies_boxoffice]{movies_boxoffice()}},
\code{\link[=movies_related]{movies_related()}},
\code{\link[=movies_releases]{movies_releases()}},
\code{\link[=movies_summary]{movies_summary()}},
\code{\link[=people_media]{people_media()}},
\code{\link{played_media}},
\code{\link{popular_media}},
\code{\link{trending_media}},
\code{\link{watched_media}}

Other show data:
\code{\link{collected_media}},
\code{\link{media_aliases}},
\code{\link{media_comments}},
\code{\link{media_people}},
\code{\link[=media_ratings]{media_ratings()}},
\code{\link[=media_stats]{media_stats()}},
\code{\link{media_translations}},
\code{\link{media_watching}},
\code{\link[=people_media]{people_media()}},
\code{\link{played_media}},
\code{\link[=shows_next_episode]{shows_next_episode()}},
\code{\link[=shows_related]{shows_related()}},
\code{\link[=shows_summary]{shows_summary()}}

Other season data:
\code{\link{media_comments}},
\code{\link{media_people}},
\code{\link[=media_ratings]{media_ratings()}},
\code{\link[=media_stats]{media_stats()}},
\code{\link[=seasons_episodes]{seasons_episodes()}},
\code{\link[=seasons_season]{seasons_season()}},
\code{\link[=seasons_summary]{seasons_summary()}}

Other episode data:
\code{\link[=episodes_summary]{episodes_summary()}},
\code{\link{media_comments}},
\code{\link{media_people}},
\code{\link[=media_ratings]{media_ratings()}},
\code{\link[=media_stats]{media_stats()}},
\code{\link{media_translations}},
\code{\link{media_watching}},
\code{\link[=seasons_episodes]{seasons_episodes()}},
\code{\link[=seasons_summary]{seasons_summary()}},
\code{\link[=shows_next_episode]{shows_next_episode()}}

Other people data:
\code{\link{media_people}},
\code{\link[=people_media]{people_media()}},
\code{\link[=people_summary]{people_summary()}}
}
\concept{episode data}
\concept{list methods}
\concept{movie data}
\concept{people data}
\concept{season data}
\concept{show data}
