% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_people.R
\name{media_people}
\alias{media_people}
\alias{shows_people}
\alias{movies_people}
\title{Get the cast and crew of a show or movie}
\usage{
shows_people(target, extended = c("min", "full"))

movies_people(target, extended = c("min", "full"))
}
\arguments{
\item{target}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and the result is combined.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \code{list} of one or more \link[tibble:tibble-package]{tibbles} for \code{cast}
and \code{crew}. The latter \code{tibble} objects are as flat as possible.
}
\description{
Returns all cast and crew for a show/movie, depending on how much data is
available.
}
\details{
The API methods for these functions are:
\itemize{
\item \href{https://trakt.docs.apiary.io/#reference/shows/people/get-all-people-for-a-show}{/shows/:id/people}
\item \href{https://trakt.docs.apiary.io/#reference/movies/people/get-all-people-for-a-movie}{/movies/:id/people}
}

Note that as of 2019-09-30, there are two representations of \code{character[s]} and \code{job[s]}:
One is a regular character variable, and the other is a list-column. The former is
\href{https://github.com/trakt/api-help/issues/74}{deprecated and only included for compatibility reasons}.
}
\examples{
\dontrun{
shows_people("breaking-bad")
}
}
\seealso{
\link{people_media}, for the other direction: People that have credits in shows/movies.

Other people data: \code{\link{people_media}},
  \code{\link{people_summary}}
}
\concept{people data}
