% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-media-summary.R
\name{media_summary}
\alias{media_summary}
\alias{trakt.movies.summary}
\alias{trakt.shows.summary}
\title{Get show or movie summary info}
\usage{
trakt.movies.summary(target, extended = c("min", "full"))

trakt.shows.summary(target, extended = c("min", "full"))
}
\arguments{
\item{target}{\code{character(1)}: The ID of the item requested. Preferably the
\code{Trakt ID} (e.g. \code{1390}), \code{slug} (e.g. \code{"game-of-thrones"}), or \code{IMDb ID} (e.g. \code{"tt0944947"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and the result is combined.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Get show or movie summary info
}
\details{
These functions wrap the "single item" API methods:
\itemize{
\item \href{https://trakt.docs.apiary.io/#reference/shows/summary/get-a-single-show}{/shows/:id}
\item \href{https://trakt.docs.apiary.io/#reference/movies/summary/get-a-movie}{/movies/:id}
}
}
\note{
Thes function returns the same amount of informations as \link{trakt.search}.
}
\examples{
# Minimal info by default
trakt.shows.summary("breaking-bad")
\dontrun{
# More information
trakt.shows.summary("breaking-bad", extended = "full")

# Info for multiple movies
trakt.movies.summary(c("inception-2010", "the-dark-knight-2008"), extended = "full")
}
}
\concept{summary data}
