% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_watching.R
\name{media_watching}
\alias{media_watching}
\alias{movies_watching}
\alias{shows_watching}
\alias{seasons_watching}
\alias{episodes_watching}
\title{Get who's watching a thing right now}
\source{
\code{movies_watching()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/movies/watching/get-users-watching-right-now}{/movies/:id/watching}.

\code{shows_watching()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/shows/watching/get-users-watching-right-now}{/shows/:id/watching}.

\code{episodes_watching()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/episodes/watching/get-users-watching-right-now}{/shows/:id/seasons/:season/episodes/:episode/watching}.
}
\usage{
movies_watching(id, extended = "min")

shows_watching(id, extended = "min")

seasons_watching(id, season = 1L, extended = "min")

episodes_watching(id, season = 1L, episode = 1L, extended = "min")
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("tRakt")} for more details.}

\item{extended}{\code{character}: Level of detail for the API response.
\itemize{
\item \code{"min"} (default): Minimal info (title, year, IDs). Omits the \code{extended} query param.
\item \code{"full"}: Complete info including overview, ratings, runtime, etc.
\item \code{"images"}: Minimal info plus image URLs (returned as a list-column).
\item \code{"full,images"}: Complete info plus images.
\item \code{"metadata"}: Collection endpoints only; adds video/audio metadata.
}

Multiple values can be combined as a comma-separated string (e.g. \code{"full,images"})
or a character vector (e.g. \code{c("full", "images")}).}

\item{season, episode}{\code{integer(1) [1L]}: The season and episode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get who's watching a thing right now
}
\section{Functions}{
\itemize{
\item \code{movies_watching()}: Who's watching a movie.

\item \code{shows_watching()}: Who's watching a show.

\item \code{seasons_watching()}: Who's watching a season.

\item \code{episodes_watching()}: Who's watching an episode.

}}
\examples{
\dontrun{
movies_watching("deadpool-2016")
shows_watching("the-simpsons")
seasons_watching("the-simpsons", season = 6)
episodes_watching("the-simpsons", season = 6, episode = 12)
}
}
\seealso{
Other movie data: 
\code{\link{anticipated_media}},
\code{\link{collected_media}},
\code{\link{media_aliases}},
\code{\link{media_comments}},
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{media_ratings}()},
\code{\link{media_stats}()},
\code{\link{media_translations}},
\code{\link{movies_boxoffice}()},
\code{\link{movies_related}()},
\code{\link{movies_releases}()},
\code{\link{movies_summary}()},
\code{\link{people_media}()},
\code{\link{played_media}},
\code{\link{popular_media}},
\code{\link{trending_media}},
\code{\link{watched_media}}

Other show data: 
\code{\link{collected_media}},
\code{\link{media_aliases}},
\code{\link{media_comments}},
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{media_ratings}()},
\code{\link{media_stats}()},
\code{\link{media_translations}},
\code{\link{people_media}()},
\code{\link{played_media}},
\code{\link{shows_next_episode}()},
\code{\link{shows_related}()},
\code{\link{shows_summary}()}

Other episode data: 
\code{\link{episodes_summary}()},
\code{\link{media_comments}},
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{media_ratings}()},
\code{\link{media_stats}()},
\code{\link{media_translations}},
\code{\link{seasons_episodes}()},
\code{\link{seasons_summary}()},
\code{\link{shows_next_episode}()}
}
\concept{episode data}
\concept{movie data}
\concept{show data}
