% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_aliases.R
\name{movies_aliases}
\alias{movies_aliases}
\alias{shows_aliases}
\title{Get all movie / show aliases}
\usage{
movies_aliases(id)

shows_aliases(id)
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get all movie / show aliases
}
\examples{
movies_aliases(190430)
shows_aliases(104439)
}
