% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_releases.R
\name{movies_releases}
\alias{movies_releases}
\title{Get a movie's release details}
\usage{
movies_releases(id, country = NULL)
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined.}

\item{country}{Optional two letter country code to filter by.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Retrieve one or more movie's release information,
including the release date, country code (two letter, e.g. \code{us}), and
the certification (e.g. \code{PG}).
}
\details{
This function wraps the endpoint
\href{http://docs.trakt.apiary.io/reference/movies/releases/get-all-movie-releases}{movies/:id/releases/:country}
}
\examples{
\dontrun{
movies_releases("tron-legacy-2010")
}
}
\seealso{
Other movie data: \code{\link{movies_boxoffice}}
}
\concept{movie data}
