% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_translations.R
\name{movies_translations}
\alias{movies_translations}
\alias{shows_translations}
\alias{episodes_translations}
\title{Get translations for a movie, show or episode}
\usage{
movies_translations(id, languages = NULL)

shows_translations(id, languages = NULL)

episodes_translations(id, season = 1L, episode = 1L,
  languages = NULL)
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}

\item{languages}{\code{character(n)}: Two-digit language code(s).
Also see \code{\link{trakt_languages}} for table of available languages.}

\item{season}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}

\item{episode}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Get translations for a movie, show or episode
}
\examples{
# Get all translations
movies_translations("193972")

# Only get a specific language
movies_translations("193972", "de")
}
