% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-internal.R
\docType{data}
\name{networks}
\alias{networks}
\title{Available networks}
\format{An object of class \code{character} of length 1526.}
\source{
\url{https://trakt.docs.apiary.io/#reference/networks/list/get-networks}
}
\usage{
networks
}
\description{
Available networks
}
\examples{
head(networks)
}
\keyword{datasets}
