% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pad}
\alias{pad}
\title{Easy episode id padding}
\usage{
pad(x, width = 2)
}
\arguments{
\item{x}{Input episode/season number, coerced to \code{character}.}

\item{width}{The length of the padding. Defaults to 2.}
}
\value{
A \code{character} in standard \code{sXXeYY} format
}
\description{
\code{pad} pulls show stats and returns it compactly.
}
\details{
Simple function to ease the creation of \code{sXXeYY} episode ids.
}
\note{
See \href{http://trakt.tv/api-docs/show-stats}{the trakt API docs for further info}
}
\examples{
\dontrun{
season  <- 5
episode <- 1
paste0("s", pad(season), "e", pad(episode))

pad(c(1,2,3))
}
}

