% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{pad_episode}
\alias{pad_episode}
\title{Easy episode number padding}
\usage{
pad_episode(s = "0", e = "0", s_width = 2, e_width = 2)
}
\arguments{
\item{s}{Input season number, coerced to \code{character}.}

\item{e}{Input episode number, coerced to \code{character}.}

\item{s_width}{The length of the season number padding. Defaults to 2.}

\item{e_width}{The length of the episode number padding. Defaults to 2.}
}
\value{
A \code{character} in the common \code{sXXeYY} format
}
\description{
Simple function to ease the creation of \code{sXXeYY} episode ids.
Note that \code{s} and \code{e} must have the same length.
}
\note{
I like my sXXeYY format, okay?
}
\examples{
# Season 2, episode 4
pad_episode(2, 4)
pad_episode(1, 85, e_width = 3)
}
\seealso{
Other utility functions: 
\code{\link{build_trakt_url}()}
}
\concept{utility functions}
