% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{parse_trakt_url}
\alias{parse_trakt_url}
\title{Get info from a show URL}
\usage{
parse_trakt_url(url, epid = FALSE, getslug = FALSE)
}
\arguments{
\item{url}{Input URL. must be a \code{character}, but not a valid URL.}

\item{epid}{Whether the episode ID (\code{sXXeYY} format) should be extracted.
Defaults to \code{FALSE}.}

\item{getslug}{Whether the \code{slug} should be extracted. Defaults to \code{FALSE}.}
}
\value{
A \code{list} containing at least the show name.
}
\description{
\code{parse_trakt_url} extracts some info from a show URL
}
\note{
This is pointless.
}
\examples{
parse_trakt_url("http://trakt.tv/show/fargo/season/1/episode/2", TRUE, TRUE)
parse_trakt_url("http://trakt.tv/show/breaking-bad", TRUE, FALSE)
}
\seealso{
Other utility functions: \code{\link{build_trakt_url}};
  \code{\link{pad}}
}

