% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-filters.R
\name{search_filters}
\alias{search_filters}
\title{Result Filters}
\arguments{
\item{query}{\code{character(1)}: Search titles and descriptions.}

\item{years}{\code{character | integer}: 4-digit year (\code{2010}) \strong{or} range, e.g. \code{"2010-2020"}.
Can also be an integer vector of length two which will be coerced appropriately.}

\item{genres}{\code{character(n)}: Genre  slug(s). See \code{\link{trakt_genres}} for a table of genres.
Multiple values are allowed and will be concatenated.}

\item{languages}{\code{character(n)}: Two-digit language code(s). Also see \code{\link{trakt_languages}} for
table of available languages.}

\item{countries}{\code{character(n)}: Two-character country code(s). See \code{\link{trakt_countries}}.}

\item{runtimes}{\code{character | integer}: Integer range in minutes, e.g. \code{30-90}. Can also
be an integer vector of length two which will be coerced appropriately.}

\item{ratings}{\code{character | integer}:  Integer range between \code{0} and \code{100}. Can also be
an integer vector of length two which will be coerced appropriately.}

\item{certifications}{\code{character(n)}: Certification(s) like \code{pg-13}.
Multiple values are allowed. Use \code{\link{trakt_certifications}} for reference. Note that there
are different certifications for shows and movies.}

\item{networks}{\code{character(n)}: (Shows only) Network name like \code{HBO}. See \code{\link{trakt_networks}}
for a list of known networks.}

\item{status}{\code{character(n)}: (Shows only) The status of the shows. One of \code{returning series},
\code{in production}, \code{planned}, \code{canceled}, or \code{ended}}
}
\description{
You can specifiy additional search filters in some functions. These filters are
documented \href{https://trakt.docs.apiary.io/#introduction/filters}{in the official API reference}.
For convenience, all supported filters are implemented as explicit, named arguments.
}
\seealso{
\link{dynamic_lists}, where these parameters are used
}
