% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shows_next_episode.R
\name{shows_next_episode}
\alias{shows_next_episode}
\alias{shows_last_episode}
\title{Get a shows next or latest episode}
\source{
\code{shows_next_episode()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/shows/next-episode/get-next-episode}{/shows/:id/next_episode}.

\code{shows_last_episode()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/shows/last-episode/get-last-episode}{/shows/:id/last_episode}.
}
\usage{
shows_next_episode(id, extended = "min")

shows_last_episode(id, extended = "min")
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("tRakt")} for more details.}

\item{extended}{\code{character}: Level of detail for the API response.
\itemize{
\item \code{"min"} (default): Minimal info (title, year, IDs). Omits the \code{extended} query param.
\item \code{"full"}: Complete info including overview, ratings, runtime, etc.
\item \code{"images"}: Minimal info plus image URLs (returned as a list-column).
\item \code{"full,images"}: Complete info plus images.
\item \code{"metadata"}: Collection endpoints only; adds video/audio metadata.
}

Multiple values can be combined as a comma-separated string (e.g. \code{"full,images"})
or a character vector (e.g. \code{c("full", "images")}).}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get a shows next or latest episode
}
\examples{
shows_next_episode("one-piece")
shows_last_episode("one-piece", extended = "full")
}
\seealso{
Other show data: 
\code{\link{collected_media}},
\code{\link{media_aliases}},
\code{\link{media_comments}},
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{media_ratings}()},
\code{\link{media_stats}()},
\code{\link{media_translations}},
\code{\link{media_watching}},
\code{\link{people_media}()},
\code{\link{played_media}},
\code{\link{shows_related}()},
\code{\link{shows_summary}()}

Other episode data: 
\code{\link{episodes_summary}()},
\code{\link{media_comments}},
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{media_ratings}()},
\code{\link{media_stats}()},
\code{\link{media_translations}},
\code{\link{media_watching}},
\code{\link{seasons_episodes}()},
\code{\link{seasons_summary}()}

Other show data: 
\code{\link{collected_media}},
\code{\link{media_aliases}},
\code{\link{media_comments}},
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{media_ratings}()},
\code{\link{media_stats}()},
\code{\link{media_translations}},
\code{\link{media_watching}},
\code{\link{people_media}()},
\code{\link{played_media}},
\code{\link{shows_related}()},
\code{\link{shows_summary}()}

Other episode data: 
\code{\link{episodes_summary}()},
\code{\link{media_comments}},
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{media_ratings}()},
\code{\link{media_stats}()},
\code{\link{media_translations}},
\code{\link{media_watching}},
\code{\link{seasons_episodes}()},
\code{\link{seasons_summary}()}
}
\concept{episode data}
\concept{show data}
