% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.get_all_episodes.R
\name{trakt.get_all_episodes}
\alias{trakt.get_all_episodes}
\title{Get a show's episodes. All of them.}
\usage{
trakt.get_all_episodes(target, season_nums = NULL, extended = "full",
  drop.unaired = TRUE, drop.translations = TRUE)
}
\arguments{
\item{target}{The \code{id} of the show requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"}), \code{trakt id} or \code{IMDb id}.}

\item{season_nums}{Vector of season numbers, e.g. \code{c(1, 2)}. If \code{NULL}, all the seasons
are retrieved by calling \link{trakt.seasons.summary} to determine the number of seasons.
If a vector of length 1 (e.g. \code{5}) is supplied, it is extended to \code{seq_len(season_nums)}.}

\item{extended}{Use \code{full} more data. Can be \code{min}, \code{full} (default).}

\item{drop.unaired}{\code{logical(1) [TRUE]}: Episodes which have not aired yet are dropped.}

\item{drop.translations}{\code{logical(1) [TRUE]}: Remove list-column containing country-
codes for available translation. This column is unlikely to be of interest and
therefore excluded by default.}
}
\value{
A \code{tbl} containing episode details
}
\description{
\code{trakt.get_all_episodes} pulls detailed episode data.
Get details for a show's episodes, e.g. ratings, number of votes,
airdates, images, plot overviews…
}
\details{
This is basically just an extension of \link[tRakt]{trakt.seasons.season}, which is used in
this function to collect all the episode data.
If you only want the episode data for a single season anyway, \code{trakt.seasons.season}
is recommended, yet this function makes some additions.
The use case of this function is to quickly gather episode data of all seasons of a show.
}
\note{
This function is mainly for convenience.
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
# Manually specifiy seasons
breakingbad.episodes <- trakt.get_all_episodes("breaking-bad", season_nums = c(1, 2, 3, 4, 5))
# Get all the seasons
breakingbad.episodes <- trakt.get_all_episodes("breaking-bad")
# Get first to 3rd season
breakingbad.episodes <- trakt.get_all_episodes("breaking-bad", season_nums = 3)
}
}
\seealso{
Other show data: \code{\link{trakt.get_full_showdata}},
  \code{\link{trakt.seasons.season}},
  \code{\link{trakt.seasons.summary}},
  \code{\link{trakt.show.people}},
  \code{\link{trakt.show.ratings}},
  \code{\link{trakt.show.summary}},
  \code{\link{trakt.show.watching}},
  \code{\link{trakt.shows.popular}},
  \code{\link{trakt.shows.related}},
  \code{\link{trakt.shows.trending}},
  \code{\link{trakt.stats}}
}
\concept{show data}
