% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.get_full_showdata.R
\name{trakt.get_full_showdata}
\alias{trakt.get_full_showdata}
\title{Get all the show data}
\usage{
trakt.get_full_showdata(query = NULL, slug = NULL,
  drop.unaired = TRUE)
}
\arguments{
\item{query}{Keyword used for \link{trakt.search}. Optional.}

\item{slug}{Used if \code{query} is not specified. Optional, but gives exact results.}

\item{drop.unaired}{If \code{TRUE}, episodes which have not aired yet are dropped.}
}
\value{
A \code{list} containing multiple \code{lists} and \code{tibble}s with show info.
}
\description{
\code{trakt.get_full_showdata} is a combination function of multiple
functions in this package. The idea is to easily execute all major functions
required to get a full show dataset. This also uses \code{extended = "full"}, so
beware that this is going to retrieve \emph{a lot} of data. Please be nice to the API.
}
\note{
This is primarily intended to be a convenience function for the
case where you really want all that data. Usually it is advisable to only
call the specific functions for the data you really need.
}
\examples{
\dontrun{
# Use the search within the function
breakingbad <- trakt.get_full_showdata("Breaking Bad")

# Alternatively, us a slug for explicit results
breakingbad <- trakt.get_full_showdata(slug = "breaking-bad")
}
}
\seealso{
Other show data: \code{\link{trakt.get_all_episodes}},
  \code{\link{trakt.seasons.season}},
  \code{\link{trakt.seasons.summary}}
}
\concept{show data}
