% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trakt.getEpisodeData2}
\alias{trakt.getEpisodeData2}
\title{Get a show's episode information}
\usage{
trakt.getEpisodeData2(target, season_nums = NULL,
  apikey = getOption("trakt.apikey"), dropunaired = TRUE)
}
\arguments{
\item{target}{The \code{slug} or \code{tvdbid} of the show requested}

\item{season_nums}{Vector of season numbers.}

\item{apikey}{API-key used for the call.
Defaults to \code{getOption("trakt.apikey")}}

\item{dropunaired}{If \code{TRUE}, episodes which have not aired yet are dropped.}
}
\value{
A \code{data.frame} containing episode details
}
\description{
\code{trakt.getEpisodeData2} pulls detailed episode data (differently).
Get details for a show's episodes, e.g. ratings, number of votes,
airdates, urls, plot overviews…
}
\note{
This function uses much less API calls than \link{trakt.getEpisodeData}.
}
\examples{
\dontrun{
options(trakt.apikey = jsonlite::fromJSON("key.json")$apikey)
breakingbad.seasons  <- trakt.getSeasons("breaking-bad")
breakingbad.episodes <- trakt.getEpisodeData2("breaking-bad", breakingbad.seasons)
}
}

