% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trakt.getFullShowData}
\alias{trakt.getFullShowData}
\title{Get all the show data}
\usage{
trakt.getFullShowData(searchquery = NULL, tvdb_id = NULL,
  dropunaired = TRUE)
}
\arguments{
\item{searchquery}{Keyword used for \link{trakt.search}. Optional.}

\item{tvdb_id}{Used if \code{searchquery} is not specified. Optional.}

\item{dropunaired}{If \code{TRUE}, episodes which have not aired yet are dropped.}
}
\value{
A \code{list} containing multiple \code{lists} and \code{data.frames} with show info.
}
\description{
\code{trakt.getFullShowData} is a combination function of multiple
functions in this package. The idea is to easily execute all major functions
required to get a full show dataset.
}
\details{
It's assumed that \code{getOption("trakt.apikey")} is already set, otherwise
the function will stop and tell you to.
}
\note{
This is primarily intended to be a convenience function for the case where you
really want all that data. If you're just derping around, maybe you should consider interactively
calling the other functions.
}
\examples{
\dontrun{
options(trakt.apikey = jsonlite::fromJSON("key.json")$apikey)
breakingbad.seasons  <- trakt.getFullShowData("Breaking Bad")
}
}

