% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trakt.getSeasons}
\alias{trakt.getSeasons}
\title{Get a show's season information}
\usage{
trakt.getSeasons(target, dropspecials = TRUE,
  apikey = getOption("trakt.apikey"))
}
\arguments{
\item{target}{The \code{slug} or \code{tvdbid} of the show requested}

\item{apikey}{API-key used for the call. Defaults to \code{getOption("trakt.apikey")}}

\item{dropspecials}{If \code{TRUE}, special episodes (listed as 'season 0') are ignored. This is the default.}
}
\value{
A \code{data.frame} containing season details
}
\description{
\code{trakt.getSeasons} pulls season data.
Get details for a show's seasons, e.g. how many seasons there are, how many epsiodes
each season has, and seaosn posters. Unfortunately, trakt does not offer per-season ratings.
}
\note{
See \href{http://trakt.tv/api-docs/show-season}{the trakt API docs for further info}
}
\examples{
\dontrun{
options(trakt.apikey = jsonlite::fromJSON("key.json")$apikey)
breakingbad.seasons <- trakt.getSeasons("breaking-bad")
}
}

