% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.media.stats.R
\name{trakt.media.stats}
\alias{trakt.media.stats}
\alias{trakt.shows.stats}
\alias{trakt.movies.stats}
\title{Get a show or movie's stats}
\usage{
trakt.media.stats(type = c("shows", "movies"), target)

trakt.shows.stats(target)

trakt.movies.stats(target)
}
\arguments{
\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}

\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"} for a show, or \code{"bryan-cranston"} for a person),
\code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"} or \code{"nm0186505"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df} if possible.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
The data includes show/movie ratings, scrobbles, checkins, plays, comments...
}
\examples{
\dontrun{
trakt.shows.stats("breaking-bad")
}
}
