% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.media.summary.R
\name{trakt.media.summary}
\alias{trakt.media.summary}
\alias{trakt.movies.summary}
\alias{trakt.shows.summary}
\title{Get show or movie summary info}
\usage{
trakt.media.summary(type, target, extended = c("min", "full"))

trakt.movies.summary(target, extended = c("min", "full"))

trakt.shows.summary(target, extended = c("min", "full"))
}
\arguments{
\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}

\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"} for a show, or \code{"bryan-cranston"} for a person),
\code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"} or \code{"nm0186505"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df} if possible.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Note that setting \code{extended} to \code{min} makes this function
return about as much informations as \link[tRakt]{trakt.search}
}
\examples{
\dontrun{
# Minimal info by default
trakt.shows.summary("breaking-bad")

# More information
trakt.shows.summary("breaking-bad", extended = "full")

# Info for multiple movies
trakt.movies.summary(c("inception-2010", "the-dark-knight-2008"), extended = "full")
}
}
\concept{summary data}
