% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.people.R
\name{trakt.movie.people}
\alias{trakt.movie.people}
\title{Get the cast and crew of a movie}
\usage{
trakt.movie.people(target, extended = "min")
}
\arguments{
\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"}), \code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \code{list}.
}
\description{
\code{trakt.movie.people} pulls movie people data.
}
\details{
Returns all cast and crew for a movie, depending on how much data is available.
}
\note{
See \href{http://docs.trakt.apiary.io/#reference/movies/people/get-all-people-for-a-movie}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
tron.people <- trakt.movie.people("tron-legacy-2010")
}
}
\seealso{
Other movie data: \code{\link{trakt.movie.releases}},
  \code{\link{trakt.movie.summary}},
  \code{\link{trakt.stats}}

Other people data: \code{\link{trakt.people.movies}},
  \code{\link{trakt.people.shows}},
  \code{\link{trakt.people.summary}},
  \code{\link{trakt.show.people}}
}
\concept{movie data}
\concept{people data}
