% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trakt.summary.R
\name{trakt.movie.summary}
\alias{trakt.movie.summary}
\title{Get a single movie's details}
\usage{
trakt.movie.summary(target, extended = "min")
}
\arguments{
\item{target}{The \code{id} of the movie requested. Either the \code{slug}
(e.g. \code{"tron-legacy-2010"}), \code{trakt id} or \code{IMDb id}}

\item{extended}{Whether extended info should be provided.
Defaults to \code{"min"}, can either be \code{"min"} or \code{"full"}}
}
\value{
A \code{list} containing movie information
}
\description{
\code{trakt.movie.summary} returns a single movie's summary information.
}
\note{
See \href{http://docs.trakt.apiary.io/reference/movies/summary/get-a-movie}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
trakt.movie.summary("tron-legacy-2010")
}
}
\seealso{
Other movie: \code{\link{trakt.movie.people}};
  \code{\link{trakt.movie.ratings}};
  \code{\link{trakt.movie.releases}};
  \code{\link{trakt.movie.watching}};
  \code{\link{trakt.movies.popular}};
  \code{\link{trakt.movies.related}};
  \code{\link{trakt.movies.trending}}
}

