% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-movies-releases-boxoffice.R
\name{trakt.movies.boxoffice}
\alias{trakt.movies.boxoffice}
\title{Get the weekend box office}
\usage{
trakt.movies.boxoffice(extended = c("min", "full"))
}
\arguments{
\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Returns the top 10 grossing movies in the U.S. box office last weekend.
Updated every Monday morning.
}
\details{
This function wraps
\href{https://trakt.docs.apiary.io/#reference/movies/box-office/get-the-weekend-box-office}{movies/boxoffice}.
}
\examples{
trakt.movies.boxoffice()
}
\seealso{
Other movie data: \code{\link{trakt.movies.releases}}
}
\concept{movie data}
