% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.popular.R
\name{trakt.movies.popular}
\alias{trakt.movies.popular}
\alias{trakt.popular}
\alias{trakt.shows.popular}
\title{Get popular movies}
\usage{
trakt.movies.popular(limit = 10, page = 1, extended = "min")

trakt.shows.popular(limit = 10, page = 1, extended = "min")
}
\arguments{
\item{limit}{Number of movies to return. Is coerced to \code{integer} and must be greater than 0.}

\item{page}{Page to return (default is \code{1})
for \href{http://docs.trakt.apiary.io/#introduction/pagination}{pagination}.}

\item{extended}{\code{[character(1)]}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
\code{trakt.[movies|shows].popular} returns a list of the most popular movies on trakt.tv.
According to the API docs, opularity is calculated based both ratings and number of ratings.
}
\section{Functions}{
\itemize{
\item \code{trakt.shows.popular}: Get popular shows
}}

\examples{
\dontrun{
trakt.movies.popular(5)
trakt.shows.popular(5)
}
}
\concept{Automated lists}
