% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.people.R
\name{trakt.people.shows}
\alias{trakt.people.shows}
\title{Get a single person's show credits}
\usage{
trakt.people.shows(target, extended = c("min", "full"))
}
\arguments{
\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"}), \code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \code{list}.
}
\description{
\code{trakt.people.shows} pulls show people data.
}
\details{
Returns all shows where this person is in the cast or crew.
}
\note{
See \href{http://docs.trakt.apiary.io/reference/people/shows/get-show-credits}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
person <- trakt.people.shows("bryan-cranston")
}
}
\seealso{
Other people data: \code{\link{trakt.movie.people}},
  \code{\link{trakt.people.movies}},
  \code{\link{trakt.people.summary}},
  \code{\link{trakt.show.people}}
}
\concept{people data}
