% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.popular-trending.R
\name{trakt.popular}
\alias{trakt.popular}
\alias{trakt.trending}
\alias{trakt.movies.trending}
\alias{trakt.shows.trending}
\alias{trakt.movies.popular}
\alias{trakt.shows.popular}
\title{Get popular and trending movies and shows}
\usage{
trakt.popular(type, limit = 10, page = 1, extended = c("min",
  "full"))

trakt.trending(type, limit = 10, page = 1, extended = c("min",
  "full"))

trakt.movies.trending(limit = 10, page = 1, extended = c("min",
  "full"))

trakt.shows.trending(limit = 10, page = 1, extended = c("min",
  "full"))

trakt.movies.popular(limit = 10, page = 1, extended = c("min",
  "full"))

trakt.shows.popular(limit = 10, page = 1, extended = c("min",
  "full"))
}
\arguments{
\item{type}{\code{[character(1)]}: Either \code{"show"} or \code{"movie"}.}

\item{limit}{Number of movies/shows to return. Is coerced to \code{integer} and must be
greater than 0.}

\item{page}{Page to return (default is \code{1})
for \href{http://docs.trakt.apiary.io/#introduction/pagination}{pagination}.}

\item{extended}{\code{[character(1)]}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
According to the API docs, popularity is calculated based both ratings
and number of ratings. Trending items are those being watched right now, where
items with the most users watching are returned first.
}
\examples{
\dontrun{
# Get popular movies and shows
trakt.movies.popular(5)
trakt.shows.popular(5)

# Get trending movies and shows
trakt.movies.trending(5)
trakt.shows.trending(5)
}
}
\concept{Automated lists}
