% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.ratings.R
\name{trakt.ratings}
\alias{trakt.ratings}
\alias{trakt.shows.ratings}
\alias{trakt.movies.ratings}
\title{Get show or movie user ratings}
\usage{
trakt.ratings(type = c("shows", "movies"), target)

trakt.shows.ratings(target)

trakt.movies.ratings(target)
}
\arguments{
\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}

\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"}), \code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df}.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Returns a show's or movie's rating and ratings distribution.
}
\note{
See \href{http://docs.trakt.apiary.io/reference/shows/ratings/get-show-ratings}{the trakt API docs for further info}
}
\examples{
\dontrun{
trakt.movies.ratings("tron-legacy-2010")
trakt.shows.ratings("game-of-thrones")
}
}
