% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{tRakt}
\alias{tRakt}
\alias{tRakt-package}
\title{Pull data from trakt.tv}
\description{
tRakt is a package that grew out of my initial attempt to compare Stargate SG-1 episode ratings.
\href{http://stuff.wurstmannberg.de/tRakt}{Here are some of the resulting graphs}.
This package is meant to provide the basic functions required to pull series data, which
then can be used either for interactive use and standalone graphs, or for more complex applications.
}

