% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trakt.related.show}
\alias{trakt.related.show}
\title{Search for related shows}
\usage{
trakt.related.show(target, apikey = getOption("trakt.apikey"))
}
\arguments{
\item{target}{The show's tvdb id to be used.}

\item{apikey}{API-key used for the call. Defaults to \code{getOption("trakt.apikey")}}
}
\value{
A \code{data.frame} containing search results
}
\description{
\code{trakt.related.show} returns related shows to the search query.
}
\details{
Search for related shows to the show you specified.
}
\note{
See \href{http://trakt.tv/api-docs/show-related}{the trakt API docs for further info}
}
\examples{
\dontrun{
options(trakt.apikey = jsonlite::fromJSON("key.json")$apikey)
related <- trakt.related.show(78874)
}
}

