% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.search.R
\name{trakt.search}
\alias{trakt.search}
\title{Search for a show via text query}
\usage{
trakt.search(query, type = "show", year = NULL)
}
\arguments{
\item{query}{The keyword used for the search. If the query ends with a 4 digit numbers,
this will be used as \code{year} parameter.}

\item{type}{The type of data you're looking for. Defaults to \code{show}, can also be
\code{movie}, \code{episode}, \code{person} or \code{list}.}

\item{year}{Optionally filter by year.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing a single search result.
If no results are found, the \code{tibble} has 0 rows.
}
\description{
\code{trakt.search} returns search result info.
}
\details{
Search for a show or movie with a keyword (e.g. \code{"Breaking Bad"}) and receive basic info of the first
search result. It's main use is to retrieve the ids or proper show/movie title for further use, as well
as receiving a quick overview of a show/movie.
}
\note{
See \href{http://docs.trakt.apiary.io/reference/search/text-query}{the trakt API docs for
further info}
}
\examples{
\dontrun{
breakingbad <- trakt.search("Breaking Bad")
}
}
\seealso{
Other API-basics: \code{\link{trakt.api.call}},
  \code{\link{trakt.search.byid}},
  \code{\link{trakt_credentials}}

Other search functions: \code{\link{trakt.search.byid}}
}
\concept{API-basics}
\concept{search functions}
