% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-seasons.R
\name{trakt.seasons.season}
\alias{trakt.seasons.season}
\title{Get a season of a show}
\usage{
trakt.seasons.season(target, seasons = 1L, extended = c("min", "full"))
}
\arguments{
\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"} for a show, or \code{"bryan-cranston"} for a person),
\code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"} or \code{"nm0186505"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df} if possible.}

\item{seasons}{\code{integer(1) [1L]}: The season(s) to get. Use 0 for special episodes.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
\code{trakt.seasons.season} pulls a season's full data.
Similar to \link{trakt.seasons.summary}, but this function returns full data for
a single season, i.e. all the episodes of the season.
See \href{http://docs.trakt.apiary.io/#introduction/extended-info}{the API docs} for possible values of
\code{extended} to customize output amount.
}
\note{
See \href{http://docs.trakt.apiary.io/reference/seasons/season/get-single-season-for-a-show}{the trakt API docs for further info}.
If you want to quickly gather data of multiple seasons, see \link[tRakt]{trakt.get_all_episodes}
}
\examples{
\dontrun{
breakingbad.seasons <- trakt.seasons.season("breaking-bad", 1)
}
}
\seealso{
Other show data: \code{\link{trakt.get_all_episodes}},
  \code{\link{trakt.seasons.summary}}
}
\concept{show data}
