% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trakt.show.season}
\alias{trakt.show.season}
\title{Get a season of a show}
\usage{
trakt.show.season(target, apikey = getOption("trakt.apikey"), season = 1)
}
\arguments{
\item{target}{The \code{slug} or \code{tvdbid} of the show requested}

\item{apikey}{API-key used for the call. Defaults to \code{getOption("trakt.apikey")}}

\item{season}{The season to get. Defaults to 1. Use 0 for special episodes.}
}
\value{
A \code{data.frame} containing all of a season's episodes
}
\description{
\code{trakt.show.season} pulls a full season's data.
Similar to \link{trakt.getSeasons}, but this function returns full data for
a single season, i.e. all the episodes of the season.
}
\note{
See \href{http://trakt.tv/api-docs/show-season}{the trakt API docs for further info}
}
\examples{
\dontrun{
options(trakt.apikey = jsonlite::fromJSON("key.json")$apikey)
breakingbad.seasons <- trakt.show.season("breaking-bad", 1)
}
}

