% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trakt.show.summaries}
\alias{trakt.show.summaries}
\title{Get show summary info}
\usage{
trakt.show.summaries(target, apikey = getOption("trakt.apikey"),
  extended = NULL)
}
\arguments{
\item{target}{The \code{slug} or \code{tvdbid} of the show requested}

\item{apikey}{API-key used for the call.
Defaults to \code{getOption("trakt.apikey")}}

\item{extended}{Whether extended info should be provided.
Defaults to \code{NULL}, can be \code{"normal"} or \code{"full"}}
}
\value{
A \code{data.frame} containing summary info
}
\description{
\code{trakt.show.summaries} pulls show summary data and returns it compactly.
}
\note{
See \href{http://trakt.tv/api-docs/show-summaries}{the trakt API docs for further info}
Not to be confused with \href{http://trakt.tv/api-docs/show-summary}{show/summary}
}
\examples{
\dontrun{
options(trakt.apikey = jsonlite::fromJSON("key.json")$apikey)
breakingbad.summary_small <- trakt.show.summaries("breaking-bad")
}
}

