% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.summary.R
\name{trakt.show.summary}
\alias{trakt.show.summary}
\title{Get show summary info}
\usage{
trakt.show.summary(target, extended = c("min", "full"),
  force_data_frame = FALSE)
}
\arguments{
\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"}), \code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}

\item{force_data_frame}{If \code{TRUE}, the \code{list} is unnested as much as possible, resulting
in a flat \link[tibble:tibble-package]{tibble} suitable to \code{rbind} with other summary results.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
\code{trakt.show.summary} pulls show summary data and returns it compactly.
}
\details{
Note that setting \code{extended} to \code{min} makes this function
return about as much informations as \link[tRakt]{trakt.search}
}
\note{
See \href{http://docs.trakt.apiary.io/reference/shows/summary}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
breakingbad.summary <- trakt.show.summary("breaking-bad")
}
}
\seealso{
Other show data: \code{\link{trakt.get_all_episodes}},
  \code{\link{trakt.get_full_showdata}},
  \code{\link{trakt.seasons.season}},
  \code{\link{trakt.seasons.summary}},
  \code{\link{trakt.stats}}
}
\concept{show data}
