% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trakt.shows.trending.R
\name{trakt.shows.trending}
\alias{trakt.shows.trending}
\title{Get trending shows}
\usage{
trakt.shows.trending(limit = 10)
}
\arguments{
\item{limit}{Number of shows to return. Is coerced to \code{integer} and must be greater than 0.}
}
\value{
A \code{data.frame} containing trending shows with their name and ids
}
\description{
\code{trakt.shows.trending} returns a list of trending shows on trakt.tv.
According to the API docs, it returns all shows being watched right now, where
shows with the most users are returned first.
}
\note{
See \href{http://docs.trakt.apiary.io/#reference/shows/get-trending-shows}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
trakt.shows.trending(5)
}
}

