% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.user.friendfollow.R
\name{trakt.user.network}
\alias{trakt.user.network}
\alias{trakt.user.friends}
\alias{trakt.user.followers}
\alias{trakt.user.following}
\title{Get a user's friends, followers or followings}
\usage{
trakt.user.network(relationship = c("friends", "followers", "following"),
  user = getOption("trakt.username"), extended = c("min", "full"))

trakt.user.friends(user = getOption("trakt.username"),
  extended = c("min", "full"))

trakt.user.followers(user = getOption("trakt.username"),
  extended = c("min", "full"))

trakt.user.following(user = getOption("trakt.username"),
  extended = c("min", "full"))
}
\arguments{
\item{user}{Target user. Defaults to \code{getOption("trakt.username")}. If multiple users
are specified, the results will be combined with \link{rbind}.}

\item{extended}{\code{[character(1)]}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
\code{trakt.user.friends} retrieves a user's friends, the two-way relationship
of both following and being followed by a user.
}
\note{
Since no OAuth2 methods are supported yet, the specified user must not be private.
}
\examples{
\dontrun{
trakt.user.friends("jemus42")
trakt.user.followers("jemus42")
trakt.user.following("jemus42")
}
}
\seealso{
Other user data: \code{\link{trakt.user.collection}},
  \code{\link{trakt.user.ratings}},
  \code{\link{trakt.user.stats}},
  \code{\link{trakt.user.watched}},
  \code{\link{trakt.user.watchlist}}
}
\concept{user data}
