% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trakt.user.watched.R
\name{trakt.user.watched}
\alias{trakt.user.watched}
\title{Get a user's watched shows or movies}
\usage{
trakt.user.watched(user = getOption("trakt.username"), type = "shows")
}
\arguments{
\item{user}{Target user. Defaults to \code{getOption("trakt.username")}}

\item{type}{Either \code{shows} (default), or \code{movies}}
}
\value{
A \code{data.frame} containing stats.
}
\description{
\code{trakt.user.watched} pulls a user's stats.
}
\note{
See \href{http://docs.trakt.apiary.io/reference/users/stats/get-stats}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
mystats   <- trakt.user.watched() # Defaults to your username if set
seanstats <- trakt.user.watched(user = "sean")
}
}

