% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-common.R
\name{trakt_api_common_parameters}
\alias{trakt_api_common_parameters}
\title{Common API parameters}
\arguments{
\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"} for a show, or \code{"bryan-cranston"} for a person),
\code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"} or \code{"nm0186505"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df} if possible.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}

\item{user}{Target user. Defaults to \code{getOption("trakt.username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and combined using \link[purrr:map_df]{purrr::map_df}.}

\item{period}{\code{character(1) ["weekly"]}: Which period to filter by. Possible values
are \code{"weekly"}, \code{"monthly"}, \code{"yearly"}, \code{"all"}.}

\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced to \code{integer}.}
}
\description{
These parameters are used \emph{a lot} throughout this package, and they tend to be
alsways the same. Here there are all documented in one place, which is intended
to make the individual function documentation more consistent.
As a side-effect, this is also visible to the user, which I believe to be acceptable.
}
