% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-oauth.R
\name{trakt_credentials}
\alias{trakt_credentials}
\title{trakt.tv credentials}
\usage{
trakt_credentials()
}
\description{
Authentication has changed with the migration to \link[httr2:httr2-package]{httr2::httr2}, so this function is just a
placeholder that shows information about available credentials.
}
