% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-internal.R
\docType{data}
\name{trakt_datasets}
\alias{trakt_datasets}
\alias{trakt_genres}
\alias{trakt_languages}
\alias{trakt_networks}
\alias{trakt_countries}
\alias{trakt_certifications}
\title{Cached filter datasets}
\format{
Every dataset is a \link[tibble:tibble-package]{tibble()}.
The following list includes the dataset topic with a link to the API documentation,
a short description and a list of variables with example values:
\itemize{
\item \href{https://trakt.docs.apiary.io/#reference/genres/list/get-genres}{Genres}:
Genres for shows and movies (with their two-letter codes) trakt.tv knows.
\itemize{
\item 3 Variables: \code{name} ("Action"), \code{slug} ("action"), \code{type} ("movies")
}
\item \href{https://trakt.docs.apiary.io/#reference/languages/list}{Languages}:
Languages (and two-letter codes) trakt.tv knows.
\itemize{
\item 3 Variables: \code{name} ("Arabic"), \code{code} ("ar"), \code{type} ("movies")
}
\item \href{https://trakt.docs.apiary.io/#reference/networks/list}{Networks}:
TV networks trakt.tv knows.
\itemize{
\item 2 Variables: \code{name} ("TBS"), \code{name_clean} ("tbs") (lower-case, no trailing whitespaces)
}
\item \href{https://trakt.docs.apiary.io/#reference/countries/list}{Countries}:
Country names (and two-letter codes).
\itemize{
\item 3 Variables: \code{name} ("Belarus"), \code{code} ("by"), \code{type} ("movies")
}
\item \href{https://trakt.docs.apiary.io/#reference/certifications/list}{Certifications}:
TV and movie certifications (e.g. "PG-13" and the likes).
\itemize{
\item 5 Variables: \code{country} ("us" only), \code{name} ("TV-PG"),
\code{slug} ("tv-pg"), \code{description} ("Parental Guidance Suggested"), \code{type} ("shows")
}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 64 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 243 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3045 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 349 rows and 3 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 5 columns.
}
\usage{
trakt_genres

trakt_languages

trakt_networks

trakt_countries

trakt_certifications
}
\description{
These datasets are used internally to check the optional
\href{https://trakt.docs.apiary.io/#introduction/filters}{filter parameters} for certain
functions (see \link{search_query} or the dynamic lists like \link{shows_popular}).
They are unlikely to change often and are therefore included as package datasets.
}
\details{
The datasets are prefixed with \code{trakt_} purely to avoid confusion or masking for
filter arguments of the same name.
}
\note{
Currently only US certifications are available.
}
\examples{
head(trakt_genres)
head(trakt_languages)
head(trakt_networks)
head(trakt_countries)
trakt_certifications
}
\keyword{datasets}
