% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-internal.R
\docType{data}
\name{trakt_genres}
\alias{trakt_genres}
\title{Genres}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36 rows and 2 columns.}
\source{
\url{https://trakt.docs.apiary.io/#reference/genres/list/get-genres}
}
\usage{
trakt_genres
}
\description{
Genres
}
\examples{
head(trakt_genres)
}
\keyword{datasets}
