% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-requests.R
\name{trakt_get}
\alias{trakt_get}
\title{Make an API call and receive parsed output}
\usage{
trakt_get(url)
}
\arguments{
\item{url}{\code{character(1)}: The API endpoint. Either a full URL like
\code{"https://api.trakt.tv/shows/breaking-bad"} or just the endpoint like
\code{shows/breaking-bad}.}
}
\value{
The parsed content of the API response.
An empty \link[tibble:tibble-package]{tibble()} if the response is an empty
\code{JSON} array.
}
\description{
The most basic form of API interaction: Querying a specific URL and getting
its parsed result. If the response is empty, the function returns an empty
\link[tibble:tibble-package]{tibble()}, and if there are date-time variables
present in the response, they are converted to \code{POSIXct} via
\code{\link[lubridate:ymd_hms]{lubridate::ymd_hms()}} or to \code{Date} via \code{\link[lubridate:as_date]{lubridate::as_date()}} if the
variable only contains date information.
}
\details{
See \href{https://trakt.docs.apiary.io}{the official API reference} for a detailed
overview of available methods. Most methods of potential interest for data
collection have dedicated functions in this package.
}
\examples{
# A simple request to a direct URL
trakt_get("https://api.trakt.tv/shows/breaking-bad")

# Optionally be lazy about URL specification by dropping the hostname:
trakt_get("shows/game-of-thrones")
}
\concept{API-basics}
