% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{trakt_get_token}
\alias{trakt_get_token}
\title{Get a trakt.tv OAuth2 token}
\usage{
trakt_get_token()
}
\value{
An OAuth2 token object. See \link[httr:oauth2.0_token]{oauth2.0_token}.
}
\description{
This is used internally for authenticated requests.
}
\seealso{
Other API-basics: \code{\link{search_query}},
  \code{\link{trakt_credentials}}, \code{\link{trakt_get}}
}
\concept{API-basics}
\keyword{internal}
