% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-internal.R
\docType{data}
\name{trakt_languages}
\alias{trakt_languages}
\title{Language codes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 151 rows and 2 columns.}
\source{
\url{https://trakt.docs.apiary.io/#reference/languages/list/get-languages}
}
\usage{
trakt_languages
}
\description{
Language codes
}
\examples{
head(trakt_languages)
}
\keyword{datasets}
