% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-internal.R
\docType{data}
\name{trakt_networks}
\alias{trakt_networks}
\title{Available networks}
\format{An object of class \code{character} of length 1524.}
\source{
\url{https://trakt.docs.apiary.io/#reference/networks/list/get-networks}
}
\usage{
trakt_networks
}
\description{
Available networks
}
\examples{
head(trakt_networks)
}
\keyword{datasets}
