% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{trakt_rating_labels}
\alias{trakt_rating_labels}
\title{Trakt's rating labels}
\format{
An object of class \code{character} of length 10.
}
\usage{
trakt_rating_labels
}
\description{
These can be useful for labeling the numeric scale.
}
\keyword{internal}
