% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRakt-package.R
\name{tRakt_sitrep}
\alias{tRakt_sitrep}
\title{tRakt sitrep}
\usage{
tRakt_sitrep()
}
\description{
Get general information about set options and APi credentaials.
}
\details{
Supported package-level options are:
\itemize{
\item \code{tRakt_debug}: \code{logical(1)}: Should debug output be enabled? Default is \code{FALSE}.
\item \code{tRakt_cache_dir}: \code{character(1)}: Where should the cache be stored?
\item \code{tRakt_cache_max_age}: \code{numeric(1)}: How long should the cache be kept, in seconds? Default is one week (604800 seconds).
\item \code{tRakt_cache_max_size}: \code{numeric(1)}: How big should the cache be, in bytes? Default is 100MB (100 * 1000^2 bytes).
}

All options can be overriden with environment variables of the same name.
}
\examples{
tRakt_sitrep()
}
