% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-common.R
\name{type_shows_movies}
\alias{type_shows_movies}
\title{Type: Shows or Movies}
\arguments{
\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}
}
\description{
Type: Shows or Movies
}
