% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extract-media.R
\name{unpack_show}
\alias{unpack_show}
\title{Unpack a standard show media object}
\usage{
unpack_show(show)
}
\description{
This should work regardless of the value of \code{extended} to be sufficiently robust
}
\keyword{internal}
