% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_comments.R
\name{user_comments}
\alias{user_comments}
\title{Get a user's comments}
\usage{
user_comments(user = getOption("trakt_username"),
  comment_type = c("all", "reviews", "shouts"), type = c("all",
  "movies", "shows", "seasons", "episodes", "lists"),
  include_replies = FALSE)
}
\arguments{
\item{user}{\code{character(1)}: Target username. Defaults to \code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{comment_type}{\code{character(1) ["all"]}: The type of comment, one of "all",
"reviews" or "shouts".}

\item{type}{\code{character(1) ["all"]}: The type of media to filter by, one of
"all", "movies", "shows", "seasons", "episodes" or "lists".}

\item{include_replies}{\code{logical(1) [FALSE]}: Whether to include replies.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Get a user's comments
}
\examples{
\dontrun{
user_comments("jemus42")
}
}
