% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_history.R
\name{user_history}
\alias{user_history}
\title{Get a user's watch history}
\source{
\code{user_history()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/users/history/get-watched-history}{users/:id/history/:type/:item_id?start_at=&end_at=}.
}
\usage{
user_history(
  user = "me",
  type = c("shows", "movies", "seasons", "episodes"),
  item_id = NULL,
  limit = 10L,
  start_at = NULL,
  end_at = NULL,
  extended = c("min", "full")
)
}
\arguments{
\item{user}{\code{character(1)}: Target username (or \code{slug}). Defaults to
\code{"me"}, the OAuth user.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}. For season/episode-specific
functions, values \code{seasons} or \code{episodes} are also allowed.}

\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced via \code{as.integer()}.}

\item{start_at, end_at}{\code{character(1)}: A time-window to filter by. Must be coercible
to a datetime object of class \code{POSIXct}. See \link{ISOdate} for further information.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Retrieve a the last \code{limit} items a user has watched, including the method by
which it was logged (e.g. \emph{scrobble} or \emph{checkin}).
}
\details{
This function wraps the API method
\href{https://trakt.docs.apiary.io/#reference/users/history/get-watched-history}{\verb{/users/:id/history/:type}}.
}
\note{
For \code{type = "shows"}, the
original output contains a nested object with \code{show} and \code{episode} data,
which are unnested by this function. Due to duplicate variable names,
all episode-related variables are prefixed with \code{episode_}. This results in the
episode number having the name \code{episode_episode}, which is quite silly. Sorry.
}
\examples{
\dontrun{
# Shows user "jemus42" watched around christmas 2016
user_history(
  user = "jemus42", type = "shows", limit = 5,
  start_at = "2015-12-24", end_at = "2015-12-28"
)
}
}
\seealso{
Other user data: 
\code{\link{user_collection}()},
\code{\link{user_comments}()},
\code{\link{user_likes}()},
\code{\link{user_network}()},
\code{\link{user_profile}()},
\code{\link{user_ratings}()},
\code{\link{user_stats}()},
\code{\link{user_watched}()},
\code{\link{user_watchlist}()}
}
\concept{user data}
