% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_likes.R
\name{user_likes}
\alias{user_likes}
\title{Get items (comments, lists) a user likes}
\usage{
user_likes(type = c("comments", "lists"))
}
\arguments{
\item{type}{One of "comments", "lists".}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
This requires authentication.
}
\examples{
# Get liked lists
if (FALSE) {
user_likes("lists")
}
}
