% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_list_items.R
\name{user_list_items}
\alias{user_list_items}
\title{Get a user's list's items}
\usage{
user_list_items(user = getOption("trakt_username"), list_id,
  type = NULL, extended = c("min", "full"))
}
\arguments{
\item{user}{\code{character(1)}: Target username. Defaults to \code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{list_id}{The list identifier, either \code{trakt} ID or \code{slug} of the list.
Can be optained via the website (URL \code{slug}) or e.g. \link{user_lists}.}

\item{type}{\code{character(1) [NULL]}: If not \code{NULL}, only items of that media type are
returned. Possible values are \code{"movie"}, \code{"show"}, \code{"season"}, \code{"episode"}, \code{"person"}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See section \link[=trakt_api_common_parameters]{"Extended Information"} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Get a user's list's items
}
\examples{
\dontrun{
# A large list with various media types
# All items
user_list_items("sp1ti", list_id = "5615781", extended = "min")

# Movies only
user_list_items("sp1ti", list_id = "5615781", extended = "min", type = "movie")

# Shows...
user_list_items("sp1ti", list_id = "5615781", extended = "min", type = "shows")

# Only seasons
user_list_items("sp1ti", list_id = "5615781", extended = "min", type = "season")

# Only episodes
user_list_items("sp1ti", list_id = "5615781", extended = "min", type = "episodes")
}
}
\seealso{
Other list methods: \code{\link{lists_popular}},
  \code{\link{user_lists}}, \code{\link{user_list}}
}
\concept{list methods}
