% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_network.R
\name{user_network}
\alias{user_network}
\alias{user_followers}
\alias{user_following}
\alias{user_friends}
\title{Get a user's social connections}
\usage{
user_followers(user = getOption("trakt_user"), extended = "min")

user_following(user = getOption("trakt_user"), extended = "min")

user_friends(user = getOption("trakt_user"), extended = "min")
}
\arguments{
\item{user}{\code{character(1)}: Target username. Defaults to \code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See section \link[=trakt_api_common_parameters]{"Extended Information"} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Get followers, followings or friends (the two-way relationship).
}
\details{
The corresponding API methods are described in
\href{https://trakt.docs.apiary.io/#reference/users/followers}{the API reference}.

The relevant endpoints for this function are:
\itemize{
\item \code{user_followers}: \code{users/:id/followers}
\item \code{user_following}: \code{users/:id/following}
\item \code{user_friends}: \code{users/:id/friends}
}
}
\note{
If the specified user is private, you need to be able to make an \link[=trakt_credentials]{authenticated request} and be friends with the user.
}
\examples{
\dontrun{
user_followers(user = "sean")
}
\dontrun{
user_following(user = "sean")
}
\dontrun{
user_friends(user = "sean")
}
}
\seealso{
Other user data: \code{\link{user_collection}},
  \code{\link{user_history}}, \code{\link{user_ratings}},
  \code{\link{user_stats}}, \code{\link{user_summary}},
  \code{\link{user_watched}}, \code{\link{user_watchlist}}
}
\concept{user data}
