% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-common.R
\name{user_param}
\alias{user_param}
\title{User parameter}
\arguments{
\item{user}{Target user. Defaults to \code{getOption("trakt.username")}. If multiple users
are specified, the results will be combined with \link{rbind}.}
}
\description{
User parameter
}
