% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_stats.R
\name{user_stats}
\alias{user_stats}
\title{Get a user's stats}
\usage{
user_stats(user = getOption("trakt_username"))
}
\arguments{
\item{user}{\code{character(1)}: Target username. Defaults to \code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}
}
\value{
A \code{list} of \link[tibble:tibble-package]{tibbles} containing the following elements:
\itemize{
\item "movies"
\item "shows"
\item "seasons"
\item "episodes"
\item "network"
\item "ratings"
}
}
\description{
Data about a user's interactions with movies, shows, seasons, episodes,
as well as their social network (friends, followings, followers) and a
frequency table of the user's media ratings so far.
}
\details{
This function wraps the API method
\href{https://trakt.docs.apiary.io/#reference/users/stats/get-stats}{/users/:id/stats}.
}
\examples{
\dontrun{
user_stats(user = "sean")
}
}
\seealso{
Other user data: \code{\link{user_collection}},
  \code{\link{user_history}}, \code{\link{user_network}},
  \code{\link{user_profile}}, \code{\link{user_ratings}},
  \code{\link{user_watched}}, \code{\link{user_watchlist}}
}
\concept{user data}
