% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_summary.R
\name{user_summary}
\alias{user_summary}
\title{Get a user's profile}
\usage{
user_summary(user = getOption("trakt_username"), extended = c("min",
  "full"))
}
\arguments{
\item{user}{\code{character(1)}: Target username. Defaults to \code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See section \link[=trakt_api_common_parameters]{"Extended Information"} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Get a user's profile
}
\details{
The corresponding API methods is
\href{https://trakt.docs.apiary.io/#reference/users/profile/get-user-profile}{users/:id}.
}
\note{
If the specified user is private, you need to be able to make an \link[=trakt_credentials]{authenticated request} and be friends with the user.
}
\examples{
\dontrun{
user_summary("sean")
}
}
\seealso{
Other user data: \code{\link{user_collection}},
  \code{\link{user_history}}, \code{\link{user_network}},
  \code{\link{user_ratings}}, \code{\link{user_stats}},
  \code{\link{user_watched}}, \code{\link{user_watchlist}}
}
\concept{user data}
