% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_watchlist.R
\name{user_watchlist}
\alias{user_watchlist}
\title{Get a user's watchlist}
\usage{
user_watchlist(user = getOption("trakt_username"), type = c("movies",
  "shows"), extended = c("min", "full"))
}
\arguments{
\item{user}{\code{character(1)}: Target username. Defaults to \code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}. For season/episode-specific
functions, values \code{seasons} or \code{episodes} are also allowed.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Retrieve a user's watchlisted shows or movies.
}
\details{
This function wraps the API method
\href{https://trakt.docs.apiary.io/#reference/users/watchlist/get-watchlist}{/users/:id/watchlist/:type}.
}
\examples{
\dontrun{
# Defaults to movie watchlist and minimal info
user_watchlist(user = "sean")
}
}
\seealso{
Other user data: \code{\link{user_collection}},
  \code{\link{user_history}}, \code{\link{user_network}},
  \code{\link{user_ratings}}, \code{\link{user_stats}},
  \code{\link{user_summary}}, \code{\link{user_watched}}
}
\concept{user data}
