% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.popular-trending.R
\name{automated_lists}
\alias{automated_lists}
\alias{trakt.popular}
\alias{trakt.trending}
\alias{trakt.anticipated}
\alias{trakt.played}
\alias{trakt.watched}
\title{See which movies / shows are popular}
\usage{
trakt.popular(type = c("shows", "movies"), limit = 10, page = 1,
  extended = c("min", "full"))

trakt.trending(type = c("shows", "movies"), limit = 10, page = 1,
  extended = c("min", "full"))

trakt.anticipated(type = c("shows", "movies"), limit = 10, page = 1,
  extended = c("min", "full"))

trakt.played(type = c("shows", "movies"), limit = 10, page = 1,
  extended = c("min", "full"), period = c("weekly", "monthly",
  "yearly", "all"))

trakt.watched(type = c("shows", "movies"), limit = 10, page = 1,
  extended = c("min", "full"), period = c("weekly", "monthly",
  "yearly", "all"))
}
\arguments{
\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}

\item{limit, page}{\code{integer(1) [10L], [1L]}: Number of items and page
for paginated requests. Both values must be greater than \code{0} and will
be coerced to \code{integer}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}

\item{period}{\code{character(1) ["weekly"]}: Which period to filter by. Possible values
are \code{"weekly"}, \code{"monthly"}, \code{"yearly"}, \code{"all"}.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
According to the API docs, popularity is calculated based both ratings
and number of ratings. Trending items are those being watched right now, where
items with the most users watching are returned first. Anticipation is measured
by the number of user-created lists an items is part of while not being released yet.
}
\examples{
\dontrun{
# Get popular shows with only ids
trakt.popular(type = "shows")

# Get trending movies with extended information
trakt.trending(type = "movies", 5, extended = "full")

# Get top 5 anticipated movies
trakt.anticipated(type = "movies", 5)
}
}
