% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-automated-lists.R
\name{automated_lists}
\alias{automated_lists}
\alias{trakt.popular}
\alias{trakt.trending}
\alias{trakt.anticipated}
\alias{trakt.played}
\alias{trakt.watched}
\alias{trakt.collected}
\alias{trakt.updates}
\title{See which movies / shows are popular}
\usage{
trakt.popular(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"))

trakt.trending(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"))

trakt.anticipated(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"))

trakt.played(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), period = c("weekly", "monthly",
  "yearly", "all"))

trakt.watched(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), period = c("weekly", "monthly",
  "yearly", "all"))

trakt.collected(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), period = c("weekly", "monthly",
  "yearly", "all"))

trakt.updates(type = c("shows", "movies"), limit = 10,
  extended = c("min", "full"), start_date = Sys.Date() - 7)
}
\arguments{
\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}. For season/episode-specific
functions, values \code{seasons} or \code{episodes} are also allowed.}

\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced to \code{integer}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}

\item{period}{\code{character(1) ["weekly"]}: Which period to filter by. Possible values
are \code{"weekly"}, \code{"monthly"}, \code{"yearly"}, \code{"all"}.}

\item{start_date}{\code{character(1)}: A date in the past from which on to count updates.
If no date is supplied, the default is to use the date 7 days in the past relative
to the current date. Value must either be standard \code{YYYY-MM-DD} format or an object
of class \link[base:Dates]{Date}.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
According to the API docs, popularity is calculated based both ratings
and number of ratings. Trending items are those being watched right now, where
items with the most users watching are returned first. Anticipation is measured
by the number of user-created lists an items is part of while not being released yet.
}
\examples{
\dontrun{
# Get popular shows with only ids
trakt.popular(type = "shows")

# Get trending movies with extended information
trakt.trending(type = "movies", 5, extended = "full")

# Get top 5 anticipated movies
trakt.anticipated(type = "movies", 5)
}
}
