% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.popular-trending.R
\name{automated_lists}
\alias{automated_lists}
\alias{trakt.popular}
\alias{trakt.trending}
\alias{trakt.anticipated}
\alias{trakt.movies.trending}
\alias{trakt.shows.trending}
\alias{trakt.movies.popular}
\alias{trakt.shows.popular}
\alias{trakt.movies.anticipated}
\alias{trakt.shows.anticipated}
\title{Get popular/trending/anticipated movies and shows}
\usage{
trakt.popular(type = c("shows", "movies"), limit = 10, page = 1,
  extended = c("min", "full"))

trakt.trending(type = c("shows", "movies"), limit = 10, page = 1,
  extended = c("min", "full"))

trakt.anticipated(type = c("shows", "movies"), limit = 10, page = 1,
  extended = c("min", "full"))

trakt.movies.trending(limit = 10, page = 1, extended = c("min",
  "full"))

trakt.shows.trending(limit = 10, page = 1, extended = c("min",
  "full"))

trakt.movies.popular(limit = 10, page = 1, extended = c("min",
  "full"))

trakt.shows.popular(limit = 10, page = 1, extended = c("min",
  "full"))

trakt.movies.anticipated(limit = 10, page = 1, extended = c("min",
  "full"))

trakt.shows.anticipated(limit = 10, page = 1, extended = c("min",
  "full"))
}
\arguments{
\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}

\item{limit, page}{\code{integer(1) [10L] [1L]}: Number of items and page for paginated requests.
Bot values must be greater than \code{0} and will be coerced to \code{integer}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
According to the API docs, popularity is calculated based both ratings
and number of ratings. Trending items are those being watched right now, where
items with the most users watching are returned first. Anticipation is measured
by the number of user-created lists an items is part of while not being released yet.
}
\examples{
\dontrun{
# Get popular movies and shows
trakt.movies.popular(5)
trakt.shows.popular(5)

# Get trending movies and shows
trakt.movies.trending(5)
trakttrending(type = "movies", 5)
}

# Get anticipated movies
trakt.anticipated(type = "movies", 10)
}
