% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_username}
\alias{check_username}
\title{Check username}
\usage{
check_username(user)
}
\arguments{
\item{user}{The username input}
}
\value{
An error if the checks fail or else \code{TRUE} invisibly
}
\description{
Check username
}
\keyword{internal}
