% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_comment.R
\name{comments_comment}
\alias{comments_comment}
\alias{comments_replies}
\alias{comments_likes}
\alias{comments_item}
\title{Get a single comment}
\usage{
comments_comment(id, extended = c("min", "full"))

comments_replies(id, extended = c("min", "full"))

comments_likes(id, extended = c("min", "full"))

comments_item(id, extended = c("min", "full"))
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Get a single comment
}
\section{Functions}{
\itemize{
\item \code{comments_replies}: Get a comment's replies

\item \code{comments_likes}: Get users who liked a comment.

\item \code{comments_item}: Get the media item attached to the comment.
}}

\examples{
# A single comment
comments_comment("236397")

# Multiple comments
comments_comment(c("236397", "112561"))
\dontrun{
comments_replies("236397")
}
\dontrun{
comments_likes("236397")
}
\dontrun{
# A movie
comments_item("236397")
comments_item("236397", extended = "full")

# A show
comments_item("120768")
comments_item("120768", extended = "full")

# A season
comments_item("140265")
comments_item("140265", extended = "full")

# An episode
comments_item("136632")
comments_item("136632", extended = "full")
}
}
