% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments_trending.R
\name{comments_trending}
\alias{comments_trending}
\alias{comments_recent}
\alias{comments_updates}
\title{Get trending, recent or recently updated comments}
\usage{
comments_trending(comment_type = c("all", "reviews", "shouts"),
  type = c("all", "movies", "shows", "seasons", "episodes", "lists"),
  include_replies = FALSE, limit = 10L)

comments_recent(comment_type = c("all", "reviews", "shouts"),
  type = c("all", "movies", "shows", "seasons", "episodes", "lists"),
  include_replies = FALSE, limit = 10L)

comments_updates(comment_type = c("all", "reviews", "shouts"),
  type = c("all", "movies", "shows", "seasons", "episodes", "lists"),
  include_replies = FALSE, limit = 10L)
}
\arguments{
\item{comment_type}{\code{character(1) ["all"]}: The type of comment, one of "all",
"reviews" or "shouts".}

\item{type}{\code{character(1) ["all"]}: The type of media to filter by, one of
"all", "movies", "shows", "seasons", "episodes" or "lists".}

\item{include_replies}{\code{logical(1) [FALSE]}: Whether to include replies.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Get trending, recent or recently updated comments
}
\section{Functions}{
\itemize{
\item \code{comments_trending}: Get trending comments.

\item \code{comments_recent}: Get recently made comments.

\item \code{comments_updates}: Get recently updated comments.
}}

\examples{
\dontrun{
comments_trending("reviews")
comments_recent("shouts")
comments_updates()
}
}
