% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-internal.R
\docType{data}
\name{countries}
\alias{countries}
\title{Country codes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 219 rows and 2 columns.}
\source{
\url{https://trakt.docs.apiary.io/#reference/countries/list/get-countries}
}
\usage{
countries
}
\description{
Country codes
}
\examples{
head(countries)
}
\keyword{datasets}
