% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes_summary.R
\name{episodes_summary}
\alias{episodes_summary}
\title{Get a single episode's details}
\source{
\href{https://trakt.docs.apiary.io/#reference/episodes/summary/get-a-single-episode-for-a-show}{shows/:id/seasons/:season/episodes/:episode}
}
\usage{
episodes_summary(id, season = 1L, episode = 1L, extended = c("min",
  "full"))
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}

\item{season}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}

\item{episode}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
This retrieves a single episode. See \link{seasons_season} for a whole season, and
\link{seasons_summary} for (potentially) all episodes of a show.
}
\examples{
# Get just this one episode with its ratings, votes, etc.
episodes_summary("breaking-bad", season = 1, episode = 1, extended = "full")
}
\seealso{
Other episode data: \code{\link{media_comments}},
  \code{\link{media_lists}}, \code{\link{media_people}},
  \code{\link{media_ratings}}, \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{media_watching}},
  \code{\link{seasons_season}},
  \code{\link{seasons_summary}},
  \code{\link{shows_next_episode}}

Other summary methods: \code{\link{comments_comment}},
  \code{\link{movies_summary}},
  \code{\link{people_summary}},
  \code{\link{seasons_summary}},
  \code{\link{shows_summary}}, \code{\link{user_profile}}
}
\concept{episode data}
\concept{summary methods}
