% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-episodes.R
\docType{data}
\name{futurama}
\alias{futurama}
\title{Futurama episodes}
\format{A \link[tibble:tibble-package]{tibble()} with 124 rows and 18 variables:
\describe{
\item{episode, season}{Episode within each season and season number}
\item{title}{Episode title}
\item{episode_abs}{Overall episode number}
\item{overview}{Episode summary}
\item{rating}{Rating (1-10) on  trakt.tv}
\item{votes}{Number of votes ontrakt.tv}
\item{comment_count}{Number of comments on episode page}
\item{first_aired, updated_at}{Original air date and last update in UTC as \code{POSIXct}}
\item{runtime}{Runtime in minutes}
\item{trakt, tvdb, tmdb}{Episode IDs for trakt.tv, TVDb, and TMDb}
\item{available_translations}{List-column of available translation on trakt.tv}
}}
\usage{
futurama
}
\description{
This data comes from \url{https://trakt.tv} and serves as an example of episode data output.
}
\examples{
futurama
}
\seealso{
Other Episode datasets: 
\code{\link{gameofthrones}}
}
\concept{Episode datasets}
\keyword{datasets}
