% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gameofthrones}
\alias{gameofthrones}
\title{Game of Thrones episodes}
\format{A tibble with 60 rows and 21 variables:
\describe{
\item{episode_abs}{Overall episode number}
\item{episode, season}{Episode within each season and season number}
\item{epid}{Episode ID in \code{s00e00} format}
\item{runtime}{Runtime in minutes}
\item{title}{Episode title}
\item{year}{Year of first airing}
\item{overview}{Episode summary}
\item{rating_}{Rating (1-10) on IMDb and trakt.tv}
\item{votes}{Number of votes and trakt.tv}
\item{viewers}{Number of viewers (US) in millions}
\item{director, writer}{Direction and writing credits}
\item{first_aired, first_aired.string}{Original airdate in UTC as
\code{POSIXct} and \code{character}}
\item{trakt, tvdb, tmdb}{Episode IDs for trakt.tv, TVDb, and TMDb}
}}
\usage{
gameofthrones
}
\description{
This data comes from \url{https://trakt.tv} and
\url{https://en.wikipedia.org/wiki/List_of_House_episodes}.
}
\examples{
gameofthrones
}
\seealso{
Other Episode datasets: \code{\link{futurama}}
}
\concept{Episode datasets}
\keyword{datasets}
