% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-internal.R
\docType{data}
\name{genres}
\alias{genres}
\title{Genres}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36 rows and 2 columns.}
\usage{
genres
}
\description{
Genres
}
\examples{
genres
}
\keyword{datasets}
