% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{get_trakt_credentials}
\alias{get_trakt_credentials}
\title{Get the trakt.tv credentials}
\usage{
get_trakt_credentials(apikey = NULL, username = NULL)
}
\arguments{
\item{apikey}{Optional. Directly set your API key}

\item{username}{Optional. Also set your trakt.tv username (Not used yet)}
}
\value{
Nothing
}
\description{
\code{get_trakt_credentials} searches for your credentials and stores them
in the appropriate \code{option} variables
}
\note{
Not yet implemented for the APIv2
}
\examples{
\dontrun{
get_trakt_credentials()
}
}

