% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_trakt_credentials}
\alias{get_trakt_credentials}
\title{Set the required trakt.tv API credentials}
\usage{
get_trakt_credentials(username = "", client.id = "",
  set.headers = TRUE, silent = TRUE)
}
\arguments{
\item{username}{\code{character(1)}: Explicitly set your trakt.tv username (optional).}

\item{client.id}{\code{character(1)}: Explicitly set your APIv2 client id (required for API interaciton).}

\item{set.headers}{\code{logical(1) [TRUE]}: Sets the \code{httr} headers
for \code{GET} requests to the APIv2.}

\item{silent}{\code{logical(1) [TRUE]}: No messages are printed showing you the API information.
Mostly for debug purposes.}
}
\value{
Nothing. Only messages.
}
\description{
\code{get_trakt_credentials} searches for your credentials and stores them
in the appropriate \code{option} variables.
It also sets the HTTP header required for v2 API calls.
To make this work automatically, place your key as environment variables in
\code{~/.Renviron}.
Arguments to this function take precedence over any key file.
}
\details{
Set appropriate values in your \code{~/.Renviron} like this:\preformatted{# tRakt
trakt_username=jemus42
trakt_client_id=12[...]f2
}
}
\note{
Please note that no OAuth2 methods are supported yet,
only client id really matters.
}
\examples{
\dontrun{
# Use a values set in ~/.Renviron in an R session:
# (This is automatically executed when calling library(tRakt))
get_trakt_credentials()

# Explicitly set values in an R session, overriding .Renviron values f present
get_trakt_credentials(
  username = "sean",
  client.id = "12fc1de7671c7f2fb4a8ac08ba7c9f45b447f4d5bad5e11e3490823d629afdf2"
)
}
}
\seealso{
Other API-basics: \code{\link{trakt.api.call}},
  \code{\link{trakt.search.byid}},
  \code{\link{trakt.search}}
}
\concept{API-basics}
