% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initializeEpisodes}
\alias{initializeEpisodes}
\title{Initialize an empty episode dataset}
\usage{
initializeEpisodes(show.seasons = NULL)
}
\arguments{
\item{show.seasons}{Show seasons dataset, normally provided by \link{trakt.getSeasons}}
}
\value{
A \code{data.frame} containing episode placeholders.
}
\description{
\code{initializeEpisodes} uses the show season info provided by \link{trakt.getSeasons}
to initialize a \code{data.frame} with a row for each episode of the show.
}
\note{
This is not a regular trakt-function, hence the "trakt."-prefix is omitted.
}
\examples{
\dontrun{
options(trakt.apikey = jsonlite::fromJSON("key.json")$apikey)
breakingbad.seasons <- trakt.getSeasons("breaking-bad")
breakingbad.episodes <- initializeEpisodes(breakingbad.seasons)
}
}

