% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_popular.R
\name{lists_popular}
\alias{lists_popular}
\alias{lists_trending}
\title{Get popular / trending lists}
\usage{
lists_popular(limit = 10)

lists_trending(limit = 10)
}
\arguments{
\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced via \code{as.integer()}.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get popular / trending lists
}
\examples{
\dontrun{
lists_popular()
lists_trending()
}
}
\seealso{
\link{user_list_items} For the actual content of a list.

Other list methods: \code{\link{media_lists}},
  \code{\link{user_list_comments}},
  \code{\link{user_list_items}}, \code{\link{user_lists}},
  \code{\link{user_list}}
}
\concept{list methods}
