% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_popular.R
\name{lists_popular}
\alias{lists_popular}
\alias{lists_trending}
\title{Get popular / trending lists}
\source{
\href{https://trakt.docs.apiary.io/#reference/lists/popular/get-popular-lists}{lists/popular}

\href{https://trakt.docs.apiary.io/#reference/lists/trending/get-trending-lists}{lists/trending}
}
\usage{
lists_popular(limit = 10)

lists_trending(limit = 10)
}
\arguments{
\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced via \code{as.integer()}.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get popular / trending lists
}
\examples{
\dontrun{
lists_popular()
lists_trending()
}
}
\seealso{
\code{\link[=user_list_items]{user_list_items()}} For the actual content of a list.

Other list methods: \code{\link{media_lists}},
  \code{\link{user_list_comments}},
  \code{\link{user_list_items}}, \code{\link{user_lists}},
  \code{\link{user_list}}

Other dynamic lists: \code{\link{anticipated_media}},
  \code{\link{collected_media}},
  \code{\link{played_media}}, \code{\link{popular_media}},
  \code{\link{trending_media}}, \code{\link{watched_media}}

Other list methods: \code{\link{media_lists}},
  \code{\link{user_list_comments}},
  \code{\link{user_list_items}}, \code{\link{user_lists}},
  \code{\link{user_list}}

Other dynamic lists: \code{\link{anticipated_media}},
  \code{\link{collected_media}},
  \code{\link{played_media}}, \code{\link{popular_media}},
  \code{\link{trending_media}}, \code{\link{watched_media}}
}
\concept{dynamic lists}
\concept{list methods}
