% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_lists.R
\name{media_lists}
\alias{media_lists}
\alias{movies_lists}
\alias{shows_lists}
\alias{seasons_lists}
\alias{episodes_lists}
\alias{people_lists}
\title{Get lists containing a movie, show, season, episode or person}
\source{
movies_lists() wraps endpoint \href{https://trakt.docs.apiary.io/#reference/movies/lists/get-lists-containing-this-movie}{movies/:id/lists/:type/:sort}

shows_lists() wraps endpoint \href{https://trakt.docs.apiary.io/#reference/shows/lists/get-lists-containing-this-show}{shows/:id/lists/:type/:sort}

seasons_lists() wraps endpoint \href{https://trakt.docs.apiary.io/#reference/seasons/lists/get-lists-containing-this-season}{shows/:id/seasons/:season/lists/:type/:sort}

episodes_lists() wraps endpoint \href{https://trakt.docs.apiary.io/#reference/episodes/lists/get-lists-containing-this-episode}{shows/:id/seasons/:season/episodes/:episode/lists/:type/:sort}
}
\usage{
movies_lists(id, type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L, extended = c("min", "full"))

shows_lists(id, type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L, extended = c("min", "full"))

seasons_lists(id, season, type = c("all", "personal", "official",
  "watchlists"), sort = c("popular", "likes", "comments", "items",
  "added", "updated"), limit = 10L, extended = c("min", "full"))

episodes_lists(id, season, episode, type = c("all", "personal",
  "official", "watchlists"), sort = c("popular", "likes", "comments",
  "items", "added", "updated"), limit = 10L, extended = c("min",
  "full"))

people_lists(id, type = c("all", "personal", "official", "watchlists"),
  sort = c("popular", "likes", "comments", "items", "added", "updated"),
  limit = 10L, extended = c("min", "full"))
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}

\item{type}{\code{character(1) ["all"]}: The type of list, one of "all", "personal",
"official" or "watchlists".}

\item{sort}{\code{character(1) ["popular"]}: Sort lists by one of "popular", "likes",
"comments", "items", "added" or "updated".}

\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced via \code{as.integer()}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}

\item{season}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}

\item{episode}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get lists containing a movie, show, season, episode or person
}
\section{Functions}{
\itemize{
\item \code{movies_lists}: Lists containing a movie.

\item \code{shows_lists}: Lists containing a show.

\item \code{seasons_lists}: Lists containing a season.

\item \code{episodes_lists}: Lists containing an episode.

\item \code{people_lists}: Lists containing a person.
}}

\examples{
\dontrun{
movies_lists("190430", type = "personal", limit = 5)
shows_lists("46241")
seasons_lists("46241", season = 1)
episodes_lists("46241", season = 1, episode = 1)
people_lists("david-tennant")
}
}
\seealso{
Other list methods: \code{\link{lists_popular}},
  \code{\link{user_list_comments}},
  \code{\link{user_list_items}}, \code{\link{user_lists}},
  \code{\link{user_list}}

Other movie data: \code{\link{anticipated_media}},
  \code{\link{collected_media}},
  \code{\link{media_aliases}},
  \code{\link{media_comments}}, \code{\link{media_people}},
  \code{\link{media_ratings}}, \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{media_watching}},
  \code{\link{movies_boxoffice}},
  \code{\link{movies_related}},
  \code{\link{movies_releases}},
  \code{\link{movies_summary}}, \code{\link{people_media}},
  \code{\link{played_media}}, \code{\link{popular_media}},
  \code{\link{trending_media}},
  \code{\link{updated_media}}, \code{\link{watched_media}}

Other list methods: \code{\link{lists_popular}},
  \code{\link{user_list_comments}},
  \code{\link{user_list_items}}, \code{\link{user_lists}},
  \code{\link{user_list}}

Other show data: \code{\link{collected_media}},
  \code{\link{media_aliases}},
  \code{\link{media_comments}}, \code{\link{media_people}},
  \code{\link{media_ratings}}, \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{media_watching}}, \code{\link{people_media}},
  \code{\link{played_media}},
  \code{\link{shows_next_episode}},
  \code{\link{shows_related}}, \code{\link{shows_summary}},
  \code{\link{updated_media}}

Other list methods: \code{\link{lists_popular}},
  \code{\link{user_list_comments}},
  \code{\link{user_list_items}}, \code{\link{user_lists}},
  \code{\link{user_list}}

Other season data: \code{\link{media_comments}},
  \code{\link{media_people}}, \code{\link{media_ratings}},
  \code{\link{media_stats}}, \code{\link{seasons_season}},
  \code{\link{seasons_summary}}

Other list methods: \code{\link{lists_popular}},
  \code{\link{user_list_comments}},
  \code{\link{user_list_items}}, \code{\link{user_lists}},
  \code{\link{user_list}}

Other episode data: \code{\link{episodes_summary}},
  \code{\link{media_comments}}, \code{\link{media_people}},
  \code{\link{media_ratings}}, \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{media_watching}},
  \code{\link{seasons_season}},
  \code{\link{seasons_summary}},
  \code{\link{shows_next_episode}}

Other list methods: \code{\link{lists_popular}},
  \code{\link{user_list_comments}},
  \code{\link{user_list_items}}, \code{\link{user_lists}},
  \code{\link{user_list}}

Other people data: \code{\link{media_people}},
  \code{\link{people_media}}, \code{\link{people_summary}}
}
\concept{episode data}
\concept{list methods}
\concept{movie data}
\concept{people data}
\concept{season data}
\concept{show data}
