% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_people.R
\name{media_people}
\alias{media_people}
\alias{movies_people}
\alias{shows_people}
\alias{seasons_people}
\alias{episodes_people}
\title{Get the cast and crew of a show or movie}
\source{
movies_people() wraps endpoint \href{https://trakt.docs.apiary.io/#reference/movies/people/get-all-people-for-a-movie}{movies/:id/people}

shows_people() wraps endpoint \href{https://trakt.docs.apiary.io/#reference/shows/people/get-all-people-for-a-show}{shows/:id/people}

seasons_people() wraps endpoint \href{https://trakt.docs.apiary.io/#reference/seasons/people/get-all-people-for-a-season}{shows/:id/seasons/:season/people}

episodes_people() wraps endpoint \href{https://trakt.docs.apiary.io/#reference/episodes/people/get-all-people-for-an-episode}{shows/:id/seasons/:season/episodes/:episode/people}
}
\usage{
movies_people(id, extended = c("min", "full"))

shows_people(id, guest_stars = FALSE, extended = c("min", "full"))

seasons_people(id, season = 1L, guest_stars = FALSE,
  extended = c("min", "full"))

episodes_people(id, season = 1L, episode = 1L, guest_stars = FALSE,
  extended = c("min", "full"))
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}

\item{guest_stars}{\code{logical(1) ["FALSE"]}: Also include guest stars. This returns
a lot of data, so use with care.}

\item{season}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}

\item{episode}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}
}
\value{
A \code{list} of one or more \link[tibble:tibble-package]{tibbles} for \code{cast}
and/or \code{crew}. The latter \code{tibble} objects are as flat as possible.
}
\description{
Returns all cast and crew for a show/movie, depending on how much data is
available.
}
\note{
As of 2019-09-30, there are two representations of \code{character[s]} and
\code{job[s]}:
One is a regular character variable, and the other is a list-column. The former is
\href{https://github.com/trakt/api-help/issues/74}{deprecated} and only included for
compatibility reasons.
}
\examples{
\dontrun{
movies_people("deadpool-2016")
shows_people("breaking-bad")
seasons_people("breaking-bad", season = 1)
episodes_people("breaking-bad", season = 1, episode = 1)
}
}
\seealso{
\link{people_media}, for the other direction: People that have credits
in shows/movies.

Other people data: \code{\link{media_lists}},
  \code{\link{people_media}}, \code{\link{people_summary}}

Other movie data: \code{\link{anticipated_media}},
  \code{\link{collected_media}},
  \code{\link{media_aliases}},
  \code{\link{media_comments}}, \code{\link{media_lists}},
  \code{\link{media_ratings}}, \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{media_watching}},
  \code{\link{movies_boxoffice}},
  \code{\link{movies_related}},
  \code{\link{movies_releases}},
  \code{\link{movies_summary}}, \code{\link{people_media}},
  \code{\link{played_media}}, \code{\link{popular_media}},
  \code{\link{trending_media}},
  \code{\link{updated_media}}, \code{\link{watched_media}}

Other people data: \code{\link{media_lists}},
  \code{\link{people_media}}, \code{\link{people_summary}}

Other show data: \code{\link{collected_media}},
  \code{\link{media_aliases}},
  \code{\link{media_comments}}, \code{\link{media_lists}},
  \code{\link{media_ratings}}, \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{media_watching}}, \code{\link{people_media}},
  \code{\link{played_media}},
  \code{\link{shows_next_episode}},
  \code{\link{shows_related}}, \code{\link{shows_summary}},
  \code{\link{updated_media}}

Other people data: \code{\link{media_lists}},
  \code{\link{people_media}}, \code{\link{people_summary}}

Other season data: \code{\link{media_comments}},
  \code{\link{media_lists}}, \code{\link{media_ratings}},
  \code{\link{media_stats}}, \code{\link{seasons_season}},
  \code{\link{seasons_summary}}

Other people data: \code{\link{media_lists}},
  \code{\link{people_media}}, \code{\link{people_summary}}

Other episode data: \code{\link{episodes_summary}},
  \code{\link{media_comments}}, \code{\link{media_lists}},
  \code{\link{media_ratings}}, \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{media_watching}},
  \code{\link{seasons_season}},
  \code{\link{seasons_summary}},
  \code{\link{shows_next_episode}}

Other people data: \code{\link{media_lists}},
  \code{\link{people_media}}, \code{\link{people_summary}}
}
\concept{episode data}
\concept{movie data}
\concept{people data}
\concept{season data}
\concept{show data}
