% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_ratings.R
\name{media_ratings}
\alias{media_ratings}
\alias{shows_ratings}
\alias{movies_ratings}
\alias{seasons_ratings}
\alias{episodes_ratings}
\title{Media user ratings}
\source{
\href{https://trakt.docs.apiary.io/#reference/shows/ratings/get-show-ratings}{shows/:ids/ratings}

\href{https://trakt.docs.apiary.io/#reference/movies/ratings/get-movie-ratings}{movies/:id/ratings}

\href{https://trakt.docs.apiary.io/#reference/seasons/ratings/get-season-ratings}{shows/:id/seasons/:season/ratings}

\href{https://trakt.docs.apiary.io/#reference/episodes/ratings/get-episode-ratings}{shows/:id/seasons/:season/episodes/:episode/ratings}
}
\usage{
shows_ratings(id)

movies_ratings(id)

seasons_ratings(id, season = 1L)

episodes_ratings(id, season = 1L, episode = 1L)
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}

\item{season}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}

\item{episode}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Returns a movie's or show's (or season's, or episode's) rating and ratings distribution.
If you \emph{do not} want the full ratings distribution, it is highly advised to
just use \code{*_summary} functions or \link{seasons_season} for episode ratings.
}
\note{
Since this function is able to work on multi-length inputs for
\code{id}, \code{season} and \code{episode}, it is possible to get a lot of data, \emph{but} at the cost
of one API call \emph{per element in each argument}. Please be kind to the API.
}
\examples{
# A movie's ratings
movies_ratings("tron-legacy-2010")

# A show's ratings
shows_ratings("game-of-thrones")
\dontrun{
# Ratings for seasons 1 through 5
seasons_ratings("futurama", season = 1:5)

# Ratings for episodes 1 through 7 of season 1
episodes_ratings("futurama", season = 1, episode = 1:7)
}
}
\seealso{
Other show data: \code{\link{media_aliases}},
  \code{\link{media_comments}}, \code{\link{media_people}},
  \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{shows_next_episode}},
  \code{\link{shows_related}}, \code{\link{shows_summary}}

Other movie data: \code{\link{media_aliases}},
  \code{\link{media_comments}}, \code{\link{media_people}},
  \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{movies_boxoffice}},
  \code{\link{movies_related}},
  \code{\link{movies_releases}},
  \code{\link{movies_summary}}

Other season data: \code{\link{media_comments}},
  \code{\link{media_people}}, \code{\link{media_stats}},
  \code{\link{seasons_season}},
  \code{\link{seasons_summary}}

Other episode data: \code{\link{episodes_summary}},
  \code{\link{media_comments}}, \code{\link{media_people}},
  \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{seasons_season}},
  \code{\link{seasons_summary}},
  \code{\link{shows_next_episode}}
}
\concept{episode data}
\concept{movie data}
\concept{season data}
\concept{show data}
