% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-media-related.R
\name{media_related}
\alias{media_related}
\alias{trakt.movies.related}
\alias{trakt.shows.related}
\title{Search for related shows or movies}
\usage{
trakt.movies.related(target, limit = 10L, extended = c("min", "full"))

trakt.shows.related(target, limit = 10L, extended = c("min", "full"))
}
\arguments{
\item{target}{\code{character(1)}: The ID of the item requested. Preferably the
\code{Trakt ID} (e.g. \code{1390}), \code{slug} (e.g. \code{"game-of-thrones"}), or \code{IMDb ID} (e.g. \code{"tt0944947"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and the result is combined.}

\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced to \code{integer}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Receive a set of media items that are related to a specific show or movie.
}
\examples{
trakt.shows.related("breaking-bad", limit = 5)
}
