% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_related.R
\name{media_related}
\alias{media_related}
\alias{movies_related}
\alias{shows_related}
\title{Search for related shows or movies}
\usage{
movies_related(target, limit = 10L, extended = c("min", "full"))

shows_related(target, limit = 10L, extended = c("min", "full"))
}
\arguments{
\item{target}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and the result is combined.}

\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced to \code{integer}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Receive a set of media items that are related to a specific show or movie.
}
\details{
The API methods wrapped are:
\itemize{
\item \href{https://trakt.docs.apiary.io/#reference/shows/related/get-related-shows}{/shows/:id/related}
\item \href{https://trakt.docs.apiary.io/#reference/movies/related/get-related-movies}{/movies/:id/related}
}
}
\examples{
shows_related("breaking-bad", limit = 5)
}
