% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-media-stats.R
\name{media_stats}
\alias{media_stats}
\alias{trakt.shows.stats}
\alias{trakt.movies.stats}
\alias{trakt.season.stats}
\alias{trakt.episodes.stats}
\title{Get a show or movie's stats}
\usage{
trakt.shows.stats(target)

trakt.movies.stats(target)

trakt.season.stats(target, season = 1L)

trakt.episodes.stats(target, season = 1L, episode = 1L)
}
\arguments{
\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"} for a show, or \code{"bryan-cranston"} for a person),
\code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"} or \code{"nm0186505"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df} if possible.}

\item{season}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined via \link[purrr:map_df]{purrr::map_df}. This may result in \emph{a lot} of API calls. Use wisely.}

\item{episode}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined via \link[purrr:map_df]{purrr::map_df}. This may result in \emph{a lot} of API calls. Use wisely.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
The data contains watchers, playes, collectors, comments, lists, and votes.
}
\examples{
# Stats for a movie
trakt.movies.stats("inception-2010")
\dontrun{
# Stats for multiple shows at once
trakt.shows.stats(c("breaking-bad", "game-of-thrones"))

# Stats for multiple episodes
trakt.episodes.stats("futurama", season = 1, episode = 1:7)
}
}
