% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_summary.R
\name{media_summary}
\alias{media_summary}
\alias{movies_summary}
\alias{shows_summary}
\title{Get show or movie summary info}
\usage{
movies_summary(id, extended = c("min", "full"))

shows_summary(id, extended = c("min", "full"))
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Get show or movie summary info
}
\details{
These functions wrap the "single item" API methods:
\itemize{
\item \href{https://trakt.docs.apiary.io/#reference/shows/summary/get-a-single-show}{/shows/:id}
\item \href{https://trakt.docs.apiary.io/#reference/movies/summary/get-a-movie}{/movies/:id}
}
}
\note{
These functions return the same amount of information as \link{trakt.search}.
}
\examples{
# Minimal info by default
shows_summary("breaking-bad")
\dontrun{
# More information
shows_summary("breaking-bad", extended = "full")

# Info for multiple movies
movies_summary(c("inception-2010", "the-dark-knight-2008"), extended = "full")
}
}
\concept{summary data}
