% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_watching.R
\name{media_watching}
\alias{media_watching}
\alias{movies_watching}
\alias{shows_watching}
\alias{seasons_watching}
\alias{episodes_watching}
\title{Get who's watching a thing right now}
\source{
\code{movies_watching()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/movies/watching/get-users-watching-right-now}{movies/:id/watching}.

\code{shows_watching()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/shows/watching/get-users-watching-right-now}{shows/:id/watching}.

\code{episodes_watching()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/episodes/watching/get-users-watching-right-now}{shows/:id/seasons/:season/episodes/:episode/watching}.
}
\usage{
movies_watching(id, extended = c("min", "full"))

shows_watching(id, extended = c("min", "full"))

seasons_watching(id, season = 1L, extended = c("min", "full"))

episodes_watching(id, season = 1L, episode = 1L, extended = c("min",
  "full"))
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}

\item{season}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}

\item{episode}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get who's watching a thing right now
}
\section{Functions}{
\itemize{
\item \code{movies_watching}: Who's watching a movie.

\item \code{shows_watching}: Who's watching a show.

\item \code{seasons_watching}: Who's watching a season.

\item \code{episodes_watching}: Who's watching an episode.
}}

\examples{
\dontrun{
movies_watching("deadpool-2016")
shows_watching("the-simpsons")
seasons_watching("the-simpsons", season = 6)
episodes_watching("the-simpsons", season = 6, episode = 12)
}
}
\seealso{
Other movie data: \code{\link{anticipated_media}},
  \code{\link{collected_media}},
  \code{\link{media_aliases}},
  \code{\link{media_comments}}, \code{\link{media_lists}},
  \code{\link{media_people}}, \code{\link{media_ratings}},
  \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{movies_boxoffice}},
  \code{\link{movies_related}},
  \code{\link{movies_releases}},
  \code{\link{movies_summary}}, \code{\link{people_media}},
  \code{\link{played_media}}, \code{\link{popular_media}},
  \code{\link{trending_media}},
  \code{\link{updated_media}}, \code{\link{watched_media}}

Other show data: \code{\link{collected_media}},
  \code{\link{media_aliases}},
  \code{\link{media_comments}}, \code{\link{media_lists}},
  \code{\link{media_people}}, \code{\link{media_ratings}},
  \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{people_media}}, \code{\link{played_media}},
  \code{\link{shows_next_episode}},
  \code{\link{shows_related}}, \code{\link{shows_summary}},
  \code{\link{updated_media}}

Other episode data: \code{\link{episodes_summary}},
  \code{\link{media_comments}}, \code{\link{media_lists}},
  \code{\link{media_people}}, \code{\link{media_ratings}},
  \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{seasons_season}},
  \code{\link{seasons_summary}},
  \code{\link{shows_next_episode}}
}
\concept{episode data}
\concept{movie data}
\concept{show data}
