% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_watching.R
\name{media_watching}
\alias{media_watching}
\alias{movies_watching}
\alias{shows_watching}
\alias{seasons_watching}
\alias{episodes_watching}
\title{Get who's watching a thing right now}
\usage{
movies_watching(id, extended = c("min", "full"))

shows_watching(id, extended = c("min", "full"))

seasons_watching(id, season = 1L, extended = c("min", "full"))

episodes_watching(id, season = 1L, episode = 1L, extended = c("min",
  "full"))
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}

\item{season}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}

\item{episode}{\code{integer(1) [1L]}: The season and eisode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get who's watching a thing right now
}
\section{Functions}{
\itemize{
\item \code{movies_watching}: Who's watching a movie.

\item \code{shows_watching}: Who's watching a show.

\item \code{seasons_watching}: Who's watching a season.

\item \code{episodes_watching}: Who's watching an episode.
}}

\examples{
\dontrun{
movies_watching("deadpool-2016")
shows_watching("the-simpsons")
seasons_watching("the-simpsons", season = 6)
episodes_watching("the-simpsons", season = 6, episode = 12)
}
}
