% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_boxoffice.R
\name{movies_boxoffice}
\alias{movies_boxoffice}
\title{Get the weekend box office}
\source{
\code{movies_boxoffice()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/movies/box-office/get-the-weekend-box-office}{movies/boxoffice}.
}
\usage{
movies_boxoffice(extended = c("min", "full"))
}
\arguments{
\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Returns the top 10 grossing movies in the U.S. box office last weekend.
Updated every Monday morning.
}
\examples{
movies_boxoffice()
}
\seealso{
Other movie data: 
\code{\link{anticipated_media}},
\code{\link{collected_media}},
\code{\link{media_aliases}},
\code{\link{media_comments}},
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{media_ratings}()},
\code{\link{media_stats}()},
\code{\link{media_translations}},
\code{\link{media_watching}},
\code{\link{movies_related}()},
\code{\link{movies_releases}()},
\code{\link{movies_summary}()},
\code{\link{people_media}()},
\code{\link{played_media}},
\code{\link{popular_media}},
\code{\link{trending_media}},
\code{\link{updated_media}},
\code{\link{watched_media}}
}
\concept{movie data}
