% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_related.R
\name{movies_related}
\alias{movies_related}
\title{Get similiar(ish) movies}
\source{
\href{https://trakt.docs.apiary.io/#reference/movies/related/get-related-movies}{movies/:id/related}
}
\usage{
movies_related(id, limit = 10L, extended = c("min", "full"))
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}

\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced via \code{as.integer()}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get similiar(ish) movies
}
\examples{
movies_related("the-avengers-2012", limit = 5)
}
\seealso{
Other movie data: \code{\link{media_aliases}},
  \code{\link{media_comments}}, \code{\link{media_people}},
  \code{\link{media_ratings}}, \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{movies_boxoffice}},
  \code{\link{movies_releases}},
  \code{\link{movies_summary}}
}
\concept{movie data}
