% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies_releases.R
\name{movies_releases}
\alias{movies_releases}
\title{Get a movie's release details}
\source{
\code{movies_releases()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/movies/releases/get-all-movie-releases}{movies/:id/releases/:country}.
}
\usage{
movies_releases(id, country = NULL)
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}

\item{country}{Optional two letter country code to filter by. See \link{trakt_countries}
for a table of country codes.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Retrieve one or more movie's release information,
including the release date, country code (two letter, e.g. \code{us}), and
the certification (e.g. \code{PG}).
}
\examples{
\dontrun{
movies_releases("tron-legacy-2010")
}
}
\seealso{
Other movie data: 
\code{\link{anticipated_media}},
\code{\link{collected_media}},
\code{\link{media_aliases}},
\code{\link{media_comments}},
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{media_ratings}()},
\code{\link{media_stats}()},
\code{\link{media_translations}},
\code{\link{media_watching}},
\code{\link{movies_boxoffice}()},
\code{\link{movies_related}()},
\code{\link{movies_summary}()},
\code{\link{people_media}()},
\code{\link{played_media}},
\code{\link{popular_media}},
\code{\link{trending_media}},
\code{\link{updated_media}},
\code{\link{watched_media}}
}
\concept{movie data}
