% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_summary.R
\name{movies_summary}
\alias{movies_summary}
\title{Get a single movie}
\source{
movies_summary() wraps endpoint \href{https://trakt.docs.apiary.io/#reference/movies/summary/get-a-movie}{movies/:id}
}
\usage{
movies_summary(id, extended = c("min", "full"))
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get a single movie
}
\examples{
# Minimal info by default
movies_summary("inception-2010")
\dontrun{
# Full information,  multiple movies
movies_summary(c("inception-2010", "the-dark-knight-2008"), extended = "full")
}
}
\seealso{
Other movie data: \code{\link{anticipated_media}},
  \code{\link{collected_media}},
  \code{\link{media_aliases}},
  \code{\link{media_comments}}, \code{\link{media_lists}},
  \code{\link{media_people}}, \code{\link{media_ratings}},
  \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{media_watching}},
  \code{\link{movies_boxoffice}},
  \code{\link{movies_related}},
  \code{\link{movies_releases}},
  \code{\link{people_media}}, \code{\link{played_media}},
  \code{\link{popular_media}},
  \code{\link{trending_media}},
  \code{\link{updated_media}}, \code{\link{watched_media}}

Other summary methods: \code{\link{comments_comment}},
  \code{\link{episodes_summary}},
  \code{\link{people_summary}},
  \code{\link{seasons_summary}},
  \code{\link{shows_summary}}, \code{\link{user_profile}}
}
\concept{movie data}
\concept{summary methods}
