% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-internal.R
\docType{data}
\name{networks}
\alias{networks}
\title{Available networks}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1328 rows and 1 columns.}
\usage{
networks
}
\description{
Available networks
}
\examples{
networks
}
\keyword{datasets}
