% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/misc.R
\name{pad}
\alias{pad}
\title{Easy episode id padding}
\usage{
pad(s = "0", e = "0", width = 2)
}
\arguments{
\item{s}{Input season number, coerced to \code{character}.}

\item{e}{Input episode number, coerced to \code{character}.}

\item{width}{The length of the padding. Defaults to 2.}
}
\value{
A \code{character} in standard \code{sXXeYY} format
}
\description{
Simple function to ease the creation of \code{sXXeYY} episode ids.
}
\note{
See \href{http://trakt.tv/api-docs/show-stats}{the trakt API docs for further info}
}
\examples{
pad(2, 4)
}

