% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pad}
\alias{pad}
\title{Easy episode id padding}
\usage{
pad(s = "0", e = "0", width = 2)
}
\arguments{
\item{s}{Input season number, coerced to \code{character}.}

\item{e}{Input episode number, coerced to \code{character}.}

\item{width}{The length of the padding. Defaults to 2.}
}
\value{
A \code{character} in standard \code{sXXeYY} format
}
\description{
Simple function to ease the creation of \code{sXXeYY} episode ids.
}
\note{
I like my sXXeYY format, okay?
}
\examples{
pad(2, 4) # Returns "s02e04"
}
\seealso{
Other utility functions: \code{\link{build_trakt_url}},
  \code{\link{parse_trakt_url}}
}
\concept{utility functions}
