% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_media.R
\name{people_media}
\alias{people_media}
\alias{people_movies}
\alias{people_shows}
\title{Get a single person's movie or show credits}
\source{
\code{people_movies()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/people/movies/get-movie-credits}{people/:id/movies}.

\code{people_shows()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/people/shows/get-show-credits}{people/:id/shows}.
}
\usage{
people_movies(id, extended = c("min", "full"))

people_shows(id, extended = c("min", "full"))
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things")} for more details.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things")} for more details.}
}
\value{
A \code{list} of one or more \link[tibble:tibble-package]{tibbles} for \code{cast}
and \code{crew}. The latter \code{tibble} objects are as flat as possible.
}
\description{
Returns all movies or shows where this person is in the cast or crew.
}
\details{
Note that as of 2019-09-30, there are two representations of \code{character[s]} and \code{job[s]}:
One is a regular character variable, and the other is a list-column. The singular is
\href{https://github.com/trakt/api-help/issues/74}{deprecated and only included for compatibility reasons}.
}
\examples{
\dontrun{
people_movies("christopher-nolan")

people_shows("kit-harington")
}
}
\seealso{
\link{media_people}, for the other direction: Media that has people.

Other movie data: 
\code{\link{anticipated_media}},
\code{\link{collected_media}},
\code{\link{media_aliases}},
\code{\link{media_comments}},
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{media_ratings}()},
\code{\link{media_stats}()},
\code{\link{media_translations}},
\code{\link{media_watching}},
\code{\link{movies_boxoffice}()},
\code{\link{movies_related}()},
\code{\link{movies_releases}()},
\code{\link{movies_summary}()},
\code{\link{played_media}},
\code{\link{popular_media}},
\code{\link{trending_media}},
\code{\link{watched_media}}

Other people data: 
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{people_summary}()}

Other show data: 
\code{\link{collected_media}},
\code{\link{media_aliases}},
\code{\link{media_comments}},
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{media_ratings}()},
\code{\link{media_stats}()},
\code{\link{media_translations}},
\code{\link{media_watching}},
\code{\link{played_media}},
\code{\link{shows_next_episode}()},
\code{\link{shows_related}()},
\code{\link{shows_summary}()}

Other people data: 
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{people_summary}()}
}
\concept{movie data}
\concept{people data}
\concept{show data}
