% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-people.R
\name{people_media}
\alias{people_media}
\alias{trakt.people.movies}
\alias{trakt.people.shows}
\title{Get a single person's movie or show credits}
\usage{
trakt.people.movies(target, extended = c("min", "full"))

trakt.people.shows(target, extended = c("min", "full"))
}
\arguments{
\item{target}{\code{character(1)}: The ID of the item requested. Preferably the
\code{Trakt ID} (e.g. \code{1390}), \code{slug} (e.g. \code{"game-of-thrones"}), or \code{IMDb ID} (e.g. \code{"tt0944947"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and the result is combined.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \code{list} of one or more \link[tibble:tibble-package]{tibbles} for \code{cast}
and \code{crew}. The latter \code{tibble} objects are as flat as possible.
}
\description{
Returns all movies or shows where this person is in the cast or crew.
}
\examples{
\dontrun{
trakt.people.movies("christopher-nolan")

trakt.people.shows("kit-harington")
}
}
\seealso{
\link{media_people}, for the other direction: Media that has people.

Other people data: \code{\link{media_people}},
  \code{\link{trakt.people.summary}}
}
\concept{people data}
