% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_media.R
\name{people_media}
\alias{people_media}
\alias{people_movies}
\alias{people_shows}
\title{Get a single person's movie or show credits}
\usage{
people_movies(id, extended = c("min", "full"))

people_shows(id, extended = c("min", "full"))
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See section
\link[=trakt_api_common_parameters]{"Item identifiers"} for more details.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See section \link[=trakt_api_common_parameters]{"Extended Information"} for more details.}
}
\value{
A \code{list} of one or more \link[tibble:tibble-package]{tibbles} for \code{cast}
and \code{crew}. The latter \code{tibble} objects are as flat as possible.
}
\description{
Returns all movies or shows where this person is in the cast or crew.
}
\details{
The API methods wrapped are
\itemize{
\item \href{https://trakt.docs.apiary.io/#reference/people/movies/get-movie-credits}{/people/:id/movies}
\item \href{https://trakt.docs.apiary.io/#reference/people/shows/get-show-credits}{/people/:id/shows}
}

Note that as of 2019-09-30, there are two representations of \code{character[s]} and \code{job[s]}:
One is a regular character variable, and the other is a list-column. The singular is
\href{https://github.com/trakt/api-help/issues/74}{deprecated and only included for compatibility reasons}.
}
\examples{
\dontrun{
people_movies("christopher-nolan")

people_shows("kit-harington")
}
}
\seealso{
\link{media_people}, for the other direction: Media that has people.

Other people data: \code{\link{media_people}},
  \code{\link{people_summary}}
}
\concept{people data}
