% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people_summary.R
\name{people_summary}
\alias{people_summary}
\title{Get a single person's details}
\source{
\code{people_summary()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/people/summary/get-a-single-person}{people/:id}.
}
\usage{
people_summary(id, extended = c("min", "full"))
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get a single person's details, like their various IDs. If \code{extended} is
\code{"full"}, there will also be biographical data if available, e.g. their
birthday.
}
\examples{
# A single person's extended information
people_summary("bryan-cranston", "full")

# Multiple people
people_summary(c("kit-harington", "emilia-clarke"))
}
\seealso{
Other people data: \code{\link{media_lists}},
  \code{\link{media_people}}, \code{\link{people_media}}

Other summary methods: \code{\link{comments_comment}},
  \code{\link{episodes_summary}},
  \code{\link{movies_summary}},
  \code{\link{seasons_summary}},
  \code{\link{shows_summary}}, \code{\link{user_profile}}
}
\concept{people data}
\concept{summary methods}
