% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons_season.R
\name{seasons_season}
\alias{seasons_season}
\title{Get a season of a show}
\usage{
seasons_season(id, seasons = 1L, extended = c("min", "full"))
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}

\item{seasons}{\code{integer(1) [1L]}: The season(s) to get. Use \code{0} for specials.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Similar to \link{seasons_summary}, but this function returns full data for
a single season, i.e. all the episodes of the season
}
\details{
This function accesses \href{https://trakt.docs.apiary.io/#reference/seasons/season/get-single-season-for-a-show}{this API method}
at the endpoint \code{/shows/:show_id/seasons/:season_num}.
}
\note{
If you want to quickly gather episode data of all available seasons,
see \link{seasons_summary} and use the \code{episodes = TRUE} parameter.
}
\examples{
\dontrun{
seasons_season("breaking-bad", 1)

# Including all episode data:
seasons_season("breaking-bad", 1, extended = "full")
}
}
\seealso{
Other show data: \code{\link{episodes_summary}},
  \code{\link{seasons_summary}}
}
\concept{show data}
