% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasons_summary.R
\name{seasons_summary}
\alias{seasons_summary}
\title{Get a show's seasons}
\source{
\href{https://trakt.docs.apiary.io/#reference/seasons/summary/get-all-seasons-for-a-show}{shows/:id/seasons/seasons}
}
\usage{
seasons_summary(id, episodes = FALSE, drop_specials = TRUE,
  drop_unaired = TRUE, extended = c("min", "full"))
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}

\item{episodes}{\code{logical(1) [FALSE]}: If \code{TRUE}, all episodes for each season
are appended as a list-column, with the amount of variables depending on \code{extended}.}

\item{drop_specials}{\code{logical(1) [TRUE]}: Special episodes (season 0) are dropped}

\item{drop_unaired}{\code{logical(1) [TRUE]}: Seasons without aired episodes are dropped.
Only works if \code{extended} is \code{"full"}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get details for a show's seasons, e.g. how many seasons there are and  how many epsiodes
each season has. With \code{episodes == TRUE} and \code{extended == "full"}, this function
is also suitable to retrieve all episode data for all seasons of a show with
just a single API call.
}
\examples{
# Get just the season numbers and their IDs
seasons_summary("breaking-bad", extended = "min")
\dontrun{
# Get season numbers, ratings, votes, titles and other metadata as well as
# a list-column containing all episode data
seasons_summary("utopia", extended = "full", episodes = TRUE)
}
}
\seealso{
Other season data: \code{\link{media_people}},
  \code{\link{media_ratings}}, \code{\link{media_stats}},
  \code{\link{seasons_season}}

Other episode data: \code{\link{episodes_summary}},
  \code{\link{media_people}}, \code{\link{media_ratings}},
  \code{\link{media_stats}}, \code{\link{seasons_season}}

Other summary methods: \code{\link{episodes_summary}},
  \code{\link{movies_summary}},
  \code{\link{people_summary}}, \code{\link{shows_summary}}
}
\concept{episode data}
\concept{season data}
\concept{summary methods}
