% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shows_nest_episode.R
\name{shows_next_episode}
\alias{shows_next_episode}
\alias{shows_last_episode}
\title{Get a shows next or latest episode}
\source{
\code{shows_next_episode()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/shows/next-episode/get-next-episode}{shows/:id/next_episode}.

\code{shows_last_episode()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/shows/last-episode/get-last-episode}{shows/:id/last_episode}.
}
\usage{
shows_next_episode(id, extended = c("min", "full"))

shows_last_episode(id, extended = c("min", "full"))
}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things", package = "tRakt")} for more details.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get a shows next or latest episode
}
\examples{
shows_next_episode("one-piece")
shows_last_episode("one-piece")
}
\seealso{
Other show data: 
\code{\link{collected_media}},
\code{\link{media_aliases}},
\code{\link{media_comments}},
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{media_ratings}()},
\code{\link{media_stats}()},
\code{\link{media_translations}},
\code{\link{media_watching}},
\code{\link{people_media}()},
\code{\link{played_media}},
\code{\link{shows_related}()},
\code{\link{shows_summary}()},
\code{\link{updated_media}}

Other episode data: 
\code{\link{episodes_summary}()},
\code{\link{media_comments}},
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{media_ratings}()},
\code{\link{media_stats}()},
\code{\link{media_translations}},
\code{\link{media_watching}},
\code{\link{seasons_season}()},
\code{\link{seasons_summary}()}

Other show data: 
\code{\link{collected_media}},
\code{\link{media_aliases}},
\code{\link{media_comments}},
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{media_ratings}()},
\code{\link{media_stats}()},
\code{\link{media_translations}},
\code{\link{media_watching}},
\code{\link{people_media}()},
\code{\link{played_media}},
\code{\link{shows_related}()},
\code{\link{shows_summary}()},
\code{\link{updated_media}}

Other episode data: 
\code{\link{episodes_summary}()},
\code{\link{media_comments}},
\code{\link{media_lists}},
\code{\link{media_people}},
\code{\link{media_ratings}()},
\code{\link{media_stats}()},
\code{\link{media_translations}},
\code{\link{media_watching}},
\code{\link{seasons_season}()},
\code{\link{seasons_summary}()}
}
\concept{episode data}
\concept{show data}
