% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tRakt-package.R
\docType{package}
\name{tRakt-package}
\alias{tRakt}
\alias{tRakt-package}
\title{tRakt: Get Data from 'trakt.tv'}
\description{
A wrapper for the <https://trakt.tv> API to retrieve data
    about shows and movies, including user ratings, credits and related
    metadata. Additional functions retrieve user-related information
    including collections and history of watched items. A full
    API reference is available at <https://trakt.docs.apiary.io>.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/jemus42/tRakt}
  \item Report bugs at \url{http://github.com/jemus42/tRakt/issues}
}

}
\author{
\strong{Maintainer}: Lukas Burk \email{lukas@quantenbrot.de} (0000-0001-7528-3795)

}
\keyword{internal}
