% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{trakt.api.call}
\alias{trakt.api.call}
\title{Make an API call to any URL}
\usage{
trakt.api.call(url, headers = getOption("trakt.headers"),
  convert.datetime = TRUE)
}
\arguments{
\item{url}{APIv2 method. See \href{http://docs.trakt.apiary.io/}{the trakt API}.}

\item{headers}{HTTP headers to set. Must be result of \code{httr::add_headers}.
Default value is \code{getOption("trakt.headers")} set by \link[tRakt]{get_trakt_credentials}.}

\item{convert.datetime}{If \code{TRUE} (default), known top-level datetime variables
are converted to \code{POSIXct}. This might miss some variables and does not recurse
into nested lists or list-columns.}
}
\value{
The parsed content of the API response.
An empty \link[tibble:tibble-package]{tibble} if the response is an empty array.
}
\description{
\code{trakt.api.call} makes an API call to a specified URL and returns the parsed output.
}
\note{
This function is heavily used internally, so why not expose it to the user.
}
\examples{
\dontrun{
library(tRakt)
trakt.api.call("https://api.trakt.tv/shows/breaking-bad")
}
}
\seealso{
Other API-basics: \code{\link{get_trakt_credentials}},
  \code{\link{trakt.search.byid}},
  \code{\link{trakt.search}}
}
\concept{API-basics}
