% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{trakt.api.call}
\alias{trakt.api.call}
\title{Make an API call to any URL}
\usage{
trakt.api.call(url, client.id = getOption("trakt.client.id"),
  convert.datetime = TRUE, HEAD = FALSE)
}
\arguments{
\item{url}{APIv2 method. See \href{http://docs.trakt.apiary.io/}{the trakt API}.}

\item{client.id}{API client id. see \link{trakt_credentials} for further information.}

\item{convert.datetime}{If \code{TRUE} (default), known top-level datetime variables
are converted to \code{POSIXct}. This might miss some variables and does not recurse
into nested lists or list-columns.}

\item{HEAD}{\code{logical(1) [FALSE]}: If \code{TRUE}, only a HTTP \code{HEAD} request is performed
and its content returned. This is useful if you are only interested in status codes
or other headers, and don't want to waste resources on additional bandwidth.}
}
\value{
The \link[jsonlite:fromJSON]{parsed} content of the API response.
An empty \link[tibble:tibble-package]{tibble} if the response is an empty array.
}
\description{
\code{trakt.api.call} makes an API call to a specified URL and returns the parsed output.
}
\note{
This function is heavily used internally, so why not expose it to the user.
}
\examples{
\dontrun{
trakt.api.call("https://api.trakt.tv/shows/breaking-bad")

trakt.api.call("https://api.trakt.tv/users/jemus42", HEAD = TRUE)
}
}
\seealso{
Other API-basics: \code{\link{trakt.search.byid}},
  \code{\link{trakt.search}},
  \code{\link{trakt_credentials}}
}
\concept{API-basics}
