% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{trakt.api.call}
\alias{trakt.api.call}
\title{Make an API call to any URL}
\usage{
trakt.api.call(url, client.id = getOption("trakt.client.id"),
  HEAD = FALSE)
}
\arguments{
\item{url}{APIv2 endpoint. See \href{http://docs.trakt.apiary.io/}{the trakt API}.}

\item{client.id}{API client id. see \link{trakt_credentials} for further information.}

\item{HEAD}{\code{logical(1) [FALSE]}: If \code{TRUE}, only a HTTP \code{HEAD} request is performed
and its content returned. This is useful if you are only interested in status codes
or other headers, and don't want to waste resources on additional bandwidth.}
}
\value{
The \link[jsonlite:fromJSON]{parsed} content of the API response.
An empty \link[tibble:tibble-package]{tibble()} if the response is an empty \code{JSON} array.
}
\description{
\code{trakt.api.call} makes an API call to a specified URL and returns the parsed output.
}
\examples{
# A simple request to a direct URL
trakt.api.call("https://api.trakt.tv/shows/breaking-bad")

# A HEAD-only request, useful for validating a URL exists or the API is accessable
trakt.api.call("https://api.trakt.tv/users/jemus42", HEAD = TRUE)

# Optionally be lazy about URL specification by dropping the hostname:
trakt.api.call("shows/game-of-thrones")
}
\seealso{
Other API-basics: \code{\link{trakt.search}},
  \code{\link{trakt_credentials}}
}
\concept{API-basics}
