% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trakt.getEpisodeData}
\alias{trakt.getEpisodeData}
\title{Get a show's episode information}
\usage{
trakt.getEpisodeData(target, show.episodes = NULL,
  apikey = getOption("trakt.apikey"), dropunaired = TRUE)
}
\arguments{
\item{target}{The \code{slug} or \code{tvdbid} of the show requested}

\item{show.episodes}{Show episodes dataset, normally provided by \link{initializeEpisodes}}

\item{apikey}{API-key used for the call.
Defaults to \code{getOption("trakt.apikey")}}

\item{dropunaired}{If \code{TRUE}, episodes which have not aired yet are dropped.}
}
\value{
A \code{data.frame} containing episode details
}
\description{
\code{trakt.getEpisodeData} pulls detailed episode data.
Get details for a show's episodes, e.g. ratings, number of votes,
airdates, urls, plot overviews…
}
\note{
See \href{http://trakt.tv/api-docs/show-episode-summary}{the trakt API docs for further info}
}
\examples{
\dontrun{
options(trakt.apikey = jsonlite::fromJSON("key.json")$apikey)
breakingbad.seasons  <- trakt.getSeasons("breaking-bad")
breakingbad.episodes <- initializeEpisodes(breakingbad.seasons)
breakingbad.episodes <- trakt.getEpisodeData("breaking-bad", breakingbad.episodes)
}
}

