% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trakt.getSeasons.R
\name{trakt.getSeasons}
\alias{trakt.getSeasons}
\title{Get a show's season information}
\usage{
trakt.getSeasons(target, extended = "full,images", dropspecials = TRUE)
}
\arguments{
\item{target}{The \code{slug} of the show requested}

\item{extended}{Defaults to \code{full,images} to get season posters. Can be
\code{min}, \code{images}, \code{full}, \code{full,images}}

\item{dropspecials}{If \code{TRUE} (default), special episodes (listed as 'season 0') are dropped
This is the default.}
}
\value{
A \code{data.frame} containing season details (nested in \code{list} objects)
}
\description{
\code{trakt.getSeasons} pulls season data.
Get details for a show's seasons, e.g. how many seasons there are, how many epsiodes
each season has, and season posters.
See \hred{http://docs.trakt.apiary.io/#introduction/extended-info} for possible values of
\code{extended} to customize output amount.
}
\note{
See \href{http://docs.trakt.apiary.io/reference/seasons/summary}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
breakingbad.seasons <- trakt.getSeasons("breaking-bad", extended = "min")
}
}

