% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.summary.R
\name{trakt.movie.summary}
\alias{trakt.movie.summary}
\title{Get a single movie's details}
\usage{
trakt.movie.summary(target, extended = c("min", "full"),
  force_data_frame = FALSE)
}
\arguments{
\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"}), \code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}

\item{force_data_frame}{If \code{TRUE}, the \code{list} is unnested as much as possible, resulting
in a flat \code{tibble} suitable to be \code{rbind}ed to other summary results.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
\code{trakt.movie.summary} returns a single movie's summary information.
}
\note{
See \href{http://docs.trakt.apiary.io/reference/movies/summary/get-a-movie}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
trakt.movie.summary("tron-legacy-2010")
}
}
\seealso{
Other movie data: \code{\link{trakt.movie.people}},
  \code{\link{trakt.movie.releases}},
  \code{\link{trakt.stats}}
}
\concept{movie data}
