% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trakt.popular.R
\name{trakt.movies.popular}
\alias{trakt.movies.popular}
\title{Get popular movies}
\usage{
trakt.movies.popular(limit = 10, page = 1, extended = "min")
}
\arguments{
\item{limit}{Number of movies to return. Is coerced to \code{integer} and must be greater than 0.}

\item{page}{Page to return (default is \code{1})
for \href{http://docs.trakt.apiary.io/#introduction/pagination}{pagination}.}

\item{extended}{Whether extended info should be provided.
Defaults to \code{"min"}, can either be \code{"min"} or \code{"full"}}
}
\value{
A \code{data.frame} containing popular movies with their name and ids
}
\description{
\code{trakt.movies.popular} returns a list of the most popular movies on trakt.tv.
According to the API docs, opularity is calculated based both ratings and number of ratings.
}
\note{
See \href{http://docs.trakt.apiary.io/reference/movies/popular/get-popular-movies}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
trakt.movies.popular(5)
}
}
\seealso{
Other aggregate: \code{\link{trakt.movies.related}};
  \code{\link{trakt.movies.trending}};
  \code{\link{trakt.shows.popular}};
  \code{\link{trakt.shows.related}};
  \code{\link{trakt.shows.trending}}

Other movie: \code{\link{trakt.movie.people}};
  \code{\link{trakt.movie.ratings}};
  \code{\link{trakt.movie.summary}};
  \code{\link{trakt.movies.related}};
  \code{\link{trakt.movies.trending}}
}

