% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.related.R
\name{trakt.movies.related}
\alias{trakt.movies.related}
\title{Search for related movies}
\usage{
trakt.movies.related(target, extended = c("min", "full"))
}
\arguments{
\item{target}{The \code{id} of the movie requested. Either the \code{slug}
(e.g. \code{"tron-legacy-2010"}), \code{trakt id} or \code{IMDb id}. If multiple \code{target}s are
provided, the results will be \code{rbind}ed together and a \code{source} column as appended,
containing the provided \code{id} of the input.}

\item{extended}{\code{[character(1)]}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
\code{trakt.movies.related} returns movies related to the input movie.
}
\details{
Receive a set of movies that are related to a specific movie.
}
\note{
See \href{http://docs.trakt.apiary.io/reference/movies/related/get-related-movies}{the trakt API docs for further info}
}
\examples{
\dontrun{
library(tRakt)
related <- trakt.movies.related("tron-legacy-2010")
}
}
\seealso{
Other movie data: \code{\link{trakt.movie.people}},
  \code{\link{trakt.movie.ratings}},
  \code{\link{trakt.movie.releases}},
  \code{\link{trakt.movie.summary}},
  \code{\link{trakt.stats}}

Other aggregated data: \code{\link{trakt.shows.related}}
}
\concept{aggregated data}
\concept{movie data}
