% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.movie.releases.R
\name{trakt.movies.releases}
\alias{trakt.movies.releases}
\title{Get a movie's release details}
\usage{
trakt.movies.releases(target, country = NULL)
}
\arguments{
\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"}), \code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df}.}

\item{country}{Optional two letter country code to filter by.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Retrieve one or more movie's release information,
including the release date, country code (two letter, e.g. \code{us}), and
the certification (e.g. \code{PG}).
}
\note{
See \href{http://docs.trakt.apiary.io/reference/movies/releases/get-all-movie-releases}{the
trakt API docs for further info}
}
\examples{
\dontrun{
trakt.movies.releases("tron-legacy-2010")
}
}
\seealso{
Other movie data: \code{\link{trakt.stats}}
}
\concept{movie data}
