% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.people.R
\name{trakt.people.media}
\alias{trakt.people.media}
\alias{trakt.people.movies}
\alias{trakt.people.shows}
\title{Get a single person's movie or show credits}
\usage{
trakt.people.media(type = c("shows", "movies"), target,
  extended = c("min", "full"))

trakt.people.movies(target, extended = c("min", "full"))

trakt.people.shows(target, extended = c("min", "full"))
}
\arguments{
\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}

\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"} for a show, or \code{"bryan-cranston"} for a person),
\code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"} or \code{"nm0186505"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df} if possible.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \code{list}.
}
\description{
Returns all movies or shows where this person is in the cast or crew.
}
\examples{
\dontrun{
trakt.people.movies("bryan-cranston")
}
}
\seealso{
\link{trakt.media.people}, for the other direction: Media that has people.

Other people data: \code{\link{trakt.media.people}},
  \code{\link{trakt.people.summary}}
}
\concept{people data}
