% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.people.R
\name{trakt.people.shows}
\alias{trakt.people.shows}
\title{Get a single person's show credits}
\usage{
trakt.people.shows(target, extended = "min")
}
\arguments{
\item{target}{The \code{id} of the person requested. Either the \code{slug}
(e.g. \code{"bryan-cranston"}), \code{trakt id} or \code{IMDb id}}

\item{extended}{Whether extended info should be provided.
Defaults to \code{"min"}, can either be \code{"min"} or \code{"full"}}
}
\value{
A \code{list}.
}
\description{
\code{trakt.people.shows} pulls show people data.
}
\details{
Returns all shows where this person is in the cast or crew.
}
\note{
See \href{http://docs.trakt.apiary.io/reference/people/shows/get-show-credits}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
person <- trakt.people.shows("bryan-cranston")
}
}
\seealso{
Other people data: \code{\link{trakt.movie.people}},
  \code{\link{trakt.people.movies}},
  \code{\link{trakt.people.summary}},
  \code{\link{trakt.show.people}}
}
\concept{people data}
