% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.people.R
\name{trakt.people.summary}
\alias{trakt.people.summary}
\title{Get a single person's details}
\usage{
trakt.people.summary(target, extended = c("min", "full"))
}
\arguments{
\item{target}{\code{character(1)}: The \code{id} of the person requested. Either the \code{slug}
(e.g. \code{"bryan-cranston"}), \code{Trakt ID} (e.g. \code{297737}) or \code{IMDB ID} (e.g. \code{"nm0186505"}).}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Get a single person's details, like their various IDs. If \code{extended} is \code{"full"},
there will also be biographical data if available.
}
\examples{
\dontrun{
person <- trakt.people.summary("bryan-cranston", "full")
}
}
\seealso{
Other people data: \code{\link{trakt.media.people}},
  \code{\link{trakt.people.media}}
}
\concept{people data}
