% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-people.R
\name{trakt.people.summary}
\alias{trakt.people.summary}
\title{Get a single person's details}
\usage{
trakt.people.summary(target, extended = c("min", "full"))
}
\arguments{
\item{target}{\code{character(1)}: The ID of the item requested. Preferably the
\code{Trakt ID} (e.g. \code{1390}), \code{slug} (e.g. \code{"game-of-thrones"}), or \code{IMDb ID} (e.g. \code{"tt0944947"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and the result is combined.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Get a single person's details, like their various IDs. If \code{extended} is \code{"full"},
there will also be biographical data if available.
}
\examples{
# A single person's extended information
trakt.people.summary("bryan-cranston", "full")

# Multiple people
trakt.people.summary(c("kit-harington", "emilia-clarke"))
}
\seealso{
Other people data: \code{\link{media_people}},
  \code{\link{people_media}}
}
\concept{people data}
