% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-people.R
\name{trakt.people.summary}
\alias{trakt.people.summary}
\title{Get a single person's details}
\usage{
trakt.people.summary(target, extended = c("min", "full"))
}
\arguments{
\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"} for a show, or \code{"bryan-cranston"} for a person),
\code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"} or \code{"nm0186505"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df} if possible.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Get a single person's details, like their various IDs. If \code{extended} is \code{"full"},
there will also be biographical data if available.
}
\examples{
# A single person's extended information
trakt.people.summary("bryan-cranston", "full")

# Multiple people
trakt.people.summary(c("kit-harington", "emilia-clarke"))
}
\seealso{
Other people data: \code{\link{media_people}},
  \code{\link{people_media}}
}
\concept{people data}
