% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-people.R
\name{trakt.people.summary}
\alias{trakt.people.summary}
\title{Get a single person's details}
\usage{
trakt.people.summary(target, extended = c("min", "full"))
}
\arguments{
\item{target}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and the result is combined.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Get a single person's details, like their various IDs. If \code{extended} is
\code{"full"}, there will also be biographical data if available, e.g. their
birthday.
}
\details{
This function wraps the API method
\href{https://trakt.docs.apiary.io/#reference/people/summary/get-a-single-person}{/people/:id}.
}
\examples{
# A single person's extended information
trakt.people.summary("bryan-cranston", "full")

# Multiple people
trakt.people.summary(c("kit-harington", "emilia-clarke"))
}
\seealso{
Other people data: \code{\link{media_people}},
  \code{\link{people_media}}
}
\concept{people data}
