% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.ratings.R
\name{trakt.ratings}
\alias{trakt.ratings}
\alias{trakt.shows.ratings}
\alias{trakt.movies.ratings}
\title{Get a single show's ratings}
\usage{
trakt.ratings(type = c("shows", "movies"), target)

trakt.shows.ratings(target)

trakt.movies.ratings(target)
}
\arguments{
\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}

\item{target}{The \code{id} of the show requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"}), \code{trakt id} or \code{IMDb id}}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Returns a single show's or movie's rating and ratings distribution.
}
\note{
See \href{http://docs.trakt.apiary.io/reference/shows/ratings/get-show-ratings}{the trakt API docs for further info}
}
\examples{
\dontrun{
trakt.movies.ratings("tron-legacy-2010")
trakt.shows.ratings("game-of-thrones")
}
}
