% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-media-related.R
\name{trakt.related}
\alias{trakt.related}
\alias{trakt.movies.related}
\alias{trakt.shows.related}
\title{Search for related shows or movies}
\usage{
trakt.related(target, type = c("shows", "movies"), limit = 10L,
  extended = c("min", "full"))

trakt.movies.related(target, limit = 10L, extended = c("min", "full"))

trakt.shows.related(target, limit = 10L, extended = c("min", "full"))
}
\arguments{
\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"} for a show, or \code{"bryan-cranston"} for a person),
\code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"} or \code{"nm0186505"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df} if possible.}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}

\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced to \code{integer}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Receive a set of media items that are related to a specific show or movie.
}
\examples{
\dontrun{
trakt.related("breaking-bad", "shows", limit = 5)
}
}
