% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.related.R
\name{trakt.related}
\alias{trakt.related}
\alias{trakt.movies.related}
\alias{trakt.shows.related}
\title{Search for related shows or movies}
\usage{
trakt.related(target, type = c("shows", "movies"), extended = c("min",
  "full"))

trakt.movies.related(target, extended = c("min", "full"))

trakt.shows.related(target, extended = c("min", "full"))
}
\arguments{
\item{target}{The \code{id} of the show/movie requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"}), \code{trakt id} or \code{IMDb id}}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Receive a set of shows that are related to a specific show/movie
}
\examples{
\dontrun{
trakt.related("breaking-bad", "shows")
}
}
