% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.related.R
\name{trakt.related}
\alias{trakt.related}
\title{Search for related shows or movies}
\usage{
trakt.related(target, type, extended = c("min", "full"))
}
\arguments{
\item{target}{The \code{id} of the show/movie requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"}), \code{trakt id} or \code{IMDb id}}

\item{extended}{\code{[character(1)]}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\description{
\code{trakt.related} returns shows or movies related to the input show/movie.
Receive a set of shows that are related to a specific show/movie
}
\keyword{internal}
