% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trakt.related.show.R
\name{trakt.related.show}
\alias{trakt.related.show}
\title{Search for related shows}
\usage{
trakt.related.show(target)
}
\arguments{
\item{target}{The show's \code{slug} to be used.}
}
\value{
A \code{data.frame} containing search results
}
\description{
\code{trakt.related.show} returns related shows to the search query.
}
\details{
Search for related shows to the show you specified.
}
\note{
See \href{http://docs.trakt.apiary.io/#reference/shows/related/get-related-shows}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
related <- trakt.related.show(78874)
}
}

