% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trakt.search}
\alias{trakt.search}
\title{Search for a show}
\usage{
trakt.search(query, apikey = getOption("trakt.apikey"), limit = 1)
}
\arguments{
\item{query}{The keyword used for the search. Should be as URL-compatible as possible.}

\item{apikey}{API-key used for the call. Defaults to \code{getOption("trakt.apikey")}}

\item{limit}{The number of results to be returned. Defaults to 1.}
}
\value{
A \code{data.frame} containing search results
}
\description{
\code{trakt.search} pulls show stats and returns it compactly.
}
\details{
Search for a show with a keyword (e.g. \code{"Breaking Bad"}) and receive basic info of the
first search result. It's main use is to retrieve the tvdbid or proper show title for further use,
as well as receiving a quick overview of a show.
}
\note{
See \href{http://trakt.tv/api-docs/search-shows}{the trakt API docs for further info}
}
\examples{
\dontrun{
options(trakt.apikey = jsonlite::fromJSON("key.json")$apikey)
breakingbad <- trakt.search("Breaking Bad")
}
}

