% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trakt.search.R
\name{trakt.search}
\alias{trakt.search}
\title{Search for a show via text query}
\usage{
trakt.search(query, type = "show")
}
\arguments{
\item{query}{The keyword used for the search. Should be as URL-compatible as possible.}

\item{type}{The type of data you're looking for. Defaults to \code{show}}
}
\value{
A \code{data.frame} containing a single search result. Hopefully the one you wanted.
If no result is foun, the return value is \code{list(error = "Nothing found")} and a \code{warning}
}
\description{
\code{trakt.search} pulls show stats and returns it compactly.
}
\details{
Search for a show with a keyword (e.g. \code{"Breaking Bad"}) and receive basic info of the
first search result. It's main use is to retrieve the tvdbid or proper show title for further use,
as well as receiving a quick overview of a show.
}
\note{
See \href{http://docs.trakt.apiary.io/reference/search/text-query}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
breakingbad <- trakt.search("Breaking Bad")
}
}

