% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-seasons.R
\name{trakt.seasons.season}
\alias{trakt.seasons.season}
\title{Get a season of a show}
\usage{
trakt.seasons.season(target, seasons = 1L, extended = c("min", "full"))
}
\arguments{
\item{target}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and the result is combined.}

\item{seasons}{\code{integer(1) [1L]}: The season(s) to get. Use \code{0} for specials.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
\code{trakt.seasons.season} gets a season's full data.
Similar to \link{trakt.seasons.summary}, but this function returns full data for
a single season, i.e. all the episodes of the season
}
\details{
This function accesses \href{https://trakt.docs.apiary.io/#reference/seasons/season/get-single-season-for-a-show}{this API method}
at the endpoint \code{/shows/:show_id/seasons/:season_num}.
}
\note{
If you want to quickly gather episode data of all available seasons,
see \link{trakt.seasons.summary} and use the \code{episodes = TRUE} parameter.
}
\examples{
\dontrun{
trakt.seasons.season("breaking-bad", 1)
}
}
\seealso{
Other show data: \code{\link{trakt.seasons.summary}}
}
\concept{show data}
