% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.seasons.R
\name{trakt.seasons.summary}
\alias{trakt.seasons.summary}
\title{Get a show's season information}
\usage{
trakt.seasons.summary(target, extended = c("min", "full"),
  drop.specials = TRUE, drop.unaired = TRUE)
}
\arguments{
\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"}), \code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}

\item{drop.specials}{\code{logical(1) [TRUE]}: Special episodes (season 0) are dropped}

\item{drop.unaired}{\code{logical(1) [TRUE]}: Seasons without aired episodes are dropped.
Only works if \code{extended} is \code{"full"}.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Get details for a show's seasons, e.g. how many seasons there are and  how many epsiodes
each season has.
}
\examples{
\dontrun{
breakingbad.seasons <- trakt.seasons.summary("breaking-bad", extended = "min")
breakingbad.seasons.more <- trakt.seasons.summary("breaking-bad", extended = "full")
}
}
\seealso{
Other show data: \code{\link{trakt.get_all_episodes}},
  \code{\link{trakt.get_full_showdata}},
  \code{\link{trakt.seasons.season}},
  \code{\link{trakt.stats}}
}
\concept{show data}
