% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.seasons.R
\name{trakt.seasons.summary}
\alias{trakt.seasons.summary}
\title{Get a show's season information}
\usage{
trakt.seasons.summary(target, extended = c("min", "full"),
  drop.specials = TRUE, drop.unaired = TRUE)
}
\arguments{
\item{target}{The \code{id} of the show requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"}), \code{trakt id} or \code{IMDb id}}

\item{extended}{\code{["min", "full"]}: Amount of variables to return.}

\item{drop.specials}{\code{logical(1) [TRUE]}: Special episodes (season 0) are dropped}

\item{drop.unaired}{\code{logical(1) [TRUE]}: Seasons without aired episodes are dropped.
Only works if \code{extended} is set to more than \code{min}.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
\code{trakt.seasons.summary} pulls season data.
Get details for a show's seasons, e.g. how many seasons there are, how many epsiodes
each season has, and season posters.
See \href{http://docs.trakt.apiary.io/#introduction/extended-info}{the API docs} for
possible values of \code{extended} to customize output amount.
}
\note{
See \href{http://docs.trakt.apiary.io/reference/seasons/summary}{the trakt API docs}
for further info
}
\examples{
\dontrun{
library(tRakt)
breakingbad.seasons <- trakt.seasons.summary("breaking-bad", extended = "min")
breakingbad.seasons.more <- trakt.seasons.summary("breaking-bad", extended = "full")
}
}
\seealso{
Other show data: \code{\link{trakt.get_all_episodes}},
  \code{\link{trakt.get_full_showdata}},
  \code{\link{trakt.seasons.season}},
  \code{\link{trakt.show.people}},
  \code{\link{trakt.show.ratings}},
  \code{\link{trakt.show.summary}},
  \code{\link{trakt.shows.related}},
  \code{\link{trakt.shows.trending}},
  \code{\link{trakt.stats}}
}
\concept{show data}
