% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-seasons.R
\name{trakt.seasons.summary}
\alias{trakt.seasons.summary}
\title{Get a show's season information}
\usage{
trakt.seasons.summary(target, extended = c("min", "full"),
  episodes = FALSE, drop.specials = TRUE, drop.unaired = TRUE)
}
\arguments{
\item{target}{\code{character(1)}: The ID of the item requested. Preferably the
\code{Trakt ID} (e.g. \code{1390}), \code{slug} (e.g. \code{"game-of-thrones"}), or \code{IMDb ID} (e.g. \code{"tt0944947"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and the result is combined.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}

\item{episodes}{\code{logical(1) [FALSE]}: If \code{TRUE}, all episodes for each season
are appended as a list-column, with the amount of variables depending on \code{extended}.}

\item{drop.specials}{\code{logical(1) [TRUE]}: Special episodes (season 0) are dropped}

\item{drop.unaired}{\code{logical(1) [TRUE]}: Seasons without aired episodes are dropped.
Only works if \code{extended} is \code{"full"}.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Get details for a show's seasons, e.g. how many seasons there are and  how many epsiodes
each season has. With \code{episodes == TRUE} and \code{extended == "full"}, this function
is also suitable to retrieve all episode ratings for all seasons of a show with
just a single API call.
}
\details{
This function wraps \href{https://trakt.docs.apiary.io/#reference/seasons/summary/get-all-seasons-for-a-show}{this API method}
with the endpoint \code{/shows/:show_id/seasons}.
}
\examples{
\dontrun{
trakt.seasons.summary("breaking-bad", extended = "min")
trakt.seasons.summary("utopia", extended = "full", episodes = TRUE)
}
}
\seealso{
Other show data: \code{\link{trakt.seasons.season}}
}
\concept{show data}
