% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.people.R
\name{trakt.show.people}
\alias{trakt.show.people}
\title{Get the cast and crew of a show}
\usage{
trakt.show.people(target, extended = "min")
}
\arguments{
\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"}), \code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \code{list}.
}
\description{
\code{trakt.show.people} pulls show people data.
}
\details{
Returns all cast and crew for a show, depending on how much data is available.
}
\note{
See \href{http://docs.trakt.apiary.io/#reference/shows/people/get-all-people-for-a-show}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
breakingbad.people <- trakt.show.people("breaking-bad")
}
}
\seealso{
Other show data: \code{\link{trakt.get_all_episodes}},
  \code{\link{trakt.get_full_showdata}},
  \code{\link{trakt.seasons.season}},
  \code{\link{trakt.seasons.summary}},
  \code{\link{trakt.show.summary}},
  \code{\link{trakt.stats}}

Other people data: \code{\link{trakt.movie.people}},
  \code{\link{trakt.people.movies}},
  \code{\link{trakt.people.shows}},
  \code{\link{trakt.people.summary}}
}
\concept{people data}
\concept{show data}
