% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trakt.show.stats}
\alias{trakt.show.stats}
\title{Get a show's summary info}
\usage{
trakt.show.stats(target, apikey = getOption("trakt.apikey"))
}
\arguments{
\item{target}{The \code{slug} or \code{tvdbid} of the show requested}

\item{apikey}{API-key used for the call. Defaults to \code{getOption("trakt.apikey")}}
}
\value{
A \code{list} containing show stats
}
\description{
\code{trakt.show.stats} pulls show stats and returns it compactly.
}
\details{
The data includes show ratings, scrobbles, checkins, plays, comments…
}
\note{
See \href{http://trakt.tv/api-docs/show-stats}{the trakt API docs for further info}
}
\examples{
\dontrun{
options(trakt.apikey = jsonlite::fromJSON("key.json")$apikey)
breakingbad.stats <- trakt.show.stats("breaking-bad")
}
}

