% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trakt.show.stats.R
\name{trakt.show.stats}
\alias{trakt.show.stats}
\title{[Defunct] Get a show's summary info}
\usage{
trakt.show.stats(target, extended = "min")
}
\arguments{
\item{target}{The \code{slug} of the show requested}

\item{extended}{Whether extended info should be provided.
Defaults to \code{"min"}, can either be \code{"min"} or \code{"full"}}
}
\value{
A \code{list} containing show stats
}
\description{
DEFUNCT as of 2015-02-11,
see \href{http://docs.trakt.apiary.io/reference/shows/stats/get-show-stats}{their API docs}
}
\details{
\code{trakt.show.stats} pulls show stats and returns it compactly.
The data includes show ratings, scrobbles, checkins, plays, comments…
}
\note{
See \href{http://docs.trakt.apiary.io/#reference/shows/stats}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
breakingbad.stats <- trakt.show.stats("breaking-bad")
}
}
\seealso{
Other show: \code{\link{trakt.getEpisodeData}};
  \code{\link{trakt.getFullShowData}};
  \code{\link{trakt.getSeasons}};
  \code{\link{trakt.show.people}};
  \code{\link{trakt.show.related}};
  \code{\link{trakt.show.season}};
  \code{\link{trakt.show.summary}};
  \code{\link{trakt.shows.popular}};
  \code{\link{trakt.shows.trending}}
}

