% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trakt.show.summary.R
\name{trakt.show.summary}
\alias{trakt.show.summary}
\title{Get show summary info}
\usage{
trakt.show.summary(target, apikey = getOption("trakt.apikey"),
  extended = NULL)
}
\arguments{
\item{target}{The \code{slug} or \code{tvdbid} of the show requested}

\item{apikey}{API-key used for the call.
Defaults to \code{getOption("trakt.apikey")}}

\item{extended}{Whether extended info should be provided.
Defaults to \code{NULL}, can be \code{"normal"} or \code{"full"}}
}
\value{
A \code{data.frame} containing summary info
}
\description{
\code{trakt.show.summary} pulls show summary data and returns it compactly.
}
\details{
Not to be confused with the show/summaries endpoint, this function
has the potential to return much more data.
}
\note{
See \href{http://trakt.tv/api-docs/show-summary}{the trakt API docs for further info}
Not to be confused with \href{http://trakt.tv/api-docs/show-summaries}{show/summaries}
}
\examples{
\dontrun{
options(trakt.apikey = jsonlite::fromJSON("key.json")$apikey)
breakingbad.summary_full <- trakt.show.summary("breaking-bad")
}
}

