% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trakt.show.summary.R
\name{trakt.show.summary}
\alias{trakt.show.summary}
\title{Get show summary info}
\usage{
trakt.show.summary(target, extended = "min")
}
\arguments{
\item{target}{The \code{id} of the show requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"}), \code{trakt id} or \code{IMDb id}}

\item{extended}{Whether extended info should be provided.
Defaults to \code{"min"}, can either be \code{"min"} or \code{"full"}}
}
\value{
A \code{list} containing summary info
}
\description{
\code{trakt.show.summary} pulls show summary data and returns it compactly.
}
\details{
Note that setting \code{extended} to \code{min} makes this function
return about as much informations as \link[tRakt]{trakt.search}
}
\note{
See \href{http://docs.trakt.apiary.io/reference/shows/summary}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
breakingbad.summary <- trakt.show.summary("breaking-bad")
}
}
\seealso{
Other show: \code{\link{trakt.getEpisodeData}};
  \code{\link{trakt.getFullShowData}};
  \code{\link{trakt.seasons.season}};
  \code{\link{trakt.seasons.summary}};
  \code{\link{trakt.show.people}};
  \code{\link{trakt.show.ratings}};
  \code{\link{trakt.show.stats}};
  \code{\link{trakt.shows.popular}};
  \code{\link{trakt.shows.related}};
  \code{\link{trakt.shows.trending}}
}

