% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.stats.R
\name{trakt.stats}
\alias{trakt.stats}
\title{Get a show or movie's stats}
\usage{
trakt.stats(target, type = c("shows", "movies"))
}
\arguments{
\item{target}{\code{character(1)}: The \code{id} of the item requested. Either the \code{slug}
(e.g. \code{"game-of-thrones"}), \code{Trakt ID} (e.g. \code{1390}) or \code{IMDB ID} (e.g. \code{"tt0944947"}).
Can also be of length greater than 1, in which case the function is called on all
\code{target} values separately and combined using \link[purrr:map_df]{purrr::map_df}.}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
\code{trakt.stats} pulls show stats and returns it compactly.
The data includes show ratings, scrobbles, checkins, plays, comments...
}
\note{
See \href{http://docs.trakt.apiary.io/#reference/shows/stats}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
breakingbad.stats <- trakt.stats(type = "shows", "breaking-bad")
}
}
\seealso{
Other show data: \code{\link{trakt.get_all_episodes}},
  \code{\link{trakt.get_full_showdata}},
  \code{\link{trakt.seasons.season}},
  \code{\link{trakt.seasons.summary}},
  \code{\link{trakt.show.summary}}

Other movie data: \code{\link{trakt.movie.releases}},
  \code{\link{trakt.movie.summary}}
}
\concept{movie data}
\concept{show data}
