% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.user.collection.R
\name{trakt.user.collection}
\alias{trakt.user.collection}
\title{Get a user's collected shows or movies}
\usage{
trakt.user.collection(user = getOption("trakt.username"),
  type = c("shows", "movies"), unnest_episodes = FALSE)
}
\arguments{
\item{user}{Target user. Defaults to \code{getOption("trakt.username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and combined using \link[purrr:map_df]{purrr::map_df}.}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}

\item{unnest_episodes}{\code{logical(1) [FALSE]}: Unnests episode data using
\code{[tidyr](tidyr::unnest)} and returns one row per episode rather than one row per show.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
\code{trakt.user.collection} retrieves a user's collected shows or movies.
It does not use OAuth2, so you can only get data for a user with a
public profile.
}
\examples{
\dontrun{
seans.movies <- trakt.user.collection(user = "sean", type = "movies")
}
}
\seealso{
Other user data: \code{\link{trakt.user.network}},
  \code{\link{trakt.user.ratings}},
  \code{\link{trakt.user.stats}},
  \code{\link{trakt.user.watched}},
  \code{\link{trakt.user.watchlist}}
}
\concept{user data}
