% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-user-collection.R
\name{trakt.user.collection}
\alias{trakt.user.collection}
\title{Get a user's collected shows or movies}
\usage{
trakt.user.collection(user = getOption("trakt_username"),
  type = c("shows", "movies"), unnest_episodes = FALSE)
}
\arguments{
\item{user}{Target username. Defaults to \code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}. For season/episode-specific
functions, values \code{seasons} or \code{episodes} are also allowed.}

\item{unnest_episodes}{\code{logical(1) [FALSE]}: Unnests episode data using
\code{\link[tidyr:unnest]{tidyr::unnest()}} and returns one row per episode rather than one row per show.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
\code{trakt.user.collection} retrieves a user's collected shows or movies.
It does not use OAuth2, so you can only get data for a user with a
public profile.
}
\details{
This function wraps \href{https://trakt.docs.apiary.io/#reference/users/collection/get-collection}{this API method}
at the endpoint \code{/users/:user_id/collection/:type}.
}
\note{
The \code{extended = "metadata"} API parameter is not implemented. This would
add media information \code{media_type}, \code{resolution}, \code{audio}, \code{audio_channels} and \code{3D}
to the output, which may or may not be available. If this feature is important to
you, please open an issue on GitHub.
}
\examples{
\dontrun{
trakt.user.collection(user = "sean", type = "movies")
}
}
\seealso{
Other user data: \code{\link{trakt.user.history}},
  \code{\link{trakt.user.network}},
  \code{\link{trakt.user.ratings}},
  \code{\link{trakt.user.stats}},
  \code{\link{trakt.user.watched}},
  \code{\link{trakt.user.watchlist}}
}
\concept{user data}
