% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-user-history.R
\name{trakt.user.history}
\alias{trakt.user.history}
\title{Get a user's watch history}
\usage{
trakt.user.history(user = getOption("trakt.username"),
  type = c("shows", "movies"), limit = 10L, start_at = NULL,
  end_at = NULL, extended = c("min", "full"))
}
\arguments{
\item{user}{Target username. Defaults to \code{getOption("trakt.username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}. For season/episode-specific
functions, values \code{seasons} or \code{episodes} are also allowed.}

\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced to \code{integer}.}

\item{start_at, end_at}{\code{character(1)}: A time-window to filter by. Must be coercible
to a datetime object of class \code{POSIXct}. See \link{ISOdate} for further information.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble}. For \code{type == "show"}, the \code{tibble} contains
two list-columns \code{show} and \code{episode} which are \emph{not} unnested by default due to
duplicate variable names, the preferref handling of which is left to the user.
}
\description{
Retrieve a the last \code{limit} items a user has watched, including the method by
which it was logged (e.g. \emph{scrobble} or \emph{checkin}).
}
\examples{
\dontrun{
# The last 5 movies that user "jemus42" has watched
trakt.user.history(user = "jemus42", type = "movies", limit = 5)
}
}
\seealso{
Other user data: \code{\link{trakt.user.collection}},
  \code{\link{trakt.user.network}},
  \code{\link{trakt.user.ratings}},
  \code{\link{trakt.user.stats}},
  \code{\link{trakt.user.watched}},
  \code{\link{trakt.user.watchlist}}
}
\concept{user data}
