% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-user-network.R
\name{trakt.user.network}
\alias{trakt.user.network}
\title{Get a user's social connections}
\usage{
trakt.user.network(relationship = c("friends", "followers", "following"),
  user = getOption("trakt.username"), extended = c("min", "full"))
}
\arguments{
\item{relationship}{\code{character(1) ["friends"]}: Type of user relationship. Either
\code{"friends"}, \code{"followers"}, or \code{"following"}.}

\item{user}{Target username. Defaults to \code{getOption("trakt.username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Retrieve a user's followers, followings or friends (the two-way relationship).
}
\note{
Since no OAuth2 methods are supported yet, the specified user must not be private.
}
\examples{
\dontrun{
trakt.user.network("friends", "jemus42")
trakt.user.network("following", "jemus42")
}
}
\seealso{
Other user data: \code{\link{trakt.user.collection}},
  \code{\link{trakt.user.history}},
  \code{\link{trakt.user.ratings}},
  \code{\link{trakt.user.stats}},
  \code{\link{trakt.user.watched}},
  \code{\link{trakt.user.watchlist}}
}
\concept{user data}
