% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.user.ratings.R
\name{trakt.user.ratings}
\alias{trakt.user.ratings}
\title{Get a user's ratings}
\usage{
trakt.user.ratings(user = getOption("trakt.username"),
  type = c("movies", "seasons", "shows", "episodes"), rating = NULL)
}
\arguments{
\item{user}{Target user. Defaults to \code{getOption("trakt.username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and combined using \link[purrr:map_df]{purrr::map_df}.}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}

\item{rating}{\code{integer(1L)}: Rating to filter by. Can be \code{1} through \code{10}, default is \code{NULL}}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
\code{trakt.user.ratings} retrieves a user's media ratings
}
\examples{
\dontrun{
trakt.user.ratings(user = "jemus42")
trakt.user.ratings(user = "jemus42", type = "movies")
}
}
\seealso{
Other user data: \code{\link{trakt.user.collection}},
  \code{\link{trakt.user.history}},
  \code{\link{trakt.user.network}},
  \code{\link{trakt.user.stats}},
  \code{\link{trakt.user.watched}},
  \code{\link{trakt.user.watchlist}}
}
\concept{user data}
