% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt-user-stats.R
\name{trakt.user.stats}
\alias{trakt.user.stats}
\title{Get a user's stats}
\usage{
trakt.user.stats(user = getOption("trakt.username"))
}
\arguments{
\item{user}{Target user. Defaults to \code{getOption("trakt.username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and combined using \link[purrr:map_df]{purrr::map_df}.}
}
\value{
A list of \link[tibble:tibble-package]{tibbles}.
}
\description{
Data about a user's interactions with movies, shows, seasons, episodes,
as well as their social network (friends, followings, followers) and a
frequency table of the user's media ratings so far.
}
\examples{
\dontrun{
mystats <- trakt.user.stats() # Defaults to your username if set
seanstats <- trakt.user.stats(user = "sean")
}
}
\seealso{
Other user data: \code{\link{trakt.user.collection}},
  \code{\link{trakt.user.history}},
  \code{\link{trakt.user.network}},
  \code{\link{trakt.user.ratings}},
  \code{\link{trakt.user.watched}},
  \code{\link{trakt.user.watchlist}}
}
\concept{user data}
