% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.user.watched.R
\name{trakt.user.watched}
\alias{trakt.user.watched}
\title{Get a user's watched shows or movies}
\usage{
trakt.user.watched(user = getOption("trakt.username"),
  type = c("shows", "shows.extended", "movies"))
}
\arguments{
\item{user}{Target user. Defaults to \code{getOption("trakt.username")}}

\item{type}{Either \code{shows} (default), \code{shows.extended} or \code{movies}}
}
\value{
A \code{[tibble](tibble::tibble-package)}.
}
\description{
\code{trakt.user.watched} retrieves a user's watched shows or movies.
It does not use OAuth2, so you can only get data for a user with a
\strong{public profile}.
}
\details{
If \code{type} is set to \code{shows.extended}, the resulting \link[tibble:tibble-package]{tibble}
contains play stats for \emph{every} watched episode of \emph{every} show. Otherwise,
the returned \link[tibble:tibble-package]{tibble} only contains play stats per show or movie respectively.
}
\note{
See \href{http://docs.trakt.apiary.io/reference/users/watched/get-watched}{the trakt API docs for further info}
}
\examples{
\dontrun{
myshows <- trakt.user.watched() # Defaults to your username if set
seans.shows <- trakt.user.watched(user = "sean")
}
}
\seealso{
Other user data: \code{\link{trakt.user.collection}},
  \code{\link{trakt.user.friends}},
  \code{\link{trakt.user.ratings}},
  \code{\link{trakt.user.stats}},
  \code{\link{trakt.user.watchlist}}
}
\concept{user data}
