% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.user.watched.R
\name{trakt.user.watched}
\alias{trakt.user.watched}
\title{Get a user's watched shows or movies}
\usage{
trakt.user.watched(user = getOption("trakt.username"),
  type = c("shows", "movies"), extended = c("min", "full"),
  noseasons = TRUE)
}
\arguments{
\item{user}{Target user. Defaults to \code{getOption("trakt.username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and combined using \link[purrr:map_df]{purrr::map_df}.}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}

\item{noseasons}{\code{logical(1) [TRUE]}: Only for \code{type = "show"}: Exclude detailed season
data from output. This is advisable if you do not need per-episode data and want to
be nice to the API.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
\code{trakt.user.watched} retrieves a user's watched shows or movies.
It does not use OAuth2, so you can only get data for a user with a
\strong{public profile}.
}
\details{
If \code{type} is set to \code{shows.extended}, the resulting \link[tibble:tibble-package]{tibble}
contains play stats for \emph{every} watched episode of \emph{every} show. Otherwise,
the returned \link[tibble:tibble-package]{tibble} only contains play stats per show or movie respectively.
}
\examples{
\dontrun{
myshows <- trakt.user.watched() # Defaults to your username if set

# Use noseasons = TRUE to avoid receiving detailed season/episode data
seans.shows <- trakt.user.watched(user = "sean", noseasons = TRUE)
}
}
\seealso{
Other user data: \code{\link{trakt.user.collection}},
  \code{\link{trakt.user.history}},
  \code{\link{trakt.user.network}},
  \code{\link{trakt.user.ratings}},
  \code{\link{trakt.user.stats}},
  \code{\link{trakt.user.watchlist}}
}
\concept{user data}
