% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trakt.user.watchlist.R
\name{trakt.user.watchlist}
\alias{trakt.user.watchlist}
\title{Get a user's watchlist}
\usage{
trakt.user.watchlist(user = getOption("trakt.username"), type = "shows")
}
\arguments{
\item{user}{Target user. Defaults to \code{getOption("trakt.username")}}

\item{type}{Either \code{shows} (default), \code{seasons} or \code{movies}}
}
\value{
A \code{data.frame} containing stats.
}
\description{
\code{trakt.user.watchlist} pulls a user's watchlist.
}
\note{
See \href{http://docs.trakt.apiary.io/#reference/users/ratings/get-watchlist}{the trakt API docs for further info}
}
\examples{
\dontrun{
get_trakt_credentials() # Set required API data/headers
mystats   <- trakt.user.watchlist() # Defaults to your username if set
seanstats <- trakt.user.watchlist(user = "sean")
}
}

