% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.user.watchlist.R
\name{trakt.user.watchlist}
\alias{trakt.user.watchlist}
\title{Get a user's watchlist}
\usage{
trakt.user.watchlist(user = getOption("trakt.username"),
  type = c("movies", "shows"), extended = "min")
}
\arguments{
\item{user}{Target user. Defaults to \code{getOption("trakt.username")}}

\item{type}{Either \code{shows} (default) or \code{movies}}

\item{extended}{Whether extended info should be provided.
Defaults to \code{"min"}, can either be \code{"min"} or \code{"full"}}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
\code{trakt.user.watchlist} pulls a user's watchlist.
Either all shows or movies currently watchlisted will be returned.
}
\note{
See \href{http://docs.trakt.apiary.io/#reference/users/ratings/get-watchlist}{the trakt API docs for further info}
}
\examples{
\dontrun{
library(tRakt)
movie_watchlist <- trakt.user.watchlist(type = "movies") # Defaults to your username if set
sean_movies <- trakt.user.watchlist(user = "sean") # Defaults to movie watchlist
}
}
\seealso{
Other user data: \code{\link{trakt.user.collection}},
  \code{\link{trakt.user.followers}},
  \code{\link{trakt.user.following}},
  \code{\link{trakt.user.friends}},
  \code{\link{trakt.user.ratings}},
  \code{\link{trakt.user.stats}},
  \code{\link{trakt.user.watched}}
}
\concept{user data}
