% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trakt.user.watchlist.R
\name{trakt.user.watchlist}
\alias{trakt.user.watchlist}
\title{Get a user's watchlist}
\usage{
trakt.user.watchlist(user = getOption("trakt.username"),
  type = c("movies", "shows"), extended = c("min", "full"))
}
\arguments{
\item{user}{Target user. Defaults to \code{getOption("trakt.username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and combined using \link[purrr:map_df]{purrr::map_df}.}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
\code{trakt.user.watchlist} pulls a user's watchlist.
Either all shows or movies currently watchlisted will be returned.
}
\examples{
\dontrun{
library(tRakt)
movie_watchlist <- trakt.user.watchlist(type = "movies") # Defaults to your username if set
sean_movies <- trakt.user.watchlist(user = "sean") # Defaults to movie watchlist
}
}
\seealso{
Other user data: \code{\link{trakt.user.collection}},
  \code{\link{trakt.user.history}},
  \code{\link{trakt.user.network}},
  \code{\link{trakt.user.ratings}},
  \code{\link{trakt.user.stats}},
  \code{\link{trakt.user.watched}}
}
\concept{user data}
