% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs-common.R
\name{trakt_api_common_parameters}
\alias{trakt_api_common_parameters}
\title{Common API parameters}
\arguments{
\item{id}{\code{character(1)}: The ID of the item requested. Preferably the
\code{trakt} ID (e.g. \code{1429}). Other options are the trakt.tv \code{slug} (e.g. \code{"the-wire"})
or \code{imdb} ID (e.g. \code{"tt0306414"}).
Can also be of length greater than 1, in which case the function is called on all
\code{id} values separately and the result is combined. See
\code{vignette("finding-things")} for more details.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things")} for more details.}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}. For season/episode-specific
functions, values \code{seasons} or \code{episodes} are also allowed.}

\item{user}{\code{character(1)}: Target username (or \code{slug}). Defaults to
\code{"me"}, the OAuth user.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{period}{\code{character(1) ["weekly"]}: Which period to filter by. Possible values
are \code{"weekly"}, \code{"monthly"}, \code{"yearly"}, \code{"all"}.}

\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced via \code{as.integer()}.}

\item{season, episode}{\code{integer(1) [1L]}: The season and episode number. If longer,
e.g. \code{1:5}, the function is vectorized and the output will be
combined. This may result in \emph{a lot} of API calls. Use wisely.}

\item{start_date}{\code{character(1)}: A date in the past from which
on to count updates. If no date is supplied, the default is to
use yesterday relative to the current date. Value must either
be standard \code{YYYY-MM-DD} format or an object of class \link[base:Dates]{Date},
which will then be coerced via \link[base:character]{as.character()}.}

\item{sort}{\code{character(1) ["newest"]}: Comment sort order, one of
"newest", "oldest", "likes" or "replies".}

\item{query}{\code{character(1)}: Search string for titles and descriptions.
For \code{search_query()} other fields are searched depending on the \code{type} of media.
See \href{https://trakt.docs.apiary.io/#reference/search/text-query}{the API docs} for a
full reference.}

\item{years}{\code{character | integer}: 4-digit year (\code{2010}) \strong{or} range,
e.g. \code{"2010-2020"}. Can also be an integer vector of length two which will be
coerced appropriately, e.g. \code{c(2010, 2020)}.}

\item{genres}{\code{character(n)}: Genre  slug(s).
See \code{\link{trakt_genres}} for a table of genres.
Multiple values are allowed and will be concatenated.}

\item{languages}{\code{character(n)}: Two-letter language code(s).
Also see \code{\link{trakt_languages}} for available languages (code and name).}

\item{countries}{\code{character(n)}: Two-letter country code(s).
See \code{\link{trakt_countries}}.}

\item{runtimes}{\code{character | integer}: Integer range in minutes, e.g. \code{30-90}.
Can also be an integer vector of length two which will be coerced appropriately.}

\item{ratings}{\code{character | integer}:  Integer range between \code{0} and \code{100}.
Can also be an integer vector of length two which will be coerced appropriately.
Note that user-supplied ratings are in the range of 1 to 10, yet the ratings
on the site itself are scaled to the range of 1 to 100.}

\item{certifications}{\code{character(n)}: Certification(s) like \code{pg-13}.
Multiple values are allowed. Use \code{\link{trakt_certifications}} for reference.
Note that there are different certifications for shows and movies.}

\item{networks}{\code{character(n)}: (Shows only) Network name like \code{HBO}.
See \code{\link{trakt_networks}} for a list of known networks.}

\item{status}{\code{character(n)}: (Shows only) The status of the shows.
One of \code{"returning series"}, \code{"in production"}, \code{"planned"},
\code{"canceled"}, or \code{"ended"}.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
These parameters are used extensively throughout this package as they are required
for many API methods. Here there are all documented in one place, which is
intended to make the individual function documentation more consistent.
}
\keyword{internal}
