% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{trakt_get}
\alias{trakt_get}
\title{Make an API call to any URL}
\usage{
trakt_get(url, client.id = getOption("trakt_client_id"), HEAD = FALSE)
}
\arguments{
\item{url}{\code{character(1)}: The API endpoint. Either a full URL like
\code{"https://api.trakt.tv/shows/breaking-bad"} or just the endpoint like \code{shows/breaking-bad}.}

\item{client.id}{\code{character(1)}: API client id. see \link{trakt_credentials} for further information.}

\item{HEAD}{\code{logical(1) [FALSE]}: If \code{TRUE}, only a HTTP \code{HEAD} request is performed
and its content returned. This is useful if you are only interested in status codes
or other headers, and don't want to waste resources on additional bandwidth.}
}
\value{
The \link[jsonlite:fromJSON]{parsed} content of the API response.
An empty \link[tibble:tibble-package]{tibble()} if the response is an empty \code{JSON} array.
}
\description{
\code{trakt_get} makes an API call to a specified URL and returns the parsed output.
}
\examples{
# A simple request to a direct URL
trakt_get("https://api.trakt.tv/shows/breaking-bad")

# A HEAD-only request, useful for validating a URL exists or the API is accessable
trakt_get("https://api.trakt.tv/users/jemus42", HEAD = TRUE)

# Optionally be lazy about URL specification by dropping the hostname:
trakt_get("shows/game-of-thrones")
}
\seealso{
Other API-basics: \code{\link{trakt.search}},
  \code{\link{trakt_credentials}}
}
\concept{API-basics}
