% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic-lists.R
\name{updated_media}
\alias{updated_media}
\alias{movies_updates}
\alias{shows_updates}
\title{Recently updated media}
\source{
movies_updates() wraps endpoint \href{https://trakt.docs.apiary.io/#reference/movies/updates/get-recently-updated-movies}{movies/updates/:start_date}

shows_updates() wraps endpoint \href{https://trakt.docs.apiary.io/#reference/shows/updates/get-recently-updated-shows}{shows/updates/:start_date}
}
\usage{
movies_updates(limit = 10, extended = c("min", "full"),
  start_date = Sys.Date() - 1)

shows_updates(limit = 10, extended = c("min", "full"),
  start_date = Sys.Date() - 1)
}
\arguments{
\item{limit}{\code{integer(1) [10L]}: Number of items to return. Must be greater
than \code{0} and will be coerced via \code{as.integer()}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}

\item{start_date}{\code{character(1)}: A date in the past from which
on to count updates. If no date is supplied, the default is to
use yesterday relative to the current date. Value must either
be standard \code{YYYY-MM-DD} format or an object of class \link[base:Dates]{Date},
which will then be coerced via \link[base:as.character]{as.character()}.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
These functions return recently updated movies/shows on trakt.tv.
}
\note{
\code{shows_updates()} and \code{movies_updates()} do not support filters.
}
\section{The Dynamic Lists on trakt.tv}{


These functions access the automatically updated lists provided by trakt.tv.
Each function comes in two flavors: Shows or movies. The following descriptions
are adapted directly from the
\href{https://trakt.docs.apiary.io/#reference/movies/popular/get-popular-movies}{API reference}.
\itemize{
\item \link[=popular_media]{Popular}: Popularity is calculated using the rating percentage and the number of
ratings.
\item \link[=trending_media]{Trending}: Returns all movies/shows being watched right now.
Movies/shows with the most users are returned first.
\item \link[=played_media]{Played}: Returns the most played (a single user can watch multiple times)
movies/shows in the specified time \code{period}.
\item \link[=watched_media]{Watched}: Returns the most watched (unique users) movies/shows in the specified
time \code{period}.
\item \link[=collected_media]{Collected}: Returns the most collected (unique users) movies/shows in the
specified time \code{period}.
\item \link[=anticipated_media]{Anticipated}: Returns the most anticipated movies/shows based on the number of
lists a movie/show appears on.
\item \link[=updated_media]{Updates}: Returns all movies/shows updated since the specified UTC \code{start_date}.
In this case, the upper bound for \code{limit} is 100.
}

The functions for \strong{Played}, \strong{Watched}, \strong{Collected} and \strong{Played} each return
the same additional variables besides the media information: \code{watcher_count},
\code{play_count}, \code{collected_count}, \code{collector_count}.
}

\seealso{
Other movie data: \code{\link{anticipated_media}},
  \code{\link{collected_media}},
  \code{\link{media_aliases}},
  \code{\link{media_comments}}, \code{\link{media_lists}},
  \code{\link{media_people}}, \code{\link{media_ratings}},
  \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{media_watching}},
  \code{\link{movies_boxoffice}},
  \code{\link{movies_related}},
  \code{\link{movies_releases}},
  \code{\link{movies_summary}}, \code{\link{people_media}},
  \code{\link{played_media}}, \code{\link{popular_media}},
  \code{\link{trending_media}}, \code{\link{watched_media}}

Other show data: \code{\link{collected_media}},
  \code{\link{media_aliases}},
  \code{\link{media_comments}}, \code{\link{media_lists}},
  \code{\link{media_people}}, \code{\link{media_ratings}},
  \code{\link{media_stats}},
  \code{\link{media_translations}},
  \code{\link{media_watching}}, \code{\link{people_media}},
  \code{\link{played_media}},
  \code{\link{shows_next_episode}},
  \code{\link{shows_related}}, \code{\link{shows_summary}}
}
\concept{movie data}
\concept{show data}
