% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_collection.R
\name{user_collection}
\alias{user_collection}
\title{Get a user's collected shows or movies}
\source{
\code{user_collection()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/users/collection/get-collection}{users/:id/collection/:type}.
}
\usage{
user_collection(user = getOption("trakt_username"), type = c("shows",
  "movies"), unnest_episodes = FALSE, extended = c("min", "full"))
}
\arguments{
\item{user}{\code{character(1)}: Target username (or \code{slug}). Defaults to
\code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}. For season/episode-specific
functions, values \code{seasons} or \code{episodes} are also allowed.}

\item{unnest_episodes}{\code{logical(1) [FALSE]}: Unnests episode data using
\code{\link[tidyr:unnest]{tidyr::unnest()}} and returns one row per episode rather than one row per show.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get a user's collected shows or movies
}
\details{
This function wraps the API method
\href{https://trakt.docs.apiary.io/#reference/users/collection/get-collection}{/users/:user_id/collection/:type}.
}
\note{
The \code{extended = "metadata"} API parameter is not implemented. This would
add media information \code{media_type}, \code{resolution}, \code{audio}, \code{audio_channels} and \code{3D}
to the output, which may or may not be available. If this feature is important to
you, please open an issue on GitHub.
}
\examples{
\dontrun{
user_collection(user = "sean", type = "movies")
user_collection(user = "sean", type = "shows")
}
}
\seealso{
Other user data: \code{\link{user_comments}},
  \code{\link{user_history}}, \code{\link{user_likes}},
  \code{\link{user_network}}, \code{\link{user_profile}},
  \code{\link{user_ratings}}, \code{\link{user_stats}},
  \code{\link{user_watched}}, \code{\link{user_watchlist}}
}
\concept{user data}
