% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_comments.R
\name{user_comments}
\alias{user_comments}
\title{Get a user's comments}
\source{
user_comments() wraps endpoint \href{https://trakt.docs.apiary.io/#reference/users/comments/get-comments}{users/:id/comments/:comment_type/:type?include_replies=}
}
\usage{
user_comments(user = getOption("trakt_username"),
  comment_type = c("all", "reviews", "shouts"), type = c("all",
  "movies", "shows", "seasons", "episodes", "lists"),
  include_replies = FALSE)
}
\arguments{
\item{user}{\code{character(1)}: Target username (or \code{slug}). Defaults to
\code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{comment_type}{\code{character(1) ["all"]}: The type of comment, one of "all",
"reviews" or "shouts".}

\item{type}{\code{character(1) ["all"]}: The type of media to filter by, one of
"all", "movies", "shows", "seasons", "episodes" or "lists".}

\item{include_replies}{\code{logical(1) [FALSE]}: Whether to include replies.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get a user's comments
}
\examples{
\dontrun{
user_comments("jemus42")
}
}
\seealso{
Other user data: \code{\link{user_collection}},
  \code{\link{user_history}}, \code{\link{user_likes}},
  \code{\link{user_network}}, \code{\link{user_profile}},
  \code{\link{user_ratings}}, \code{\link{user_stats}},
  \code{\link{user_watched}}, \code{\link{user_watchlist}}

Other comment methods: \code{\link{comments_comment}},
  \code{\link{comments_trending}},
  \code{\link{comments_updates}},
  \code{\link{media_comments}},
  \code{\link{user_list_comments}}
}
\concept{comment methods}
\concept{user data}
