% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_likes.R
\name{user_likes}
\alias{user_likes}
\title{Get items (comments, lists) a user likes}
\source{
\href{https://trakt.docs.apiary.io/#reference/users/likes/get-likes}{users/likes/:type}
}
\usage{
user_likes(type = c("comments", "lists"))
}
\arguments{
\item{type}{One of "comments", "lists".}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
This requires authentication.
}
\examples{
# Get liked lists
if (FALSE) {
  user_likes("lists")
}
}
\seealso{
Other user data: \code{\link{user_collection}},
  \code{\link{user_comments}}, \code{\link{user_history}},
  \code{\link{user_network}}, \code{\link{user_profile}},
  \code{\link{user_ratings}}, \code{\link{user_stats}},
  \code{\link{user_watched}}, \code{\link{user_watchlist}}
}
\concept{user data}
