% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_lists.R
\name{user_list}
\alias{user_list}
\title{Get a single list}
\usage{
user_list(user = getOption("trakt_username"), list_id,
  extended = c("min", "full"))
}
\arguments{
\item{user}{\code{character(1)}: Target username. Defaults to \code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{list_id}{The list identifier, either \code{trakt} ID or \code{slug} of the list.
Can be optained via the website (URL \code{slug}) or e.g. \link{user_lists}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Retrieve a single list a user has created together with information about the user.
Use \code{extended = "full"} to retrieve all user profile data, similiar to \link{user_profile}.
The returned variables \code{trakt} (list ID) or \code{slug} (list slug) can be used to
retrieve the list's items via \link{user_list_items}.
}
\note{
In the embedded user data, \code{name} is renamed to \code{user_name} due
to duplication with e.g. list names,
and \code{slug} is renamed to \code{user_slug} for the same reason.
}
\examples{
\dontrun{
user_list("jemus42", list_id = 2121308)
}
}
\seealso{
\link{user_list} for only a single list.

\link{user_list_items} For the actual content of a list.

Other list methods: \code{\link{lists_popular}},
  \code{\link{media_lists}},
  \code{\link{user_list_comments}},
  \code{\link{user_list_items}}, \code{\link{user_lists}}
}
\concept{list methods}
