% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_list_comments.R
\name{user_list_comments}
\alias{user_list_comments}
\title{Get comments on a user-created list}
\usage{
user_list_comments(user = getOption("trakt_username"), list_id,
  sort = c("newest", "oldest", "likes", "replies"), extended = c("min",
  "full"))
}
\arguments{
\item{user}{\code{character(1)}: Target username. Defaults to \code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{list_id}{The list identifier, either \code{trakt} ID or \code{slug} of the list.
Can be optained via the website (URL \code{slug}) or e.g. \link{user_lists}.}

\item{sort}{\code{character(1) ["newest"]}: Comment sort order, one of
"newest", "oldest", "likes" or "replies".}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See section \link[=trakt_api_common_parameters]{"Extended Information"} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Get comments on a user-created list
}
\examples{
\dontrun{
user_list_comments("donxy", "1248149")
}
}
\seealso{
Other comment methods: \code{\link{media_comments}}

Other list methods: \code{\link{lists_popular}},
  \code{\link{media_lists}}, \code{\link{user_list_items}},
  \code{\link{user_lists}}, \code{\link{user_list}}
}
\concept{comment methods}
\concept{list methods}
