% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_list_comments.R
\name{user_list_comments}
\alias{user_list_comments}
\title{Get comments on a user-created list}
\source{
user_profile() wraps endpoint \href{https://trakt.docs.apiary.io/#reference/users/list-comments/get-all-list-comments}{users/:id/lists/:list_id/comments/:sort}
}
\usage{
user_list_comments(user = getOption("trakt_username"), list_id,
  sort = c("newest", "oldest", "likes", "replies"), extended = c("min",
  "full"))
}
\arguments{
\item{user}{\code{character(1)}: Target username (or \code{slug}). Defaults to
\code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{list_id}{The list identifier, either \code{trakt} ID or \code{slug} of the list.
Can be optained via the website (URL \code{slug}) or e.g. \link{user_lists}.}

\item{sort}{\code{character(1) ["newest"]}: Comment sort order, one of
"newest", "oldest", "likes" or "replies".}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get comments on a user-created list
}
\examples{
\dontrun{
user_list_comments("donxy", "1248149")
}
}
\seealso{
Other comment methods: \code{\link{comments_comment}},
  \code{\link{comments_trending}},
  \code{\link{comments_updates}},
  \code{\link{media_comments}}, \code{\link{user_comments}}

Other list methods: \code{\link{lists_popular}},
  \code{\link{media_lists}}, \code{\link{user_list_items}},
  \code{\link{user_lists}}, \code{\link{user_list}}
}
\concept{comment methods}
\concept{list methods}
