% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_list_items.R
\name{user_list_items}
\alias{user_list_items}
\title{Get a user's list's items}
\source{
\code{user_list_items()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/users/list-items/get-items-on-a-custom-list}{users/:id/lists/:list_id/items/:type}.
}
\usage{
user_list_items(user = getOption("trakt_username"), list_id,
  type = NULL, extended = c("min", "full"))
}
\arguments{
\item{user}{\code{character(1)}: Target username (or \code{slug}). Defaults to
\code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{list_id}{The list identifier, either \code{trakt} ID or \code{slug} of the list.
Can be optained via the website (URL \code{slug}) or e.g. \link{user_lists}.}

\item{type}{\code{character(1) [NULL]}: If not \code{NULL}, only items of that media type are
returned. Possible values are \code{"movie"}, \code{"show"}, \code{"season"}, \code{"episode"}, \code{"person"}.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get a user's list's items
}
\examples{
\dontrun{
# A large list with various media types
# All items
user_list_items("sp1ti", list_id = "5615781", extended = "min")

# Movies only
user_list_items("sp1ti", list_id = "5615781", extended = "min", type = "movie")

# Shows...
user_list_items("sp1ti", list_id = "5615781", extended = "min", type = "shows")

# Only seasons
user_list_items("sp1ti", list_id = "5615781", extended = "min", type = "season")

# Only episodes
user_list_items("sp1ti", list_id = "5615781", extended = "min", type = "episodes")
}
}
\seealso{
Other list methods: \code{\link{lists_popular}},
  \code{\link{media_lists}},
  \code{\link{user_list_comments}},
  \code{\link{user_lists}}, \code{\link{user_list}}
}
\concept{list methods}
