% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_lists.R
\name{user_lists}
\alias{user_lists}
\title{Get a user's lists}
\usage{
user_lists(user = getOption("trakt_username"), extended = c("min",
  "full"))
}
\arguments{
\item{user}{\code{character(1)}: Target username. Defaults to \code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See section \link[=trakt_api_common_parameters]{"Extended Information"} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Retrieve all lists a user has created together with information about the user.
Use \code{extended = "full"} to retrieve all user profile data, similiar to \link{user_profile}.
The returned variables \code{trakt} (list ID) or \code{slug} (list slug) can be used to
retrieve the list's items via \link{user_list_items}.
}
\note{
In the embedded user data, \code{name} is renamed to \code{user_name} due
to duplication with e.g. list names,
and \code{slug} is renamed to \code{user_slug} for the same reason.
}
\examples{
\dontrun{
user_lists("jemus42")
}
}
\seealso{
\link{user_lists} for all lists a user has.

\link{user_list_items} For the actual content of a list.

Other list methods: \code{\link{user_list_items}},
  \code{\link{user_list}}
}
\concept{list methods}
