% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_network.R
\name{user_network}
\alias{user_network}
\alias{user_followers}
\alias{user_following}
\alias{user_friends}
\title{Get a user's social connections}
\source{
\code{user_followers()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/users/followers/get-followers}{/users/:id/followers}.

\code{user_following()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/users/following/get-following}{/users/:id/following}.

\code{user_friends()} wraps endpoint \href{https://trakt.docs.apiary.io/#reference/users/friends/get-friends}{/users/:id/friends}.
}
\usage{
user_followers(user = getOption("trakt_user"), extended = "min")

user_following(user = getOption("trakt_user"), extended = "min")

user_friends(user = getOption("trakt_user"), extended = "min")
}
\arguments{
\item{user}{\code{character(1)}: Target username (or \code{slug}). Defaults to
\code{"me"}, the OAuth user.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get followers, followings or friends (the two-way relationship).
}
\note{
If the specified user is private, you need to be able to make an \link[=trakt_credentials]{authenticated request} and be friends with the user.
}
\examples{
\dontrun{
user_followers(user = "sean")
}
\dontrun{
user_following(user = "sean")
}
\dontrun{
user_friends(user = "sean")
}
}
\seealso{
Other user data: 
\code{\link{user_collection}()},
\code{\link{user_comments}()},
\code{\link{user_history}()},
\code{\link{user_likes}()},
\code{\link{user_profile}()},
\code{\link{user_ratings}()},
\code{\link{user_stats}()},
\code{\link{user_watched}()},
\code{\link{user_watchlist}()}
}
\concept{user data}
