% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_ratings.R
\name{user_ratings}
\alias{user_ratings}
\title{Get a user's ratings}
\usage{
user_ratings(user = getOption("trakt_username"), type = c("movies",
  "seasons", "shows", "episodes"), rating = NULL, extended = c("min",
  "full"))
}
\arguments{
\item{user}{\code{character(1)}: Target username. Defaults to \code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}. For season/episode-specific
functions, values \code{seasons} or \code{episodes} are also allowed.}

\item{rating}{\code{integer(1L)}: Optional rating to filter by. Can be \code{1} through \code{10},
default is \code{NULL} for no filtering.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.}
}
\value{
A \link[tibble:tibble-package]{tibble()}.
}
\description{
Retrieve a user's media ratings
}
\details{
This function wraps the API method
\href{https://trakt.docs.apiary.io/#reference/users/ratings/get-ratings}{/users/:id/ratings/:type/:rating}.
}
\examples{
\dontrun{
user_ratings(user = "jemus42")
user_ratings(user = "sean", type = "movies")
}
}
\seealso{
Other user data: \code{\link{user_collection}},
  \code{\link{user_history}}, \code{\link{user_network}},
  \code{\link{user_stats}}, \code{\link{user_summary}},
  \code{\link{user_watched}}, \code{\link{user_watchlist}}
}
\concept{user data}
