% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_watchlist.R
\name{user_watchlist}
\alias{user_watchlist}
\title{Get a user's watchlist}
\source{
user_watchlist() wraps endpoint \href{https://trakt.docs.apiary.io/#reference/users/watchlist/get-watchlist}{users/:id/watchlist/:type}
}
\usage{
user_watchlist(user = getOption("trakt_username"), type = c("movies",
  "shows"), extended = c("min", "full"))
}
\arguments{
\item{user}{\code{character(1)}: Target username (or \code{slug}). Defaults to
\code{getOption("trakt_username")}.
Can also be of length greater than 1, in which case the function is called on all
\code{user} values separately and the result is combined.}

\item{type}{\code{character(1)}: Either \code{"shows"} or \code{"movies"}. For season/episode-specific
functions, values \code{seasons} or \code{episodes} are also allowed.}

\item{extended}{\code{character(1)}: Either \code{"min"} (API default) or \code{"full"}. The latter
returns more variables and should generally only be used if required.
See \code{vignette("finding-things", package = "tRakt")} for more details.}
}
\value{
A \link[tibble:tibble-package]{tibble()}. If the function has a \code{limit} parameter
(defaulting to \code{10}), this will be the (maximum) number of rows of the \code{tibble}.
If there are no results (or the API is unreachable), an empty \code{tibble()} is returned.
}
\description{
Get a user's watchlist
}
\examples{
\dontrun{
# Defaults to movie watchlist and minimal info
user_watchlist(user = "sean")
}
}
\seealso{
Other user data: \code{\link{user_collection}},
  \code{\link{user_comments}}, \code{\link{user_history}},
  \code{\link{user_likes}}, \code{\link{user_network}},
  \code{\link{user_profile}}, \code{\link{user_ratings}},
  \code{\link{user_stats}}, \code{\link{user_watched}}
}
\concept{user data}
